@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2021 Snowflake Computing Inc. All rights reserved.
+// Copyright (c) 2017-2022 Snowflake Computing Inc. All rights reserved.
 
 package gosnowflake
 
@@ -8,21 +8,28 @@ import (
 	"crypto/x509"
 	"encoding/base64"
 	"encoding/json"
+	"errors"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"net/url"
+	"os"
 	"runtime"
 	"strconv"
 	"strings"
 	"time"
 
-	"github.com/form3tech-oss/jwt-go"
-	"github.com/google/uuid"
+	"github.com/golang-jwt/jwt/v5"
 )
 
 const (
 	clientType = "Go"
+)
+
+const (
+	clientStoreTemporaryCredential = "CLIENT_STORE_TEMPORARY_CREDENTIAL"
+	clientRequestMfaToken          = "CLIENT_REQUEST_MFA_TOKEN"
+	idTokenAuthenticator           = "ID_TOKEN"
 )
 
 // AuthType indicates the type of authentication in Snowflake
@@ -41,6 +48,10 @@ const (
 	AuthTypeJwt
 	// AuthTypeTokenAccessor is to use the provided token accessor and bypass authentication
 	AuthTypeTokenAccessor
+	// AuthTypeUsernamePasswordMFA is to use username and password with mfa
+	AuthTypeUsernamePasswordMFA
+	// AuthTypePat is to use programmatic access token
+	AuthTypePat
 )
 
 func determineAuthenticatorType(cfg *Config, value string) error {
@@ -57,6 +68,15 @@ func determineAuthenticatorType(cfg *Config, value string) error {
 		return nil
 	} else if upperCaseValue == AuthTypeExternalBrowser.String() {
 		cfg.Authenticator = AuthTypeExternalBrowser
+		return nil
+	} else if upperCaseValue == AuthTypeUsernamePasswordMFA.String() {
+		cfg.Authenticator = AuthTypeUsernamePasswordMFA
+		return nil
+	} else if upperCaseValue == AuthTypeTokenAccessor.String() {
+		cfg.Authenticator = AuthTypeTokenAccessor
+		return nil
+	} else if upperCaseValue == AuthTypePat.String() && experimentalAuthEnabled() {
+		cfg.Authenticator = AuthTypePat
 		return nil
 	} else {
 		// possibly Okta case
@@ -78,7 +98,7 @@ func determineAuthenticatorType(cfg *Config, value string) error {
 			}
 		}
 
-		if oktaURL.Scheme != "https" || !strings.HasSuffix(oktaURL.Host, "okta.com") {
+		if oktaURL.Scheme != "https" {
 			return &SnowflakeError{
 				Number:      ErrCodeFailedToParseAuthenticator,
 				Message:     errMsgFailedToParseAuthenticator,
@@ -105,6 +125,10 @@ func (authType AuthType) String() string {
 		return "SNOWFLAKE_JWT"
 	case AuthTypeTokenAccessor:
 		return "TOKENACCESSOR"
+	case AuthTypeUsernamePasswordMFA:
+		return "USERNAME_PASSWORD_MFA"
+	case AuthTypePat:
+		return "PROGRAMMATIC_ACCESS_TOKEN"
 	default:
 		return "UNKNOWN"
 	}
@@ -130,7 +154,9 @@ type authRequestClientEnvironment struct {
 	Os          string `json:"OS"`
 	OsVersion   string `json:"OS_VERSION"`
 	OCSPMode    string `json:"OCSP_MODE"`
+	GoVersion   string `json:"GO_VERSION"`
 }
+
 type authRequestData struct {
 	ClientAppID             string                       `json:"CLIENT_APP_ID"`
 	ClientAppVersion        string                       `json:"CLIENT_APP_VERSION"`
@@ -169,6 +195,10 @@ type authResponseMain struct {
 	Validity            time.Duration           `json:"validityInSeconds,omitempty"`
 	MasterToken         string                  `json:"masterToken,omitempty"`
 	MasterValidity      time.Duration           `json:"masterValidityInSeconds"`
+	MfaToken            string                  `json:"mfaToken,omitempty"`
+	MfaTokenValidity    time.Duration           `json:"mfaTokenValidityInSeconds"`
+	IDToken             string                  `json:"idToken,omitempty"`
+	IDTokenValidity     time.Duration           `json:"idTokenValidityInSeconds"`
 	DisplayUserName     string                  `json:"displayUserName"`
 	ServerVersion       string                  `json:"serverVersion"`
 	FirstLogin          bool                    `json:"firstLogin"`
@@ -183,6 +213,7 @@ type authResponseMain struct {
 	SSOURL              string                  `json:"ssoUrl,omitempty"`
 	ProofKey            string                  `json:"proofKey,omitempty"`
 }
+
 type authResponse struct {
 	Data    authResponseMain `json:"data"`
 	Message string           `json:"message"`
@@ -193,17 +224,18 @@ type authResponse struct {
 func postAuth(
 	ctx context.Context,
 	sr *snowflakeRestful,
+	client *http.Client,
 	params *url.Values,
 	headers map[string]string,
-	body []byte,
+	bodyCreator bodyCreatorType,
 	timeout time.Duration) (
 	data *authResponse, err error) {
-	params.Add(requestIDKey, getOrGenerateRequestIDFromContext(ctx).String())
-	params.Add(requestGUIDKey, uuid.New().String())
+	params.Set(requestIDKey, getOrGenerateRequestIDFromContext(ctx).String())
+	params.Set(requestGUIDKey, NewUUID().String())
 
 	fullURL := sr.getFullURL(loginRequestPath, params)
-	logger.Infof("full URL: %v", fullURL)
-	resp, err := sr.FuncPost(ctx, sr, fullURL, headers, body, timeout, true)
+	logger.WithContext(ctx).Infof("full URL: %v", fullURL)
+	resp, err := sr.FuncAuthPost(ctx, client, fullURL, headers, bodyCreator, timeout, sr.MaxRetryCount)
 	if err != nil {
 		return nil, err
 	}
@@ -212,7 +244,7 @@ func postAuth(
 		var respd authResponse
 		err = json.NewDecoder(resp.Body).Decode(&respd)
 		if err != nil {
-			logger.Error("failed to decode JSON. err: %v", err)
+			logger.WithContext(ctx).Errorf("failed to decode JSON. err: %v", err)
 			return nil, err
 		}
 		return &respd, nil
@@ -235,13 +267,13 @@ func postAuth(
 			MessageArgs: []interface{}{resp.StatusCode, fullURL},
 		}
 	}
-	b, err := ioutil.ReadAll(resp.Body)
+	b, err := io.ReadAll(resp.Body)
 	if err != nil {
-		logger.Errorf("failed to extract HTTP response body. err: %v", err)
+		logger.WithContext(ctx).Errorf("failed to extract HTTP response body. err: %v", err)
 		return nil, err
 	}
-	logger.Infof("HTTP: %v, URL: %v, Body: %v", resp.StatusCode, fullURL, b)
-	logger.Infof("Header: %v", resp.Header)
+	logger.WithContext(ctx).Infof("HTTP: %v, URL: %v, Body: %v", resp.StatusCode, fullURL, b)
+	logger.WithContext(ctx).Infof("Header: %v", resp.Header)
 	return nil, &SnowflakeError{
 		Number:      ErrFailedToAuth,
 		SQLState:    SQLStateConnectionRejected,
@@ -256,6 +288,8 @@ func getHeaders() map[string]string {
 	headers := make(map[string]string)
 	headers[httpHeaderContentType] = headerContentTypeApplicationJSON
 	headers[httpHeaderAccept] = headerAcceptTypeApplicationSnowflake
+	headers[httpClientAppID] = clientType
+	headers[httpClientAppVersion] = SnowflakeGoDriverVersion
 	headers[httpHeaderUserAgent] = userAgent
 	return headers
 }
@@ -267,63 +301,8 @@ func authenticate(
 	samlResponse []byte,
 	proofKey []byte,
 ) (resp *authResponseMain, err error) {
-	headers := getHeaders()
-	clientEnvironment := authRequestClientEnvironment{
-		Application: sc.cfg.Application,
-		Os:          operatingSystem,
-		OsVersion:   platform,
-		OCSPMode:    sc.cfg.ocspMode(),
-	}
-
-	sessionParameters := make(map[string]interface{})
-	for k, v := range sc.cfg.Params {
-		// upper casing to normalize keys
-		sessionParameters[strings.ToUpper(k)] = *v
-	}
-
-	sessionParameters[sessionClientValidateDefaultParameters] = sc.cfg.ValidateDefaultParameters != ConfigBoolFalse
-
-	requestMain := authRequestData{
-		ClientAppID:       clientType,
-		ClientAppVersion:  SnowflakeGoDriverVersion,
-		AccountName:       sc.cfg.Account,
-		SessionParameters: sessionParameters,
-		ClientEnvironment: clientEnvironment,
-	}
-
-	switch sc.cfg.Authenticator {
-	case AuthTypeExternalBrowser:
-		requestMain.ProofKey = string(proofKey)
-		requestMain.Token = string(samlResponse)
-		requestMain.LoginName = sc.cfg.User
-		requestMain.Authenticator = AuthTypeExternalBrowser.String()
-	case AuthTypeOAuth:
-		requestMain.LoginName = sc.cfg.User
-		requestMain.Authenticator = AuthTypeOAuth.String()
-		requestMain.Token = sc.cfg.Token
-	case AuthTypeOkta:
-		requestMain.RawSAMLResponse = string(samlResponse)
-	case AuthTypeJwt:
-		requestMain.Authenticator = AuthTypeJwt.String()
-
-		jwtTokenString, err := prepareJWTToken(sc.cfg)
-		if err != nil {
-			return nil, err
-		}
-		requestMain.Token = jwtTokenString
-	case AuthTypeSnowflake:
-		logger.Info("Username and password")
-		requestMain.LoginName = sc.cfg.User
-		requestMain.Password = sc.cfg.Password
-		switch {
-		case sc.cfg.PasscodeInPassword:
-			requestMain.ExtAuthnDuoMethod = "passcode"
-		case sc.cfg.Passcode != "":
-			requestMain.Passcode = sc.cfg.Passcode
-			requestMain.ExtAuthnDuoMethod = "passcode"
-		}
-	case AuthTypeTokenAccessor:
-		logger.Info("Bypass authentication using existing token from token accessor")
+	if sc.cfg.Authenticator == AuthTypeTokenAccessor {
+		logger.WithContext(ctx).Info("Bypass authentication using existing token from token accessor")
 		sessionInfo := authResponseSessionInfo{
 			DatabaseName:  sc.cfg.Database,
 			SchemaName:    sc.cfg.Schema,
@@ -339,9 +318,34 @@ func authenticate(
 		}, nil
 	}
 
-	authRequest := authRequest{
-		Data: requestMain,
+	headers := getHeaders()
+	clientEnvironment := authRequestClientEnvironment{
+		Application: sc.cfg.Application,
+		Os:          operatingSystem,
+		OsVersion:   platform,
+		OCSPMode:    sc.cfg.ocspMode(),
+		GoVersion:   runtime.Version(),
 	}
+
+	sessionParameters := make(map[string]interface{})
+	paramsMutex.Lock()
+	for k, v := range sc.cfg.Params {
+		// upper casing to normalize keys
+		sessionParameters[strings.ToUpper(k)] = *v
+	}
+	paramsMutex.Unlock()
+
+	sessionParameters[sessionClientValidateDefaultParameters] = sc.cfg.ValidateDefaultParameters != ConfigBoolFalse
+	if sc.cfg.ClientRequestMfaToken == ConfigBoolTrue {
+		sessionParameters[clientRequestMfaToken] = true
+	}
+	if sc.cfg.ClientStoreTemporaryCredential == ConfigBoolTrue {
+		sessionParameters[clientStoreTemporaryCredential] = true
+	}
+	bodyCreator := func() ([]byte, error) {
+		return createRequestBody(sc, sessionParameters, clientEnvironment, proofKey, samlResponse)
+	}
+
 	params := &url.Values{}
 	if sc.cfg.Database != "" {
 		params.Add("databaseName", sc.cfg.Database)
@@ -356,24 +360,24 @@ func authenticate(
 		params.Add("roleName", sc.cfg.Role)
 	}
 
-	jsonBody, err := json.Marshal(authRequest)
-	if err != nil {
-		return
-	}
-
-	logger.WithContext(sc.ctx).Infof("PARAMS for Auth: %v, %v, %v, %v, %v, %v",
+	logger.WithContext(ctx).WithContext(sc.ctx).Infof("PARAMS for Auth: %v, %v, %v, %v, %v, %v",
 		params, sc.rest.Protocol, sc.rest.Host, sc.rest.Port, sc.rest.LoginTimeout, sc.cfg.Authenticator.String())
 
-	respd, err := sc.rest.FuncPostAuth(ctx, sc.rest, params, headers, jsonBody, sc.rest.LoginTimeout)
+	respd, err := sc.rest.FuncPostAuth(ctx, sc.rest, sc.rest.getClientFor(sc.cfg.Authenticator), params, headers, bodyCreator, sc.rest.LoginTimeout)
 	if err != nil {
 		return nil, err
 	}
 	if !respd.Success {
-		logger.Errorln("Authentication FAILED")
+		logger.WithContext(ctx).Errorln("Authentication FAILED")
 		sc.rest.TokenAccessor.SetTokens("", "", -1)
+		if sessionParameters[clientRequestMfaToken] == true {
+			credentialsStorage.deleteCredential(newMfaTokenSpec(sc.cfg.Host, sc.cfg.User))
+		}
+		if sessionParameters[clientStoreTemporaryCredential] == true {
+			credentialsStorage.deleteCredential(newIDTokenSpec(sc.cfg.Host, sc.cfg.User))
+		}
 		code, err := strconv.Atoi(respd.Code)
 		if err != nil {
-			code = -1
 			return nil, err
 		}
 		return nil, (&SnowflakeError{
@@ -382,30 +386,139 @@ func authenticate(
 			Message:  respd.Message,
 		}).exceptionTelemetry(sc)
 	}
-	logger.Info("Authentication SUCCESS")
+	logger.WithContext(ctx).Info("Authentication SUCCESS")
 	sc.rest.TokenAccessor.SetTokens(respd.Data.Token, respd.Data.MasterToken, respd.Data.SessionID)
+	if sessionParameters[clientRequestMfaToken] == true {
+		token := respd.Data.MfaToken
+		credentialsStorage.setCredential(newMfaTokenSpec(sc.cfg.Host, sc.cfg.User), token)
+	}
+	if sessionParameters[clientStoreTemporaryCredential] == true {
+		token := respd.Data.IDToken
+		credentialsStorage.setCredential(newIDTokenSpec(sc.cfg.Host, sc.cfg.User), token)
+	}
 	return &respd.Data, nil
+}
+
+func createRequestBody(sc *snowflakeConn, sessionParameters map[string]interface{},
+	clientEnvironment authRequestClientEnvironment, proofKey []byte, samlResponse []byte,
+) ([]byte, error) {
+	requestMain := authRequestData{
+		ClientAppID:       clientType,
+		ClientAppVersion:  SnowflakeGoDriverVersion,
+		AccountName:       sc.cfg.Account,
+		SessionParameters: sessionParameters,
+		ClientEnvironment: clientEnvironment,
+	}
+
+	switch sc.cfg.Authenticator {
+	case AuthTypeExternalBrowser:
+		if sc.cfg.IDToken != "" {
+			requestMain.Authenticator = idTokenAuthenticator
+			requestMain.Token = sc.cfg.IDToken
+			requestMain.LoginName = sc.cfg.User
+		} else {
+			requestMain.ProofKey = string(proofKey)
+			requestMain.Token = string(samlResponse)
+			requestMain.LoginName = sc.cfg.User
+			requestMain.Authenticator = AuthTypeExternalBrowser.String()
+		}
+	case AuthTypeOAuth:
+		requestMain.LoginName = sc.cfg.User
+		requestMain.Authenticator = AuthTypeOAuth.String()
+		requestMain.Token = sc.cfg.Token
+	case AuthTypeOkta:
+		samlResponse, err := authenticateBySAML(
+			sc.ctx,
+			sc.rest,
+			sc.cfg.OktaURL,
+			sc.cfg.Application,
+			sc.cfg.Account,
+			sc.cfg.User,
+			sc.cfg.Password,
+			sc.cfg.DisableSamlURLCheck)
+		if err != nil {
+			return nil, err
+		}
+		requestMain.RawSAMLResponse = string(samlResponse)
+	case AuthTypeJwt:
+		requestMain.Authenticator = AuthTypeJwt.String()
+
+		jwtTokenString, err := prepareJWTToken(sc.cfg)
+		if err != nil {
+			return nil, err
+		}
+		requestMain.Token = jwtTokenString
+	case AuthTypePat:
+		if !experimentalAuthEnabled() {
+			return nil, errors.New("programmatic access tokens are not ready to use")
+		}
+		logger.WithContext(sc.ctx).Info("Programmatic access token")
+		requestMain.Authenticator = AuthTypePat.String()
+		requestMain.LoginName = sc.cfg.User
+		requestMain.Token = sc.cfg.Token
+		if sc.cfg.Password != "" && sc.cfg.Token == "" {
+			requestMain.Token = sc.cfg.Password
+		}
+	case AuthTypeSnowflake:
+		logger.WithContext(sc.ctx).Info("Username and password")
+		requestMain.LoginName = sc.cfg.User
+		requestMain.Password = sc.cfg.Password
+		switch {
+		case sc.cfg.PasscodeInPassword:
+			requestMain.ExtAuthnDuoMethod = "passcode"
+		case sc.cfg.Passcode != "":
+			requestMain.Passcode = sc.cfg.Passcode
+			requestMain.ExtAuthnDuoMethod = "passcode"
+		}
+	case AuthTypeUsernamePasswordMFA:
+		logger.WithContext(sc.ctx).Info("Username and password MFA")
+		requestMain.LoginName = sc.cfg.User
+		requestMain.Password = sc.cfg.Password
+		switch {
+		case sc.cfg.MfaToken != "":
+			requestMain.Token = sc.cfg.MfaToken
+		case sc.cfg.PasscodeInPassword:
+			requestMain.ExtAuthnDuoMethod = "passcode"
+		case sc.cfg.Passcode != "":
+			requestMain.Passcode = sc.cfg.Passcode
+			requestMain.ExtAuthnDuoMethod = "passcode"
+		}
+	}
+
+	authRequest := authRequest{
+		Data: requestMain,
+	}
+	jsonBody, err := json.Marshal(authRequest)
+	if err != nil {
+		return nil, err
+	}
+	return jsonBody, nil
 }
 
 // Generate a JWT token in string given the configuration
 func prepareJWTToken(config *Config) (string, error) {
+	if config.PrivateKey == nil {
+		return "", errors.New("trying to use keypair authentication, but PrivateKey was not provided in the driver config")
+	}
+	logger.Debug("preparing JWT for keypair authentication")
 	pubBytes, err := x509.MarshalPKIXPublicKey(config.PrivateKey.Public())
 	if err != nil {
 		return "", err
 	}
 	hash := sha256.Sum256(pubBytes)
 
-	accountName := strings.ToUpper(config.Account)
+	accountName := extractAccountName(config.Account)
 	userName := strings.ToUpper(config.User)
 
 	issueAtTime := time.Now().UTC()
-	token := jwt.NewWithClaims(jwt.SigningMethodRS256, jwt.MapClaims{
+	jwtClaims := jwt.MapClaims{
 		"iss": fmt.Sprintf("%s.%s.%s", accountName, userName, "SHA256:"+base64.StdEncoding.EncodeToString(hash[:])),
 		"sub": fmt.Sprintf("%s.%s", accountName, userName),
 		"iat": issueAtTime.Unix(),
 		"nbf": time.Date(2015, 10, 10, 12, 0, 0, 0, time.UTC).Unix(),
 		"exp": issueAtTime.Add(config.JWTExpireTimeout).Unix(),
-	})
+	}
+	token := jwt.NewWithClaims(jwt.SigningMethodRS256, jwtClaims)
 
 	tokenString, err := token.SignedString(config.PrivateKey)
 
@@ -413,6 +526,7 @@ func prepareJWTToken(config *Config) (string, error) {
 		return "", err
 	}
 
+	logger.Debugf("successfully generated JWT with following claims: %v", jwtClaims)
 	return tokenString, err
 }
 
@@ -422,33 +536,48 @@ func authenticateWithConfig(sc *snowflakeConn) error {
 	var samlResponse []byte
 	var proofKey []byte
 	var err error
-	logger.Infof("Authenticating via %v", sc.cfg.Authenticator.String())
+	//var consentCacheIdToken = true
+
+	if sc.cfg.Authenticator == AuthTypeExternalBrowser {
+		if (runtime.GOOS == "windows" || runtime.GOOS == "darwin") && sc.cfg.ClientStoreTemporaryCredential == configBoolNotSet {
+			sc.cfg.ClientStoreTemporaryCredential = ConfigBoolTrue
+		}
+		if sc.cfg.ClientStoreTemporaryCredential == ConfigBoolTrue {
+			sc.cfg.IDToken = credentialsStorage.getCredential(newIDTokenSpec(sc.cfg.Host, sc.cfg.User))
+		}
+		// Disable console login by default
+		if sc.cfg.DisableConsoleLogin == configBoolNotSet {
+			sc.cfg.DisableConsoleLogin = ConfigBoolTrue
+		}
+	}
+
+	if sc.cfg.Authenticator == AuthTypeUsernamePasswordMFA {
+		if (runtime.GOOS == "windows" || runtime.GOOS == "darwin") && sc.cfg.ClientRequestMfaToken == configBoolNotSet {
+			sc.cfg.ClientRequestMfaToken = ConfigBoolTrue
+		}
+		if sc.cfg.ClientRequestMfaToken == ConfigBoolTrue {
+			sc.cfg.MfaToken = credentialsStorage.getCredential(newMfaTokenSpec(sc.cfg.Host, sc.cfg.User))
+		}
+	}
+
+	logger.WithContext(sc.ctx).Infof("Authenticating via %v", sc.cfg.Authenticator.String())
 	switch sc.cfg.Authenticator {
 	case AuthTypeExternalBrowser:
-		samlResponse, proofKey, err = authenticateByExternalBrowser(
-			sc.ctx,
-			sc.rest,
-			sc.cfg.Authenticator.String(),
-			sc.cfg.Application,
-			sc.cfg.Account,
-			sc.cfg.User,
-			sc.cfg.Password)
-		if err != nil {
-			sc.cleanup()
-			return err
-		}
-	case AuthTypeOkta:
-		samlResponse, err = authenticateBySAML(
-			sc.ctx,
-			sc.rest,
-			sc.cfg.OktaURL,
-			sc.cfg.Application,
-			sc.cfg.Account,
-			sc.cfg.User,
-			sc.cfg.Password)
-		if err != nil {
-			sc.cleanup()
-			return err
+		if sc.cfg.IDToken == "" {
+			samlResponse, proofKey, err = authenticateByExternalBrowser(
+				sc.ctx,
+				sc.rest,
+				sc.cfg.Authenticator.String(),
+				sc.cfg.Application,
+				sc.cfg.Account,
+				sc.cfg.User,
+				sc.cfg.Password,
+				sc.cfg.ExternalBrowserTimeout,
+				sc.cfg.DisableConsoleLogin)
+			if err != nil {
+				sc.cleanup()
+				return err
+			}
 		}
 	}
 	authData, err = authenticate(
@@ -463,4 +592,9 @@ func authenticateWithConfig(sc *snowflakeConn) error {
 	sc.populateSessionParameters(authData.Parameters)
 	sc.ctx = context.WithValue(sc.ctx, SFSessionIDKey, authData.SessionID)
 	return nil
+}
+
+func experimentalAuthEnabled() bool {
+	val, ok := os.LookupEnv("ENABLE_EXPERIMENTAL_AUTHENTICATION")
+	return ok && strings.EqualFold(val, "true")
 }
