@@ -1,4 +1,4 @@
-// Copyright (c) 2021 Snowflake Computing Inc. All right reserved.
+// Copyright (c) 2021-2022 Snowflake Computing Inc. All rights reserved.
 
 package gosnowflake
 
@@ -7,11 +7,9 @@ import (
 	"compress/gzip"
 	"context"
 	"encoding/json"
+	"errors"
 	"fmt"
-	"github.com/apache/arrow/go/arrow/ipc"
-	"github.com/apache/arrow/go/arrow/memory"
 	"io"
-	"io/ioutil"
 	"math/rand"
 	"net"
 	"net/http"
@@ -19,6 +17,10 @@ import (
 	"strings"
 	"sync"
 	"time"
+
+	"github.com/apache/arrow-go/v18/arrow"
+	"github.com/apache/arrow-go/v18/arrow/ipc"
+	"github.com/apache/arrow-go/v18/arrow/memory"
 )
 
 type chunkDownloader interface {
@@ -33,11 +35,13 @@ type chunkDownloader interface {
 	getRowType() []execResponseRowType
 	setNextChunkDownloader(downloader chunkDownloader)
 	getNextChunkDownloader() chunkDownloader
+	getArrowBatches() []*ArrowBatch
 }
 
 type snowflakeChunkDownloader struct {
 	sc                 *snowflakeConn
 	ctx                context.Context
+	pool               memory.Allocator
 	Total              int64
 	TotalRowIndex      int64
 	CellCount          int
@@ -54,13 +58,15 @@ type snowflakeChunkDownloader struct {
 	ChunksFinalErrors  []*chunkError
 	ChunksMutex        *sync.Mutex
 	DoneDownloadCond   *sync.Cond
+	FirstBatch         *ArrowBatch
 	NextDownloader     chunkDownloader
 	Qrmk               string
 	QueryResultFormat  string
+	ArrowBatches       []*ArrowBatch
 	RowSet             rowSetType
 	FuncDownload       func(context.Context, *snowflakeChunkDownloader, int)
 	FuncDownloadHelper func(context.Context, *snowflakeChunkDownloader, int) error
-	FuncGet            func(context.Context, *snowflakeChunkDownloader, string, map[string]string, time.Duration) (*http.Response, error)
+	FuncGet            func(context.Context, *snowflakeConn, string, map[string]string, time.Duration) (*http.Response, error)
 }
 
 func (scd *snowflakeChunkDownloader) totalUncompressedSize() (acc int64) {
@@ -87,6 +93,9 @@ func (scd *snowflakeChunkDownloader) nextResultSet() error {
 }
 
 func (scd *snowflakeChunkDownloader) start() error {
+	if usesArrowBatches(scd.ctx) && scd.getQueryResultFormat() == arrowFormat {
+		return scd.startArrowBatches()
+	}
 	scd.CurrentChunkSize = len(scd.RowSet.JSON) // cache the size
 	scd.CurrentIndex = -1                       // initial chunks idx
 	scd.CurrentChunkIndex = -1                  // initial chunk
@@ -95,11 +104,18 @@ func (scd *snowflakeChunkDownloader) start() error {
 	populateJSONRowSet(scd.CurrentChunk, scd.RowSet.JSON)
 
 	if scd.getQueryResultFormat() == arrowFormat && scd.RowSet.RowSetBase64 != "" {
+		params, err := scd.getConfigParams()
+		if err != nil {
+			return err
+		}
 		// if the rowsetbase64 retrieved from the server is empty, move on to downloading chunks
-		var err error
-		firstArrowChunk := buildFirstArrowChunk(scd.RowSet.RowSetBase64)
+		loc := getCurrentLocation(params)
+		firstArrowChunk, err := buildFirstArrowChunk(scd.RowSet.RowSetBase64, loc, scd.pool)
+		if err != nil {
+			return err
+		}
 		higherPrecision := higherPrecisionEnabled(scd.ctx)
-		scd.CurrentChunk, err = firstArrowChunk.decodeArrowChunk(scd.RowSet.RowType, higherPrecision)
+		scd.CurrentChunk, err = firstArrowChunk.decodeArrowChunk(scd.ctx, scd.RowSet.RowType, higherPrecision, params)
 		scd.CurrentChunkSize = firstArrowChunk.rowCount
 		if err != nil {
 			return err
@@ -109,15 +125,17 @@ func (scd *snowflakeChunkDownloader) start() error {
 	// start downloading chunks if exists
 	chunkMetaLen := len(scd.ChunkMetas)
 	if chunkMetaLen > 0 {
-		logger.Debugf("MaxChunkDownloadWorkers: %v", MaxChunkDownloadWorkers)
-		logger.Debugf("chunks: %v, total bytes: %d", chunkMetaLen, scd.totalUncompressedSize())
+		logger.WithContext(scd.ctx).Debugf("MaxChunkDownloadWorkers: %v", MaxChunkDownloadWorkers)
+		logger.WithContext(scd.ctx).Debugf("chunks: %v, total bytes: %d", chunkMetaLen, scd.totalUncompressedSize())
 		scd.ChunksMutex = &sync.Mutex{}
 		scd.DoneDownloadCond = sync.NewCond(scd.ChunksMutex)
 		scd.Chunks = make(map[int][]chunkRowType)
 		scd.ChunksChan = make(chan int, chunkMetaLen)
 		scd.ChunksError = make(chan *chunkError, MaxChunkDownloadWorkers)
 		for i := 0; i < chunkMetaLen; i++ {
-			logger.Debugf("add chunk to channel ChunksChan: %v", i+1)
+			chunk := scd.ChunkMetas[i]
+			logger.WithContext(scd.ctx).Debugf("add chunk to channel ChunksChan: %v, URL: %v, RowCount: %v, UncompressedSize: %v, ChunkResultFormat: %v",
+				i+1, chunk.URL, chunk.RowCount, chunk.UncompressedSize, scd.QueryResultFormat)
 			scd.ChunksChan <- i
 		}
 		for i := 0; i < intMin(MaxChunkDownloadWorkers, chunkMetaLen); i++ {
@@ -130,30 +148,42 @@ func (scd *snowflakeChunkDownloader) start() error {
 func (scd *snowflakeChunkDownloader) schedule() {
 	select {
 	case nextIdx := <-scd.ChunksChan:
-		logger.Infof("schedule chunk: %v", nextIdx+1)
-		go scd.FuncDownload(scd.ctx, scd, nextIdx)
+		logger.WithContext(scd.ctx).Infof("schedule chunk: %v", nextIdx+1)
+		go GoroutineWrapper(
+			scd.ctx,
+			func() {
+				scd.FuncDownload(scd.ctx, scd, nextIdx)
+			},
+		)
 	default:
 		// no more download
-		logger.Info("no more download")
+		logger.WithContext(scd.ctx).Info("no more download")
 	}
 }
 
 func (scd *snowflakeChunkDownloader) checkErrorRetry() (err error) {
 	select {
 	case errc := <-scd.ChunksError:
-		if scd.ChunksErrorCounter < maxChunkDownloaderErrorCounter && errc.Error != context.Canceled {
+		if scd.ChunksErrorCounter < maxChunkDownloaderErrorCounter &&
+			errc.Error != context.Canceled &&
+			errc.Error != context.DeadlineExceeded {
 			// add the index to the chunks channel so that the download will be retried.
-			go scd.FuncDownload(scd.ctx, scd, errc.Index)
+			go GoroutineWrapper(
+				scd.ctx,
+				func() {
+					scd.FuncDownload(scd.ctx, scd, errc.Index)
+				},
+			)
 			scd.ChunksErrorCounter++
-			logger.Warningf("chunk idx: %v, err: %v. retrying (%v/%v)...",
+			logger.WithContext(scd.ctx).Warningf("chunk idx: %v, err: %v. retrying (%v/%v)...",
 				errc.Index, errc.Error, scd.ChunksErrorCounter, maxChunkDownloaderErrorCounter)
 		} else {
 			scd.ChunksFinalErrors = append(scd.ChunksFinalErrors, errc)
-			logger.Warningf("chunk idx: %v, err: %v. no further retry", errc.Index, errc.Error)
+			logger.WithContext(scd.ctx).Warningf("chunk idx: %v, err: %v. no further retry", errc.Index, errc.Error)
 			return errc.Error
 		}
 	default:
-		logger.Info("no error is detected.")
+		logger.WithContext(scd.ctx).Info("no error is detected.")
 	}
 	return nil
 }
@@ -171,16 +201,15 @@ func (scd *snowflakeChunkDownloader) next() (chunkRowType, error) {
 		}
 
 		scd.ChunksMutex.Lock()
-		if scd.CurrentChunkIndex > 1 {
+		if scd.CurrentChunkIndex > 0 {
 			scd.Chunks[scd.CurrentChunkIndex-1] = nil // detach the previously used chunk
 		}
 
 		for scd.Chunks[scd.CurrentChunkIndex] == nil {
-			logger.Debugf("waiting for chunk idx: %v/%v",
+			logger.WithContext(scd.ctx).Debugf("waiting for chunk idx: %v/%v",
 				scd.CurrentChunkIndex+1, len(scd.ChunkMetas))
 
-			err := scd.checkErrorRetry()
-			if err != nil {
+			if err := scd.checkErrorRetry(); err != nil {
 				scd.ChunksMutex.Unlock()
 				return chunkRowType{}, err
 			}
@@ -189,7 +218,7 @@ func (scd *snowflakeChunkDownloader) next() (chunkRowType, error) {
 			// 1) one chunk download finishes or 2) an error occurs.
 			scd.DoneDownloadCond.Wait()
 		}
-		logger.Debugf("ready: chunk %v", scd.CurrentChunkIndex+1)
+		logger.WithContext(scd.ctx).Debugf("ready: chunk %v", scd.CurrentChunkIndex+1)
 		scd.CurrentChunk = scd.Chunks[scd.CurrentChunkIndex]
 		scd.ChunksMutex.Unlock()
 		scd.CurrentChunkSize = len(scd.CurrentChunk)
@@ -198,7 +227,7 @@ func (scd *snowflakeChunkDownloader) next() (chunkRowType, error) {
 		scd.schedule()
 	}
 
-	logger.Debugf("no more data")
+	logger.WithContext(scd.ctx).Debugf("no more data")
 	if len(scd.ChunkMetas) > 0 {
 		close(scd.ChunksError)
 		close(scd.ChunksChan)
@@ -215,10 +244,7 @@ func (scd *snowflakeChunkDownloader) getChunkMetas() []execResponseChunk {
 }
 
 func (scd *snowflakeChunkDownloader) getQueryResultFormat() resultFormat {
-	if scd.QueryResultFormat == "json" {
-		return jsonFormat
-	}
-	return arrowFormat
+	return resultFormat(scd.QueryResultFormat)
 }
 
 func (scd *snowflakeChunkDownloader) setNextChunkDownloader(nextDownloader chunkDownloader) {
@@ -233,18 +259,73 @@ func (scd *snowflakeChunkDownloader) getRowType() []execResponseRowType {
 	return scd.RowSet.RowType
 }
 
+func (scd *snowflakeChunkDownloader) getArrowBatches() []*ArrowBatch {
+	if scd.FirstBatch == nil || scd.FirstBatch.rec == nil {
+		return scd.ArrowBatches
+	}
+	return append([]*ArrowBatch{scd.FirstBatch}, scd.ArrowBatches...)
+}
+
+func (scd *snowflakeChunkDownloader) getConfigParams() (map[string]*string, error) {
+	if scd.sc == nil || scd.sc.cfg == nil {
+		return map[string]*string{}, errors.New("failed to retrieve connection")
+	}
+	return scd.sc.cfg.Params, nil
+}
+
 func getChunk(
 	ctx context.Context,
-	scd *snowflakeChunkDownloader,
+	sc *snowflakeConn,
 	fullURL string,
 	headers map[string]string,
 	timeout time.Duration) (
-	*http.Response, error) {
+	*http.Response, error,
+) {
 	u, err := url.Parse(fullURL)
 	if err != nil {
 		return nil, err
 	}
-	return newRetryHTTP(ctx, scd.sc.rest.Client, http.NewRequest, u, headers, timeout).execute()
+	return newRetryHTTP(ctx, sc.rest.Client, http.NewRequest, u, headers, timeout, sc.rest.MaxRetryCount, sc.currentTimeProvider, sc.cfg).execute()
+}
+
+func (scd *snowflakeChunkDownloader) startArrowBatches() error {
+	var loc *time.Location
+	params, err := scd.getConfigParams()
+	if err != nil {
+		return err
+	}
+	loc = getCurrentLocation(params)
+	if scd.RowSet.RowSetBase64 != "" {
+		var err error
+		firstArrowChunk, err := buildFirstArrowChunk(scd.RowSet.RowSetBase64, loc, scd.pool)
+		if err != nil {
+			return err
+		}
+		scd.FirstBatch = &ArrowBatch{
+			idx:                0,
+			scd:                scd,
+			funcDownloadHelper: scd.FuncDownloadHelper,
+			loc:                loc,
+		}
+		// decode first chunk if possible
+		if firstArrowChunk.allocator != nil {
+			scd.FirstBatch.rec, err = firstArrowChunk.decodeArrowBatch(scd)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	chunkMetaLen := len(scd.ChunkMetas)
+	scd.ArrowBatches = make([]*ArrowBatch, chunkMetaLen)
+	for i := range scd.ArrowBatches {
+		scd.ArrowBatches[i] = &ArrowBatch{
+			idx:                i,
+			scd:                scd,
+			funcDownloadHelper: scd.FuncDownloadHelper,
+			loc:                loc,
+		}
+	}
+	return nil
 }
 
 /* largeResultSetReader is a reader that wraps the large result set with leading and tailing brackets. */
@@ -281,11 +362,11 @@ func (r *largeResultSetReader) Read(p []byte) (n int, err error) {
 }
 
 func downloadChunk(ctx context.Context, scd *snowflakeChunkDownloader, idx int) {
-	logger.Infof("download start chunk: %v", idx+1)
+	logger.WithContext(ctx).Infof("download start chunk: %v", idx+1)
 	defer scd.DoneDownloadCond.Broadcast()
 
 	if err := scd.FuncDownloadHelper(ctx, scd, idx); err != nil {
-		logger.Errorf(
+		logger.WithContext(ctx).Errorf(
 			"failed to extract HTTP response body. URL: %v, err: %v", scd.ChunkMetas[idx].URL, err)
 		scd.ChunksError <- &chunkError{Index: idx, Error: err}
 	} else if scd.ctx.Err() == context.Canceled || scd.ctx.Err() == context.DeadlineExceeded {
@@ -296,8 +377,10 @@ func downloadChunk(ctx context.Context, scd *snowflakeChunkDownloader, idx int) 
 func downloadChunkHelper(ctx context.Context, scd *snowflakeChunkDownloader, idx int) error {
 	headers := make(map[string]string)
 	if len(scd.ChunkHeader) > 0 {
-		logger.Debug("chunk header is provided.")
+		logger.WithContext(ctx).Debug("chunk header is provided.")
 		for k, v := range scd.ChunkHeader {
+			logger.WithContext(ctx).Debugf("adding header: %v, value: %v", k, v)
+
 			headers[k] = v
 		}
 	} else {
@@ -305,20 +388,20 @@ func downloadChunkHelper(ctx context.Context, scd *snowflakeChunkDownloader, idx
 		headers[headerSseCKey] = scd.Qrmk
 	}
 
-	resp, err := scd.FuncGet(ctx, scd, scd.ChunkMetas[idx].URL, headers, scd.sc.rest.RequestTimeout)
+	resp, err := scd.FuncGet(ctx, scd.sc, scd.ChunkMetas[idx].URL, headers, scd.sc.rest.RequestTimeout)
 	if err != nil {
 		return err
 	}
 	bufStream := bufio.NewReader(resp.Body)
 	defer resp.Body.Close()
-	logger.Infof("response returned chunk: %v, resp: %v", idx+1, resp)
+	logger.WithContext(ctx).Debugf("response returned chunk: %v for URL: %v", idx+1, scd.ChunkMetas[idx].URL)
 	if resp.StatusCode != http.StatusOK {
-		b, err := ioutil.ReadAll(bufStream)
+		b, err := io.ReadAll(bufStream)
 		if err != nil {
 			return err
 		}
-		logger.Infof("HTTP: %v, URL: %v, Body: %v", resp.StatusCode, scd.ChunkMetas[idx].URL, b)
-		logger.Infof("Header: %v", resp.Header)
+		logger.WithContext(ctx).Infof("HTTP: %v, URL: %v, Body: %v", resp.StatusCode, scd.ChunkMetas[idx].URL, b)
+		logger.WithContext(ctx).Infof("Header: %v", resp.Header)
 		return &SnowflakeError{
 			Number:      ErrFailedToGetChunk,
 			SQLState:    SQLStateConnectionFailure,
@@ -326,10 +409,10 @@ func downloadChunkHelper(ctx context.Context, scd *snowflakeChunkDownloader, idx
 			MessageArgs: []interface{}{idx},
 		}
 	}
-	return decodeChunk(scd, idx, bufStream)
+	return decodeChunk(ctx, scd, idx, bufStream)
 }
 
-func decodeChunk(scd *snowflakeChunkDownloader, idx int, bufStream *bufio.Reader) (err error) {
+func decodeChunk(ctx context.Context, scd *snowflakeChunkDownloader, idx int, bufStream *bufio.Reader) (err error) {
 	gzipMagic, err := bufStream.Peek(2)
 	if err != nil {
 		return err
@@ -372,23 +455,37 @@ func decodeChunk(scd *snowflakeChunkDownloader, idx int, bufStream *bufio.Reader
 		respd = make([]chunkRowType, len(decRespd))
 		populateJSONRowSet(respd, decRespd)
 	} else {
-		ipcReader, err := ipc.NewReader(source)
+		ipcReader, err := ipc.NewReader(source, ipc.WithAllocator(scd.pool))
 		if err != nil {
 			return err
 		}
+		var loc *time.Location
+		params, err := scd.getConfigParams()
+		if err != nil {
+			return err
+		}
+		loc = getCurrentLocation(params)
 		arc := arrowResultChunk{
-			*ipcReader,
+			ipcReader,
 			0,
-			int(scd.totalUncompressedSize()),
-			memory.NewGoAllocator(),
+			loc,
+			scd.pool,
+		}
+		if usesArrowBatches(scd.ctx) {
+			if scd.ArrowBatches[idx].rec, err = arc.decodeArrowBatch(scd); err != nil {
+				return err
+			}
+			// updating metadata
+			scd.ArrowBatches[idx].rowCount = countArrowBatchRows(scd.ArrowBatches[idx].rec)
+			return nil
 		}
 		highPrec := higherPrecisionEnabled(scd.ctx)
-		respd, err = arc.decodeArrowChunk(scd.RowSet.RowType, highPrec)
+		respd, err = arc.decodeArrowChunk(ctx, scd.RowSet.RowType, highPrec, params)
 		if err != nil {
 			return err
 		}
 	}
-	logger.Debugf(
+	logger.WithContext(scd.ctx).Debugf(
 		"decoded %d rows w/ %d bytes in %s (chunk %v)",
 		scd.ChunkMetas[idx].RowCount,
 		scd.ChunkMetas[idx].UncompressedSize,
@@ -433,55 +530,58 @@ func (scd *streamChunkDownloader) nextResultSet() error {
 }
 
 func (scd *streamChunkDownloader) start() error {
-	go func() {
-		var readErr = io.EOF
+	go GoroutineWrapper(
+		scd.ctx,
+		func() {
+			readErr := io.EOF
 
-		logger.WithContext(scd.ctx).Infof(
-			"start downloading. downloader id: %v, %v/%v rows, %v chunks",
-			scd.id, len(scd.RowSet.RowType), scd.Total, len(scd.ChunkMetas))
-		t := time.Now()
+			logger.WithContext(scd.ctx).Infof(
+				"start downloading. downloader id: %v, %v/%v rows, %v chunks",
+				scd.id, len(scd.RowSet.RowType), scd.Total, len(scd.ChunkMetas))
+			t := time.Now()
 
-		defer func() {
-			if readErr == io.EOF {
-				logger.WithContext(scd.ctx).Infof("downloading done. downloader id: %v", scd.id)
-			} else {
-				logger.WithContext(scd.ctx).Debugf("downloading error. downloader id: %v", scd.id)
-			}
-			scd.readErr = readErr
-			close(scd.rowStream)
-
-			if r := recover(); r != nil {
-				if err, ok := r.(error); ok {
-					readErr = err
+			defer func() {
+				if readErr == io.EOF {
+					logger.WithContext(scd.ctx).Infof("downloading done. downloader id: %v", scd.id)
 				} else {
-					readErr = fmt.Errorf("%v", r)
+					logger.WithContext(scd.ctx).Debugf("downloading error. downloader id: %v", scd.id)
 				}
-			}
-		}()
+				scd.readErr = readErr
+				close(scd.rowStream)
 
-		logger.WithContext(scd.ctx).Infof("sending initial set of rows in %vms", time.Since(t).Microseconds())
-		t = time.Now()
-		for _, row := range scd.RowSet.JSON {
-			scd.rowStream <- row
-		}
-		scd.RowSet.JSON = nil
+				if r := recover(); r != nil {
+					if err, ok := r.(error); ok {
+						readErr = err
+					} else {
+						readErr = fmt.Errorf("%v", r)
+					}
+				}
+			}()
 
-		// Download and parse one chunk at a time. The fetcher will send each
-		// parsed row to the row stream. When an error occurs, the fetcher will
-		// stop writing to the row stream so we can stop processing immediately
-		for i, chunk := range scd.ChunkMetas {
-			logger.WithContext(scd.ctx).Infof("starting chunk fetch %d (%d rows)", i, chunk.RowCount)
-			if err := scd.fetcher.fetch(chunk.URL, scd.rowStream); err != nil {
-				logger.WithContext(scd.ctx).Debugf(
-					"failed chunk fetch %d: %#v, downloader id: %v, %v/%v rows, %v chunks",
-					i, err, scd.id, len(scd.RowSet.RowType), scd.Total, len(scd.ChunkMetas))
-				readErr = fmt.Errorf("chunk fetch: %w", err)
-				break
-			}
-			logger.WithContext(scd.ctx).Infof("fetched chunk %d (%d rows) in %vms", i, chunk.RowCount, time.Since(t).Microseconds())
+			logger.WithContext(scd.ctx).Infof("sending initial set of rows in %vms", time.Since(t).Microseconds())
 			t = time.Now()
-		}
-	}()
+			for _, row := range scd.RowSet.JSON {
+				scd.rowStream <- row
+			}
+			scd.RowSet.JSON = nil
+
+			// Download and parse one chunk at a time. The fetcher will send each
+			// parsed row to the row stream. When an error occurs, the fetcher will
+			// stop writing to the row stream so we can stop processing immediately
+			for i, chunk := range scd.ChunkMetas {
+				logger.WithContext(scd.ctx).Infof("starting chunk fetch %d (%d rows)", i, chunk.RowCount)
+				if err := scd.fetcher.fetch(chunk.URL, scd.rowStream); err != nil {
+					logger.WithContext(scd.ctx).Debugf(
+						"failed chunk fetch %d: %#v, downloader id: %v, %v/%v rows, %v chunks",
+						i, err, scd.id, len(scd.RowSet.RowType), scd.Total, len(scd.ChunkMetas))
+					readErr = fmt.Errorf("chunk fetch: %w", err)
+					break
+				}
+				logger.WithContext(scd.ctx).Infof("fetched chunk %d (%d rows) in %vms", i, chunk.RowCount, time.Since(t).Microseconds())
+				t = time.Now()
+			}
+		},
+	)
 	return nil
 }
 
@@ -514,16 +614,17 @@ func (scd *streamChunkDownloader) getRowType() []execResponseRowType {
 	return scd.RowSet.RowType
 }
 
+func (scd *streamChunkDownloader) getArrowBatches() []*ArrowBatch {
+	return nil
+}
+
 func useStreamDownloader(ctx context.Context) bool {
 	val := ctx.Value(streamChunkDownload)
 	if val == nil {
 		return false
 	}
-	boolVal, ok := val.(bool)
-	if !ok {
-		return false
-	}
-	return boolVal
+	s, ok := val.(bool)
+	return s && ok
 }
 
 type streamChunkFetcher interface {
@@ -544,7 +645,8 @@ func newStreamChunkDownloader(
 	total int64,
 	rowType []execResponseRowType,
 	firstRows [][]*string,
-	chunks []execResponseChunk) *streamChunkDownloader {
+	chunks []execResponseChunk,
+) *streamChunkDownloader {
 	return &streamChunkDownloader{
 		ctx:        ctx,
 		id:         rand.Int63(),
@@ -565,14 +667,20 @@ func (f *httpStreamChunkFetcher) fetch(URL string, rows chan<- []*string) error 
 		}
 	}
 
-	fullURL, _ := url.Parse(URL)
-	res, err := newRetryHTTP(context.Background(), f.client, http.NewRequest, fullURL, f.headers, 0).execute()
+	fullURL, err := url.Parse(URL)
+	if err != nil {
+		return err
+	}
+	res, err := newRetryHTTP(context.Background(), f.client, http.NewRequest, fullURL, f.headers, 0, 0, defaultTimeProvider, nil).execute()
 	if err != nil {
 		return err
 	}
 	defer res.Body.Close()
 	if res.StatusCode != http.StatusOK {
-		b, _ := ioutil.ReadAll(res.Body)
+		b, err := io.ReadAll(res.Body)
+		if err != nil {
+			return err
+		}
 		return fmt.Errorf("status (%d): %s", res.StatusCode, string(b))
 	}
 	if err = copyChunkStream(res.Body, rows); err != nil {
@@ -625,4 +733,71 @@ func copyChunkStream(body io.Reader, rows chan<- []*string) error {
 		}
 	}
 	return nil
+}
+
+// ArrowBatch object represents a chunk of data, or subset of rows, retrievable in arrow.Record format
+type ArrowBatch struct {
+	rec                *[]arrow.Record
+	idx                int
+	rowCount           int
+	scd                *snowflakeChunkDownloader
+	funcDownloadHelper func(context.Context, *snowflakeChunkDownloader, int) error
+	ctx                context.Context
+	loc                *time.Location
+}
+
+// WithContext sets the context which will be used for this ArrowBatch.
+func (rb *ArrowBatch) WithContext(ctx context.Context) *ArrowBatch {
+	rb.ctx = ctx
+	return rb
+}
+
+// Fetch returns an array of records representing a chunk in the query
+func (rb *ArrowBatch) Fetch() (*[]arrow.Record, error) {
+	// chunk has already been downloaded
+	if rb.rec != nil {
+		// updating metadata
+		rb.rowCount = countArrowBatchRows(rb.rec)
+		return rb.rec, nil
+	}
+	var ctx context.Context
+	if rb.ctx != nil {
+		ctx = rb.ctx
+	} else {
+		ctx = context.Background()
+	}
+	if err := rb.funcDownloadHelper(ctx, rb.scd, rb.idx); err != nil {
+		return nil, err
+	}
+	return rb.rec, nil
+}
+
+// GetRowCount returns the number of rows in an arrow batch
+func (rb *ArrowBatch) GetRowCount() int {
+	return rb.rowCount
+}
+
+func getAllocator(ctx context.Context) memory.Allocator {
+	pool, ok := ctx.Value(arrowAlloc).(memory.Allocator)
+	if !ok {
+		return memory.DefaultAllocator
+	}
+	return pool
+}
+
+func usesArrowBatches(ctx context.Context) bool {
+	val := ctx.Value(arrowBatches)
+	if val == nil {
+		return false
+	}
+	a, ok := val.(bool)
+	return a && ok
+}
+
+func countArrowBatchRows(recs *[]arrow.Record) int {
+	var cnt int
+	for _, r := range *recs {
+		cnt += int(r.NumRows())
+	}
+	return cnt
 }
