@@ -1,13 +1,13 @@
-// Copyright (c) 2021 Snowflake Computing Inc. All right reserved.
+// Copyright (c) 2021-2022 Snowflake Computing Inc. All rights reserved.
 
 package gosnowflake
 
 import (
 	"bufio"
 	"compress/gzip"
+	"encoding/base64"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"math/rand"
 	"os"
 	"os/exec"
@@ -20,7 +20,12 @@ import (
 
 const timeFormat = "2006-01-02T15:04:05"
 
-func TestEncryptDecryptFile(t *testing.T) {
+type encryptDecryptTestFile struct {
+	numberOfBytesInEachRow int
+	numberOfLines          int
+}
+
+func TestEncryptDecryptFileCBC(t *testing.T) {
 	encMat := snowflakeFileEncryption{
 		"ztke8tIdVt1zmlQIZm0BMA==",
 		"123873c7-3a66-40c4-ab89-e3722fbccce1",
@@ -29,7 +34,7 @@ func TestEncryptDecryptFile(t *testing.T) {
 	data := "test data"
 	inputFile := "test_encrypt_decrypt_file"
 
-	fd, err := os.OpenFile(inputFile, os.O_CREATE|os.O_WRONLY, os.ModePerm)
+	fd, err := os.Create(inputFile)
 	if err != nil {
 		t.Error(err)
 	}
@@ -39,55 +44,106 @@ func TestEncryptDecryptFile(t *testing.T) {
 		t.Error(err)
 	}
 
-	metadata, encryptedFile, err := encryptFile(&encMat, inputFile, 0, "")
+	metadata, encryptedFile, err := encryptFileCBC(&encMat, inputFile, 0, "")
 	if err != nil {
 		t.Error(err)
 	}
 	defer os.Remove(encryptedFile)
-	decryptedFile, err := decryptFile(metadata, &encMat, encryptedFile, 0, "")
+	decryptedFile, err := decryptFileCBC(metadata, &encMat, encryptedFile, 0, "")
 	if err != nil {
 		t.Error(err)
 	}
 	defer os.Remove(decryptedFile)
 
-	fd, _ = os.OpenFile(decryptedFile, os.O_RDONLY, os.ModePerm)
+	fd, err = os.Open(decryptedFile)
+	if err != nil {
+		t.Error(err)
+	}
 	defer fd.Close()
-	content, _ := ioutil.ReadAll(fd)
+	content, err := io.ReadAll(fd)
+	if err != nil {
+		t.Error(err)
+	}
 	if string(content) != data {
 		t.Fatalf("data did not match content. expected: %v, got: %v", data, string(content))
 	}
 }
 
-func TestEncryptDecryptLargeFile(t *testing.T) {
+func TestEncryptDecryptFilePadding(t *testing.T) {
 	encMat := snowflakeFileEncryption{
 		"ztke8tIdVt1zmlQIZm0BMA==",
 		"123873c7-3a66-40c4-ab89-e3722fbccce1",
 		3112,
 	}
+
+	testcases := []encryptDecryptTestFile{
+		// File size is a multiple of 65536 bytes (chunkSize)
+		{numberOfBytesInEachRow: 8, numberOfLines: 16384},
+		{numberOfBytesInEachRow: 16, numberOfLines: 4096},
+		// File size is not a multiple of 65536 bytes (chunkSize)
+		{numberOfBytesInEachRow: 8, numberOfLines: 10240},
+		{numberOfBytesInEachRow: 16, numberOfLines: 6144},
+		// The second chunk's size is a multiple of 16 bytes (aes.BlockSize)
+		{numberOfBytesInEachRow: 16, numberOfLines: 4097},
+		// The second chunk's size is not a multiple of 16 bytes (aes.BlockSize)
+		{numberOfBytesInEachRow: 12, numberOfLines: 5462},
+		{numberOfBytesInEachRow: 10, numberOfLines: 6556},
+	}
+
+	for _, test := range testcases {
+		t.Run(fmt.Sprintf("%v_%v", test.numberOfBytesInEachRow, test.numberOfLines), func(t *testing.T) {
+			tmpDir, err := generateKLinesOfNByteRows(test.numberOfLines, test.numberOfBytesInEachRow, t.TempDir())
+			if err != nil {
+				t.Error(err)
+			}
+
+			encryptDecryptFile(t, encMat, test.numberOfLines, tmpDir)
+		})
+	}
+}
+
+func TestEncryptDecryptLargeFileCBC(t *testing.T) {
+	encMat := snowflakeFileEncryption{
+		"ztke8tIdVt1zmlQIZm0BMA==",
+		"123873c7-3a66-40c4-ab89-e3722fbccce1",
+		3112,
+	}
+
 	numberOfFiles := 1
 	numberOfLines := 10000
-	tmpDir, _ := ioutil.TempDir("", "data")
-	tmpDir = generateKLinesOfNFiles(numberOfLines, numberOfFiles, false, tmpDir)
-	defer os.RemoveAll(tmpDir)
+	tmpDir, err := generateKLinesOfNFiles(numberOfLines, numberOfFiles, false, t.TempDir())
+	if err != nil {
+		t.Error(err)
+	}
+
+	encryptDecryptFile(t, encMat, numberOfLines, tmpDir)
+}
+
+func encryptDecryptFile(t *testing.T, encMat snowflakeFileEncryption, expected int, tmpDir string) {
 	files, err := filepath.Glob(filepath.Join(tmpDir, "file*"))
 	if err != nil {
 		t.Error(err)
 	}
 	inputFile := files[0]
 
-	metadata, encryptedFile, err := encryptFile(&encMat, inputFile, 0, tmpDir)
+	metadata, encryptedFile, err := encryptFileCBC(&encMat, inputFile, 0, tmpDir)
 	if err != nil {
 		t.Error(err)
 	}
 	defer os.Remove(encryptedFile)
-	decryptedFile, err := decryptFile(metadata, &encMat, encryptedFile, 0, tmpDir)
+	decryptedFile, err := decryptFileCBC(metadata, &encMat, encryptedFile, 0, tmpDir)
 	if err != nil {
 		t.Error(err)
 	}
 	defer os.Remove(decryptedFile)
 
 	cnt := 0
-	fd, _ := os.OpenFile(decryptedFile, os.O_RDONLY, os.ModePerm)
+	fd, err := os.Open(decryptedFile)
+	if err != nil {
+		t.Error(err)
+	}
+	defer fd.Close()
+
 	scanner := bufio.NewScanner(fd)
 	for scanner.Scan() {
 		cnt++
@@ -95,18 +151,36 @@ func TestEncryptDecryptLargeFile(t *testing.T) {
 	if err = scanner.Err(); err != nil {
 		t.Error(err)
 	}
-	if cnt != numberOfLines && cnt != numberOfLines+1 {
-		t.Fatalf("incorrect number of lines. expected: %v, got: %v", numberOfLines, cnt)
+	if cnt != expected {
+		t.Fatalf("incorrect number of lines. expected: %v, got: %v", expected, cnt)
 	}
 }
 
-func generateKLinesOfNFiles(k int, n int, compress bool, tmpDir string) string {
-	if tmpDir == "" {
-		tmpDir, _ = ioutil.TempDir(tmpDir, "data")
+func generateKLinesOfNByteRows(numLines int, numBytes int, tmpDir string) (string, error) {
+	fname := path.Join(tmpDir, "file"+strconv.FormatInt(int64(numLines*numBytes), 10))
+	f, err := os.Create(fname)
+	if err != nil {
+		return "", err
 	}
+
+	for j := 0; j < numLines; j++ {
+		str := randomString(numBytes - 1) // \n is the last character
+		rec := fmt.Sprintf("%v\n", str)
+		if _, err = f.Write([]byte(rec)); err != nil {
+			return "", err
+		}
+	}
+	err = f.Close()
+	return tmpDir, err
+}
+
+func generateKLinesOfNFiles(k int, n int, compress bool, tmpDir string) (string, error) {
 	for i := 0; i < n; i++ {
 		fname := path.Join(tmpDir, "file"+strconv.FormatInt(int64(i), 10))
-		f, _ := os.OpenFile(fname, os.O_CREATE|os.O_WRONLY, os.ModePerm)
+		f, err := os.Create(fname)
+		if err != nil {
+			return "", err
+		}
 		for j := 0; j < k; j++ {
 			num := rand.Float64() * 10000
 			min := time.Date(1970, 1, 0, 0, 0, 0, 0, time.UTC).Unix()
@@ -126,28 +200,91 @@ func generateKLinesOfNFiles(k int, n int, compress bool, tmpDir string) string {
 			pct := rand.Float64() * 1000
 			ratio := fmt.Sprintf("%.2f", rand.Float64()*1000)
 			rec := fmt.Sprintf("%v,%v,%v,%v,%v,%v,%v,%v\n", num, dt, ts, tsltz, tsntz, tstz, pct, ratio)
-			f.Write([]byte(rec))
+			if _, err = f.Write([]byte(rec)); err != nil {
+				return "", err
+			}
 		}
-		f.Close()
+		if err = f.Close(); err != nil {
+			return "", err
+		}
 		if compress {
 			if !isWindows {
 				gzipCmd := exec.Command("gzip", filepath.Join(tmpDir, "file"+strconv.FormatInt(int64(i), 10)))
-				gzipOut, _ := gzipCmd.StdoutPipe()
-				gzipErr, _ := gzipCmd.StderrPipe()
-				gzipCmd.Start()
-				ioutil.ReadAll(gzipOut)
-				ioutil.ReadAll(gzipErr)
-				gzipCmd.Wait()
+				gzipOut, err := gzipCmd.StdoutPipe()
+				if err != nil {
+					return "", err
+				}
+				gzipErr, err := gzipCmd.StderrPipe()
+				if err != nil {
+					return "", err
+				}
+				if err = gzipCmd.Start(); err != nil {
+					return "", err
+				}
+				if _, err = io.ReadAll(gzipOut); err != nil {
+					return "", err
+				}
+				if _, err = io.ReadAll(gzipErr); err != nil {
+					return "", err
+				}
+				if err = gzipCmd.Wait(); err != nil {
+					return "", err
+				}
 			} else {
-				fOut, _ := os.OpenFile(fname+".gz", os.O_CREATE|os.O_WRONLY, os.ModePerm)
+				fOut, err := os.Create(fname + ".gz")
+				if err != nil {
+					return "", err
+				}
 				w := gzip.NewWriter(fOut)
-				fIn, _ := os.OpenFile(fname, os.O_RDONLY, os.ModePerm)
-				if _, err := io.Copy(w, fIn); err != nil {
-					return ""
+				fIn, err := os.Open(fname)
+				if err != nil {
+					return "", err
+				}
+				if _, err = io.Copy(w, fIn); err != nil {
+					return "", err
 				}
 				w.Close()
+				fOut.Close()
+				fIn.Close()
 			}
 		}
 	}
-	return tmpDir
+	return tmpDir, nil
+}
+
+func TestEncryptDecryptGCM(t *testing.T) {
+	input := []byte("abc")
+	iv := []byte("ab1234567890")      // pragma: allowlist secret
+	key := []byte("1234567890abcdef") // pragma: allowlist secret
+	encrypted, err := encryptGCM(iv, input, key, nil)
+	assertNilF(t, err)
+	assertEqualE(t, base64.StdEncoding.EncodeToString(encrypted), "iG+lT4o27hkzj3kblYRzQikLVQ==")
+
+	decrypted, err := decryptGCM(iv, encrypted, key, nil)
+	assertNilF(t, err)
+	assertDeepEqualE(t, decrypted, input)
+}
+
+func TestEncryptDecryptFileGCM(t *testing.T) {
+	tmpDir := os.TempDir()
+	tempFile, err := os.CreateTemp(tmpDir, "gcm")
+	assertNilF(t, err)
+	_, err = tempFile.Write([]byte("abc"))
+	assertNilF(t, err)
+
+	sfe := &snowflakeFileEncryption{
+		QueryStageMasterKey: "YWJjZGVmMTIzNDU2Nzg5MA==",
+		QueryID:             "unused",
+		SMKID:               123,
+	}
+	meta, encryptedFileName, err := encryptFileGCM(sfe, tempFile.Name(), tmpDir)
+	assertNilF(t, err)
+
+	decryptedFileName, err := decryptFileGCM(meta, sfe, encryptedFileName, tmpDir)
+	assertNilF(t, err)
+
+	fileContent, err := os.ReadFile(decryptedFileName)
+	assertNilF(t, err)
+
+	assertEqualE(t, string(fileContent), "abc")
 }
