@@ -1,4 +1,4 @@
-// Copyright (c) 2021 Snowflake Computing Inc. All right reserved.
+// Copyright (c) 2021-2022 Snowflake Computing Inc. All rights reserved.
 
 package gosnowflake
 
@@ -8,7 +8,6 @@ import (
 	"crypto/sha256"
 	"encoding/base64"
 	"io"
-	"io/ioutil"
 	"net/url"
 	"os"
 	usr "os/user"
@@ -19,58 +18,109 @@ import (
 type snowflakeFileUtil struct {
 }
 
-func (util *snowflakeFileUtil) compressFileWithGzipFromStream(srcStream **bytes.Buffer) (*bytes.Buffer, int) {
+const (
+	fileChunkSize                 = 16 * 4 * 1024
+	readWriteFileMode os.FileMode = 0666
+)
+
+func (util *snowflakeFileUtil) compressFileWithGzipFromStream(srcStream **bytes.Buffer) (*bytes.Buffer, int, error) {
 	r := getReaderFromBuffer(srcStream)
-	buf, _ := ioutil.ReadAll(r)
+	buf, err := io.ReadAll(r)
+	if err != nil {
+		return nil, -1, err
+	}
 	var c bytes.Buffer
 	w := gzip.NewWriter(&c)
-	w.Write(buf) // write buf to gzip writer
-	w.Close()
-	return &c, c.Len()
+	if _, err := w.Write(buf); err != nil { // write buf to gzip writer
+		return nil, -1, err
+	}
+	if err := w.Close(); err != nil {
+		return nil, -1, err
+	}
+	return &c, c.Len(), nil
 }
 
-func (util *snowflakeFileUtil) compressFileWithGzip(fileName string, tmpDir string) (string, int64) {
+func (util *snowflakeFileUtil) compressFileWithGzip(fileName string, tmpDir string) (gzipFileName string, size int64, err error) {
 	basename := baseName(fileName)
-	gzipFileName := filepath.Join(tmpDir, basename+"_c.gz")
+	gzipFileName = filepath.Join(tmpDir, basename+"_c.gz")
 
-	fr, _ := os.OpenFile(fileName, os.O_RDONLY, os.ModePerm)
-	defer fr.Close()
-	fw, _ := os.OpenFile(gzipFileName, os.O_WRONLY|os.O_CREATE, os.ModePerm)
+	fr, err := os.Open(fileName)
+	if err != nil {
+		return "", -1, err
+	}
+	defer func() {
+		if tmpErr := fr.Close(); tmpErr != nil {
+			err = tmpErr
+		}
+	}()
+	fw, err := os.OpenFile(gzipFileName, os.O_WRONLY|os.O_CREATE, readWriteFileMode)
+	if err != nil {
+		return "", -1, err
+	}
 	gzw := gzip.NewWriter(fw)
-	defer gzw.Close()
-	io.Copy(gzw, fr)
+	defer func() {
+		if tmpErr := gzw.Close(); tmpErr != nil {
+			err = tmpErr
+		}
+	}()
+	_, err = io.Copy(gzw, fr)
+	if err != nil {
+		return "", -1, err
+	}
 
-	stat, _ := os.Stat(gzipFileName)
-	return gzipFileName, stat.Size()
+	stat, err := os.Stat(gzipFileName)
+	if err != nil {
+		return "", -1, err
+	}
+	return gzipFileName, stat.Size(), err
 }
 
-func (util *snowflakeFileUtil) getDigestAndSize(src **bytes.Buffer) (string, int64) {
-	chunkSize := 16 * 4 * 1024
+func (util *snowflakeFileUtil) getDigestAndSizeForStream(stream **bytes.Buffer) (string, int64, error) {
 	m := sha256.New()
-	r := getReaderFromBuffer(src)
+	r := getReaderFromBuffer(stream)
+	chunk := make([]byte, fileChunkSize)
+
 	for {
-		chunk := make([]byte, chunkSize)
 		n, err := r.Read(chunk)
-		if n == 0 || err != nil {
+		if err == io.EOF {
 			break
+		} else if err != nil {
+			return "", 0, err
 		}
 		m.Write(chunk[:n])
 	}
-	return base64.StdEncoding.EncodeToString(m.Sum(nil)), int64((*src).Len())
+	return base64.StdEncoding.EncodeToString(m.Sum(nil)), int64((*stream).Len()), nil
 }
 
-func (util *snowflakeFileUtil) getDigestAndSizeForStream(stream **bytes.Buffer) (string, int64) {
-	return util.getDigestAndSize(stream)
-}
-
-func (util *snowflakeFileUtil) getDigestAndSizeForFile(fileName string) (string, int64, error) {
-	src, err := ioutil.ReadFile(fileName)
+func (util *snowflakeFileUtil) getDigestAndSizeForFile(fileName string) (digest string, size int64, err error) {
+	f, err := os.Open(fileName)
 	if err != nil {
 		return "", 0, err
 	}
-	buf := bytes.NewBuffer(src)
-	digest, size := util.getDigestAndSize(&buf)
-	return digest, size, err
+	defer func() {
+		if tmpErr := f.Close(); tmpErr != nil {
+			err = tmpErr
+		}
+	}()
+
+	var total int64
+	m := sha256.New()
+	chunk := make([]byte, fileChunkSize)
+
+	for {
+		n, err := f.Read(chunk)
+		if err == io.EOF {
+			break
+		} else if err != nil {
+			return "", 0, err
+		}
+		total += int64(n)
+		m.Write(chunk[:n])
+	}
+	if _, err = f.Seek(0, io.SeekStart); err != nil {
+		return "", -1, err
+	}
+	return base64.StdEncoding.EncodeToString(m.Sum(nil)), total, err
 }
 
 // file metadata for PUT/GET
@@ -108,6 +158,9 @@ type fileMetadata struct {
 	srcStream     *bytes.Buffer
 	realSrcStream *bytes.Buffer
 
+	/* streaming GET */
+	dstStream *bytes.Buffer
+
 	/* GCS */
 	presignedURL                *url.URL
 	gcsFileHeaderDigest         string
@@ -115,8 +168,11 @@ type fileMetadata struct {
 	gcsFileHeaderEncryptionMeta *encryptMetadata
 
 	/* mock */
-	mockUploader s3UploadAPI
-	mockHeader   s3HeaderAPI
+	mockUploader    s3UploadAPI
+	mockDownloader  s3DownloadAPI
+	mockHeader      s3HeaderAPI
+	mockGcsClient   gcsAPI
+	mockAzureClient azureAPI
 }
 
 type fileTransferResultType struct {
@@ -157,15 +213,18 @@ func baseName(path string) string {
 }
 
 // expandUser returns the argument with an initial component of ~
-func expandUser(path string) string {
-	usr, _ := usr.Current()
+func expandUser(path string) (string, error) {
+	usr, err := usr.Current()
+	if err != nil {
+		return "", err
+	}
 	dir := usr.HomeDir
 	if path == "~" {
 		path = dir
 	} else if strings.HasPrefix(path, "~/") {
 		path = filepath.Join(dir, path[2:])
 	}
-	return path
+	return path, nil
 }
 
 // getDirectory retrieves the current working directory
