@@ -3,7 +3,6 @@ package gosnowflake
 import (
 	"context"
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strconv"
@@ -15,33 +14,37 @@ func TestPutGetFileSmallDataViaUserStage(t *testing.T) {
 	if os.Getenv("AWS_ACCESS_KEY_ID") == "" {
 		t.Skip("this test requires to change the internal parameter")
 	}
-	putGetUserStage(t, "", 5, 1, false)
+	putGetUserStage(t, 5, 1, false)
 }
 
 func TestPutGetStreamSmallDataViaUserStage(t *testing.T) {
 	if os.Getenv("AWS_ACCESS_KEY_ID") == "" {
 		t.Skip("this test requires to change the internal parameter")
 	}
-	putGetUserStage(t, "", 1, 1, true)
+	putGetUserStage(t, 1, 1, true)
 }
 
-func putGetUserStage(t *testing.T, tmpDir string, numberOfFiles int, numberOfLines int, isStream bool) {
+func putGetUserStage(t *testing.T, numberOfFiles int, numberOfLines int, isStream bool) {
 	if os.Getenv("AWS_SECRET_ACCESS_KEY") == "" {
 		t.Fatal("no aws secret access key found")
 	}
-	tmpDir, _ = ioutil.TempDir(tmpDir, "data")
-	tmpDir = generateKLinesOfNFiles(numberOfLines, numberOfFiles, false, tmpDir)
-	defer os.RemoveAll(tmpDir)
+	tmpDir, err := generateKLinesOfNFiles(numberOfLines, numberOfFiles, false, t.TempDir())
+	if err != nil {
+		t.Error(err)
+	}
 	var files string
 	if isStream {
-		list, _ := ioutil.ReadDir(tmpDir)
+		list, err := os.ReadDir(tmpDir)
+		if err != nil {
+			t.Error(err)
+		}
 		file := list[0].Name()
 		files = filepath.Join(tmpDir, file)
 	} else {
 		files = filepath.Join(tmpDir, "file*")
 	}
 
-	runTests(t, dsn, func(dbt *DBTest) {
+	runDBTest(t, func(dbt *DBTest) {
 		stageName := fmt.Sprintf("%v_stage_%v_%v", dbname, numberOfFiles, numberOfLines)
 		sqlText := `create or replace table %v (aa int, dt date, ts timestamp,
 			tsltz timestamp_ltz, tsntz timestamp_ntz, tstz timestamp_tz,
@@ -49,7 +52,7 @@ func putGetUserStage(t *testing.T, tmpDir string, numberOfFiles int, numberOfLin
 		dbt.mustExec(fmt.Sprintf(sqlText, dbname))
 		userBucket := os.Getenv("SF_AWS_USER_BUCKET")
 		if userBucket == "" {
-			userBucket = fmt.Sprintf("sfc-dev1-regression/%v/reg", user)
+			userBucket = fmt.Sprintf("sfc-eng-regression/%v/reg", username)
 		}
 		sqlText = `create or replace stage %v url='s3://%v}/%v-%v-%v'
 			credentials = (AWS_KEY_ID='%v' AWS_SECRET_KEY='%v')`
@@ -61,7 +64,7 @@ func putGetUserStage(t *testing.T, tmpDir string, numberOfFiles int, numberOfLin
 		dbt.mustExec("rm @" + stageName)
 		var fs *os.File
 		if isStream {
-			fs, _ = os.OpenFile(files, os.O_RDONLY, os.ModePerm)
+			fs, _ = os.Open(files)
 			dbt.mustExecContext(WithFileStream(context.Background(), fs),
 				fmt.Sprintf("put 'file://%v' @%v", strings.ReplaceAll(
 					files, "\\", "\\\\"), stageName))
@@ -79,11 +82,17 @@ func putGetUserStage(t *testing.T, tmpDir string, numberOfFiles int, numberOfLin
 		dbt.mustExec(fmt.Sprintf("copy into %v from @%v", dbname, stageName))
 
 		rows := dbt.mustQuery("select count(*) from " + dbname)
+		defer func() {
+			assertNilF(t, rows.Close())
+		}()
 		var cnt string
 		if rows.Next() {
-			rows.Scan(&cnt)
+			assertNilF(t, rows.Scan(&cnt))
 		}
-		count, _ := strconv.Atoi(cnt)
+		count, err := strconv.Atoi(cnt)
+		if err != nil {
+			t.Error(err)
+		}
 		if count != numberOfFiles*numberOfLines {
 			t.Errorf("count did not match expected number. count: %v, expected: %v", count, numberOfFiles*numberOfLines)
 		}
@@ -91,7 +100,7 @@ func putGetUserStage(t *testing.T, tmpDir string, numberOfFiles int, numberOfLin
 }
 
 func TestPutLoadFromUserStage(t *testing.T) {
-	runTests(t, dsn, func(dbt *DBTest) {
+	runDBTest(t, func(dbt *DBTest) {
 		data, err := createTestData(dbt)
 		if err != nil {
 			t.Skip("snowflake admin account not accessible")
@@ -123,6 +132,9 @@ func TestPutLoadFromUserStage(t *testing.T) {
 		rows := dbt.mustQuery(fmt.Sprintf(`copy into gotest_putget_t2 from @%v
 			file_format = (field_delimiter = '|' error_on_column_count_mismatch
 			=false) purge=true`, data.stage))
+		defer func() {
+			assertNilF(t, rows.Close())
+		}()
 		var s0, s1, s2, s3, s4, s5 string
 		var s6, s7, s8, s9 interface{}
 		orders100 := fmt.Sprintf("s3://%v/%v/orders_100.csv.gz",
@@ -130,7 +142,7 @@ func TestPutLoadFromUserStage(t *testing.T) {
 		orders101 := fmt.Sprintf("s3://%v/%v/orders_101.csv.gz",
 			data.userBucket, data.stage)
 		for rows.Next() {
-			rows.Scan(&s0, &s1, &s2, &s3, &s4, &s5, &s6, &s7, &s8, &s9)
+			assertNilF(t, rows.Scan(&s0, &s1, &s2, &s3, &s4, &s5, &s6, &s7, &s8, &s9))
 			if s0 != orders100 && s0 != orders101 {
 				t.Fatalf("copy did not load orders files. got: %v", s0)
 			}
