@@ -1,10 +1,11 @@
-// Copyright (c) 2021 Snowflake Computing Inc. All rights reserved.
+// Copyright (c) 2021-2022 Snowflake Computing Inc. All rights reserved.
 
 package gosnowflake
 
 import (
 	"bytes"
 	"context"
+	"database/sql"
 	"database/sql/driver"
 	"fmt"
 	"reflect"
@@ -13,8 +14,8 @@ import (
 )
 
 const (
-	bindStageName   = "SYSTEM$BIND"
-	createStageStmt = "CREATE OR REPLACE TEMPORARY STAGE " + bindStageName +
+	bindStageName            = "SYSTEM$BIND"
+	createTemporaryStageStmt = "CREATE OR REPLACE TEMPORARY STAGE " + bindStageName +
 		" file_format=" + "(type=csv field_optionally_enclosed_by='\"')"
 
 	// size (in bytes) of max input stream (10MB default) as per JDBC specs
@@ -29,12 +30,26 @@ type bindUploader struct {
 	arrayBindStage string
 }
 
+type bindingSchema struct {
+	Typ      string          `json:"type"`
+	Nullable bool            `json:"nullable"`
+	Fields   []fieldMetadata `json:"fields"`
+}
+
+type bindingValue struct {
+	value  *string
+	format string
+	schema *bindingSchema
+}
+
 func (bu *bindUploader) upload(bindings []driver.NamedValue) (*execResponse, error) {
-	bindingRows, _ := bu.buildRowsAsBytes(bindings)
+	bindingRows, err := bu.buildRowsAsBytes(bindings)
+	if err != nil {
+		return nil, err
+	}
 	startIdx, numBytes, rowNum := 0, 0, 0
 	bu.fileCount = 0
 	var data *execResponse
-	var err error
 	for rowNum < len(bindingRows) {
 		for numBytes < inputStreamBufferSize && rowNum < len(bindingRows) {
 			numBytes += len(bindingRows[rowNum])
@@ -89,7 +104,7 @@ func (bu *bindUploader) createStageIfNeeded() error {
 	if bu.arrayBindStage != "" {
 		return nil
 	}
-	data, err := bu.sc.exec(bu.ctx, createStageStmt, false, false, false, []driver.NamedValue{})
+	data, err := bu.sc.exec(bu.ctx, createTemporaryStageStmt, false, false, false, []driver.NamedValue{})
 	if err != nil {
 		newThreshold := "0"
 		bu.sc.cfg.Params[sessionArrayBindStageThreshold] = &newThreshold
@@ -103,7 +118,7 @@ func (bu *bindUploader) createStageIfNeeded() error {
 		return (&SnowflakeError{
 			Number:   code,
 			SQLState: data.Data.SQLState,
-			Message:  err.Error(),
+			Message:  data.Message,
 			QueryID:  data.Data.QueryID,
 		}).exceptionTelemetry(bu.sc)
 	}
@@ -121,19 +136,29 @@ func (bu *bindUploader) buildRowsAsBytes(columns []driver.NamedValue) ([][]byte,
 		}).exceptionTelemetry(bu.sc)
 	}
 
-	_, column := snowflakeArrayToString(&columns[0], true)
+	_, column, err := snowflakeArrayToString(&columns[0], true)
+	if err != nil {
+		return nil, err
+	}
 	numRows := len(column)
 	csvRows := make([][]byte, 0)
-	rows := make([][]string, 0)
+	rows := make([][]interface{}, 0)
 	for rowIdx := 0; rowIdx < numRows; rowIdx++ {
-		rows = append(rows, make([]string, numColumns))
+		rows = append(rows, make([]interface{}, numColumns))
 	}
 
 	for rowIdx := 0; rowIdx < numRows; rowIdx++ {
-		rows[rowIdx][0] = *column[rowIdx]
+		if column[rowIdx] == nil {
+			rows[rowIdx][0] = column[rowIdx]
+		} else {
+			rows[rowIdx][0] = *column[rowIdx]
+		}
 	}
 	for colIdx := 1; colIdx < numColumns; colIdx++ {
-		_, column = snowflakeArrayToString(&columns[colIdx], true)
+		_, column, err = snowflakeArrayToString(&columns[colIdx], true)
+		if err != nil {
+			return nil, err
+		}
 		iNumRows := len(column)
 		if iNumRows != numRows {
 			return nil, (&SnowflakeError{
@@ -143,7 +168,12 @@ func (bu *bindUploader) buildRowsAsBytes(columns []driver.NamedValue) ([][]byte,
 			}).exceptionTelemetry(bu.sc)
 		}
 		for rowIdx := 0; rowIdx < numRows; rowIdx++ {
-			rows[rowIdx][colIdx] = *column[rowIdx] // length of column = number of rows
+			// length of column = number of rows
+			if column[rowIdx] == nil {
+				rows[rowIdx][colIdx] = column[rowIdx]
+			} else {
+				rows[rowIdx][colIdx] = *column[rowIdx]
+			}
 		}
 	}
 	for _, row := range rows {
@@ -152,25 +182,67 @@ func (bu *bindUploader) buildRowsAsBytes(columns []driver.NamedValue) ([][]byte,
 	return csvRows, nil
 }
 
-func (bu *bindUploader) createCSVRecord(data []string) []byte {
+func (bu *bindUploader) createCSVRecord(data []interface{}) []byte {
 	var b strings.Builder
 	b.Grow(1024)
 	for i := 0; i < len(data); i++ {
 		if i > 0 {
 			b.WriteString(",")
 		}
-		b.WriteString(escapeForCSV(data[i]))
+		value, ok := data[i].(string)
+		if ok {
+			b.WriteString(escapeForCSV(value))
+		} else if !reflect.ValueOf(data[i]).IsNil() {
+			logger.WithContext(bu.ctx).Debugf("Cannot convert value to string in createCSVRecord. value: %v", data[i])
+		}
 	}
 	b.WriteString("\n")
 	return []byte(b.String())
 }
 
-func getBindValues(bindings []driver.NamedValue) (map[string]execBindParameter, error) {
+func (sc *snowflakeConn) processBindings(
+	ctx context.Context,
+	bindings []driver.NamedValue,
+	describeOnly bool,
+	requestID UUID,
+	req *execRequest) error {
+	arrayBindThreshold := sc.getArrayBindStageThreshold()
+	numBinds, err := arrayBindValueCount(bindings)
+	if err != nil {
+		return err
+	}
+	if 0 < arrayBindThreshold && arrayBindThreshold <= numBinds && !describeOnly && isArrayBind(bindings) {
+		uploader := bindUploader{
+			sc:        sc,
+			ctx:       ctx,
+			stagePath: "@" + bindStageName + "/" + requestID.String(),
+		}
+		_, err := uploader.upload(bindings)
+		if err != nil {
+			return err
+		}
+		req.Bindings = nil
+		req.BindStage = uploader.stagePath
+	} else {
+		req.Bindings, err = getBindValues(bindings, sc.cfg.Params)
+		if err != nil {
+			return err
+		}
+		req.BindStage = ""
+	}
+	return nil
+}
+
+func getBindValues(bindings []driver.NamedValue, params map[string]*string) (map[string]execBindParameter, error) {
 	tsmode := timestampNtzType
 	idx := 1
 	var err error
 	bindValues := make(map[string]execBindParameter, len(bindings))
 	for _, binding := range bindings {
+		if tnt, ok := binding.Value.(TypedNullTime); ok {
+			tsmode = convertTzTypeToSnowflakeType(tnt.TzType)
+			binding.Value = tnt.Time
+		}
 		t := goTypeToSnowflake(binding.Value, tsmode)
 		if t == changeType {
 			tsmode, err = dataTypeMode(binding.Value)
@@ -179,21 +251,32 @@ func getBindValues(bindings []driver.NamedValue) (map[string]execBindParameter, 
 			}
 		} else {
 			var val interface{}
+			var bv bindingValue
 			if t == sliceType {
 				// retrieve array binding data
-				t, val = snowflakeArrayToString(&binding, false)
+				t, val, err = snowflakeArrayToString(&binding, false)
+				if err != nil {
+					return nil, err
+				}
 			} else {
-				val, err = valueToString(binding.Value, tsmode)
+				bv, err = valueToString(binding.Value, tsmode, params)
+				val = bv.value
 				if err != nil {
 					return nil, err
 				}
 			}
 			if t == nullType || t == unSupportedType {
 				t = textType // if null or not supported, pass to GS as text
+			} else if t == nilObjectType || t == mapType || t == nilMapType {
+				t = objectType
+			} else if t == nilArrayType {
+				t = arrayType
 			}
-			bindValues[strconv.Itoa(idx)] = execBindParameter{
-				Type:  t.String(),
-				Value: val,
+			bindValues[bindingName(binding, idx)] = execBindParameter{
+				Type:   t.String(),
+				Value:  val,
+				Format: bv.format,
+				Schema: bv.schema,
 			}
 			idx++
 		}
@@ -201,12 +284,23 @@ func getBindValues(bindings []driver.NamedValue) (map[string]execBindParameter, 
 	return bindValues, nil
 }
 
-func arrayBindValueCount(bindValues []driver.NamedValue) int {
-	if !isArrayBind(bindValues) {
-		return 0
+func bindingName(nv driver.NamedValue, idx int) string {
+	if nv.Name != "" {
+		return nv.Name
 	}
-	_, arr := snowflakeArrayToString(&bindValues[0], false)
-	return len(bindValues) * len(arr)
+	return strconv.Itoa(idx)
+}
+
+func arrayBindValueCount(bindValues []driver.NamedValue) (int, error) {
+	if !isArrayBind(bindValues) {
+		return 0, nil
+	}
+	_, arr, err := snowflakeArrayToString(&bindValues[0], false)
+	if err != nil {
+		return 0, err
+	}
+
+	return len(bindValues) * len(arr), nil
 }
 
 func isArrayBind(bindings []driver.NamedValue) bool {
@@ -233,7 +327,10 @@ func supportedArrayBind(nv *driver.NamedValue) bool {
 		return true
 	case reflect.TypeOf([]uint8{}):
 		// internal binding ts mode
-		val, _ := nv.Value.([]uint8)
+		val, ok := nv.Value.([]uint8)
+		if !ok {
+			return ok
+		}
 		if len(val) == 0 {
 			return true // for null binds
 		}
@@ -243,6 +340,38 @@ func supportedArrayBind(nv *driver.NamedValue) bool {
 		return false
 	default:
 		// TODO SNOW-176486 variant, object, array
+
+		// Support for bulk array binding insertion using []interface{}
+		if isInterfaceArrayBinding(nv.Value) {
+			return true
+		}
 		return false
 	}
+}
+
+func supportedNullBind(nv *driver.NamedValue) bool {
+	switch reflect.TypeOf(nv.Value) {
+	case reflect.TypeOf(sql.NullString{}), reflect.TypeOf(sql.NullInt64{}),
+		reflect.TypeOf(sql.NullBool{}), reflect.TypeOf(sql.NullFloat64{}), reflect.TypeOf(TypedNullTime{}):
+		return true
+	}
+	return false
+}
+
+func supportedStructuredObjectWriterBind(nv *driver.NamedValue) bool {
+	if _, ok := nv.Value.(StructuredObjectWriter); ok {
+		return true
+	}
+	_, ok := nv.Value.(reflect.Type)
+	return ok
+}
+
+func supportedStructuredArrayBind(nv *driver.NamedValue) bool {
+	typ := reflect.TypeOf(nv.Value)
+	return typ != nil && (typ.Kind() == reflect.Array || typ.Kind() == reflect.Slice)
+}
+
+func supportedStructuredMapBind(nv *driver.NamedValue) bool {
+	typ := reflect.TypeOf(nv.Value)
+	return typ != nil && (typ.Kind() == reflect.Map || typ == reflect.TypeOf(NilMapTypes{}))
 }
