@@ -1,3 +1,5 @@
+// Copyright (c) 2019-2022 Snowflake Computing Inc. All rights reserved.
+
 package gosnowflake
 
 import (
@@ -24,6 +26,18 @@ func unsetenv(k string) {
 	}
 }
 
+// deleteOCSPCacheFile deletes the OCSP response cache file
+func deleteOCSPCacheFile() {
+	os.Remove(cacheFileName)
+}
+
+// deleteOCSPCacheAll deletes all entries in the OCSP response cache on memory
+func deleteOCSPCacheAll() {
+	ocspResponseCacheLock.Lock()
+	defer ocspResponseCacheLock.Unlock()
+	ocspResponseCache = make(map[certIDKey]*certCacheValue)
+}
+
 func cleanup() {
 	deleteOCSPCacheFile()
 	deleteOCSPCacheAll()
@@ -35,6 +49,7 @@ func cleanup() {
 	unsetenv(ocspTestResponderTimeoutEnv)
 	unsetenv(ocspTestResponderURLEnv)
 	unsetenv(ocspTestNoOCSPURLEnv)
+	unsetenv(ocspRetryURLEnv)
 	unsetenv(cacheDirEnv)
 }
 
@@ -65,13 +80,20 @@ func TestOCSPFailOpen(t *testing.T) {
 	if err = db.Ping(); err == nil {
 		t.Fatalf("should fail to ping. %v", testURL)
 	}
+	if strings.Contains(err.Error(), "HTTP Status: 513. Hanging?") {
+		return
+	}
 	driverErr, ok := err.(*SnowflakeError)
 	if !ok {
 		t.Fatalf("failed to extract error SnowflakeError: %v", err)
 	}
-	if driverErr.Number != ErrCodeFailedToConnect {
+	if isFailToConnectOrAuthErr(driverErr) {
 		t.Fatalf("should failed to connect %v", err)
 	}
+}
+
+func isFailToConnectOrAuthErr(driverErr *SnowflakeError) bool {
+	return driverErr.Number != ErrCodeFailedToConnect && driverErr.Number != ErrFailedToAuth
 }
 
 // TestOCSPFailOpenWithoutFileCache ensures no file cache is used.
@@ -103,11 +125,14 @@ func TestOCSPFailOpenWithoutFileCache(t *testing.T) {
 	if err = db.Ping(); err == nil {
 		t.Fatalf("should fail to ping. %v", testURL)
 	}
+	if strings.Contains(err.Error(), "HTTP Status: 513. Hanging?") {
+		return
+	}
 	driverErr, ok := err.(*SnowflakeError)
 	if !ok {
 		t.Fatalf("failed to extract error SnowflakeError: %v", err)
 	}
-	if driverErr.Number != ErrCodeFailedToConnect {
+	if isFailToConnectOrAuthErr(driverErr) {
 		t.Fatalf("should failed to connect %v", err)
 	}
 }
@@ -142,11 +167,14 @@ func TestOCSPFailOpenValidityError(t *testing.T) {
 	if err = db.Ping(); err == nil {
 		t.Fatalf("should fail to ping. %v", testURL)
 	}
+	if strings.Contains(err.Error(), "HTTP Status: 513. Hanging?") {
+		return
+	}
 	driverErr, ok := err.(*SnowflakeError)
 	if !ok {
 		t.Fatalf("failed to extract error SnowflakeError: %v", err)
 	}
-	if driverErr.Number != ErrCodeFailedToConnect {
+	if isFailToConnectOrAuthErr(driverErr) {
 		t.Fatalf("should failed to connect %v", err)
 	}
 }
@@ -180,6 +208,9 @@ func TestOCSPFailClosedValidityError(t *testing.T) {
 	defer db.Close()
 	if err = db.Ping(); err == nil {
 		t.Fatalf("should fail to ping. %v", testURL)
+	}
+	if strings.Contains(err.Error(), "HTTP Status: 513. Hanging?") {
+		return
 	}
 	urlErr, ok := err.(*url.Error)
 	if !ok {
@@ -225,11 +256,14 @@ func TestOCSPFailOpenUnknownStatus(t *testing.T) {
 	if err = db.Ping(); err == nil {
 		t.Fatalf("should fail to ping. %v", testURL)
 	}
+	if strings.Contains(err.Error(), "HTTP Status: 513. Hanging?") {
+		return
+	}
 	driverErr, ok := err.(*SnowflakeError)
 	if !ok {
 		t.Fatalf("failed to extract error SnowflakeError: %v", err)
 	}
-	if driverErr.Number != ErrCodeFailedToConnect {
+	if isFailToConnectOrAuthErr(driverErr) {
 		t.Fatalf("should failed to connect %v", err)
 	}
 }
@@ -263,6 +297,9 @@ func TestOCSPFailClosedUnknownStatus(t *testing.T) {
 	defer db.Close()
 	if err = db.Ping(); err == nil {
 		t.Fatalf("should fail to ping. %v", testURL)
+	}
+	if strings.Contains(err.Error(), "HTTP Status: 513. Hanging?") {
+		return
 	}
 	urlErr, ok := err.(*url.Error)
 	if !ok {
@@ -309,6 +346,9 @@ func TestOCSPFailOpenRevokedStatus(t *testing.T) {
 	if err = db.Ping(); err == nil {
 		t.Fatalf("should fail to ping. %v", testURL)
 	}
+	if strings.Contains(err.Error(), "HTTP Status: 513. Hanging?") {
+		return
+	}
 	urlErr, ok := err.(*url.Error)
 	if !ok {
 		t.Fatalf("failed to extract error URL Error: %v", err)
@@ -354,6 +394,9 @@ func TestOCSPFailClosedRevokedStatus(t *testing.T) {
 	if err = db.Ping(); err == nil {
 		t.Fatalf("should fail to ping. %v", testURL)
 	}
+	if strings.Contains(err.Error(), "HTTP Status: 513. Hanging?") {
+		return
+	}
 	urlErr, ok := err.(*url.Error)
 	if !ok {
 		t.Fatalf("failed to extract error URL Error: %v", err)
@@ -373,7 +416,7 @@ func TestOCSPFailOpenCacheServerTimeout(t *testing.T) {
 	cleanup()
 	defer cleanup()
 
-	setenv(cacheServerURLEnv, "http://localhost:12345/hang")
+	setenv(cacheServerURLEnv, "http://localhost:12345/ocsp/hang")
 	setenv(ocspTestResponseCacheServerTimeoutEnv, "1000")
 
 	config := &Config{
@@ -398,11 +441,14 @@ func TestOCSPFailOpenCacheServerTimeout(t *testing.T) {
 	if err = db.Ping(); err == nil {
 		t.Fatalf("should fail to ping. %v", testURL)
 	}
+	if strings.Contains(err.Error(), "HTTP Status: 513. Hanging?") {
+		return
+	}
 	driverErr, ok := err.(*SnowflakeError)
 	if !ok {
 		t.Fatalf("failed to extract error SnowflakeError: %v", err)
 	}
-	if driverErr.Number != ErrCodeFailedToConnect {
+	if isFailToConnectOrAuthErr(driverErr) {
 		t.Fatalf("should failed to connect %v", err)
 	}
 }
@@ -412,7 +458,7 @@ func TestOCSPFailClosedCacheServerTimeout(t *testing.T) {
 	cleanup()
 	defer cleanup()
 
-	setenv(cacheServerURLEnv, "http://localhost:12345/hang")
+	setenv(cacheServerURLEnv, "http://localhost:12345/ocsp/hang")
 	setenv(ocspTestResponseCacheServerTimeoutEnv, "1000")
 
 	config := &Config{
@@ -437,12 +483,31 @@ func TestOCSPFailClosedCacheServerTimeout(t *testing.T) {
 	if err = db.Ping(); err == nil {
 		t.Fatalf("should fail to ping. %v", testURL)
 	}
-	driverErr, ok := err.(*SnowflakeError)
-	if !ok {
-		t.Fatalf("failed to extract error SnowflakeError: %v", err)
+	if err == nil {
+		t.Fatalf("should failed to connect. err:  %v", err)
 	}
-	if driverErr.Number != ErrCodeFailedToConnect {
-		t.Fatalf("should failed to connect %v", err)
+	if strings.Contains(err.Error(), "HTTP Status: 513. Hanging?") {
+		return
+	}
+
+	switch errType := err.(type) {
+	// Before Go 1.17
+	case *SnowflakeError:
+		driverErr, ok := err.(*SnowflakeError)
+		if !ok {
+			t.Fatalf("failed to extract error SnowflakeError: %v", err)
+		}
+		if isFailToConnectOrAuthErr(driverErr) {
+			t.Fatalf("should have failed to connect. err: %v", err)
+		}
+	// Go 1.18 and after rejects SHA-1 certificates, therefore a different error is returned (https://github.com/golang/go/issues/41682)
+	case *url.Error:
+		expectedErrMsg := "bad OCSP signature"
+		if !strings.Contains(err.Error(), expectedErrMsg) {
+			t.Fatalf("should have failed with bad OCSP signature. err:  %v", err)
+		}
+	default:
+		t.Fatalf("should failed to connect. err type: %v, err:  %v", errType, err)
 	}
 }
 
@@ -452,7 +517,7 @@ func TestOCSPFailOpenResponderTimeout(t *testing.T) {
 	defer cleanup()
 
 	setenv(cacheServerEnabledEnv, "false")
-	setenv(ocspTestResponderURLEnv, "http://localhost:12345/hang")
+	setenv(ocspTestResponderURLEnv, "http://localhost:12345/ocsp/hang")
 	setenv(ocspTestResponderTimeoutEnv, "1000")
 
 	config := &Config{
@@ -477,11 +542,14 @@ func TestOCSPFailOpenResponderTimeout(t *testing.T) {
 	if err = db.Ping(); err == nil {
 		t.Fatalf("should fail to ping. %v", testURL)
 	}
+	if strings.Contains(err.Error(), "HTTP Status: 513. Hanging?") {
+		return
+	}
 	driverErr, ok := err.(*SnowflakeError)
 	if !ok {
 		t.Fatalf("failed to extract error SnowflakeError: %v", err)
 	}
-	if driverErr.Number != ErrCodeFailedToConnect {
+	if isFailToConnectOrAuthErr(driverErr) {
 		t.Fatalf("should failed to connect %v", err)
 	}
 }
@@ -492,7 +560,7 @@ func TestOCSPFailClosedResponderTimeout(t *testing.T) {
 	defer cleanup()
 
 	setenv(cacheServerEnabledEnv, "false")
-	setenv(ocspTestResponderURLEnv, "http://localhost:12345/hang")
+	setenv(ocspTestResponderURLEnv, "http://localhost:12345/ocsp/hang")
 	setenv(ocspTestResponderTimeoutEnv, "1000")
 
 	config := &Config{
@@ -517,6 +585,9 @@ func TestOCSPFailClosedResponderTimeout(t *testing.T) {
 	if err = db.Ping(); err == nil {
 		t.Fatalf("should fail to ping. %v", testURL)
 	}
+	if strings.Contains(err.Error(), "HTTP Status: 513. Hanging?") {
+		return
+	}
 	urlErr, ok := err.(*url.Error)
 	if !ok {
 		t.Fatalf("failed to extract error URL Error: %v", err)
@@ -525,7 +596,7 @@ func TestOCSPFailClosedResponderTimeout(t *testing.T) {
 	if !ok {
 		t.Fatalf("failed to extract error URL Error: %v", urlErr.Err)
 	}
-	if !strings.Contains(urlErr0.Err.Error(), "Client.Timeout") {
+	if !strings.Contains(urlErr0.Err.Error(), "Client.Timeout") && !strings.Contains(urlErr0.Err.Error(), "connection refused") {
 		t.Fatalf("the root cause is not  timeout: %v", urlErr0.Err)
 	}
 }
@@ -536,7 +607,7 @@ func TestOCSPFailOpenResponder404(t *testing.T) {
 	defer cleanup()
 
 	setenv(cacheServerEnabledEnv, "false")
-	setenv(ocspTestResponderURLEnv, "http://localhost:12345/404")
+	setenv(ocspTestResponderURLEnv, "http://localhost:12345/ocsp/404")
 
 	config := &Config{
 		Account:      "fakeaccount10",
@@ -560,11 +631,14 @@ func TestOCSPFailOpenResponder404(t *testing.T) {
 	if err = db.Ping(); err == nil {
 		t.Fatalf("should fail to ping. %v", testURL)
 	}
+	if strings.Contains(err.Error(), "HTTP Status: 513. Hanging?") {
+		return
+	}
 	driverErr, ok := err.(*SnowflakeError)
 	if !ok {
 		t.Fatalf("failed to extract error SnowflakeError: %v", err)
 	}
-	if driverErr.Number != ErrCodeFailedToConnect {
+	if isFailToConnectOrAuthErr(driverErr) {
 		t.Fatalf("should failed to connect %v", err)
 	}
 }
@@ -575,7 +649,7 @@ func TestOCSPFailClosedResponder404(t *testing.T) {
 	defer cleanup()
 
 	setenv(cacheServerEnabledEnv, "false")
-	setenv(ocspTestResponderURLEnv, "http://localhost:12345/404")
+	setenv(ocspTestResponderURLEnv, "http://localhost:12345/ocsp/404")
 
 	config := &Config{
 		Account:      "fakeaccount11",
@@ -599,12 +673,15 @@ func TestOCSPFailClosedResponder404(t *testing.T) {
 	if err = db.Ping(); err == nil {
 		t.Fatalf("should fail to ping. %v", testURL)
 	}
+	if strings.Contains(err.Error(), "HTTP Status: 513. Hanging?") {
+		return
+	}
 	urlErr, ok := err.(*url.Error)
 	if !ok {
-		t.Fatalf("failed to extract error URL Error: %v", err)
+		t.Fatalf("failed to extract error SnowflakeError: %v", err)
 	}
-	if !strings.Contains(urlErr.Err.Error(), "HTTP Status: 404") {
-		t.Fatalf("the root cause is not  timeout: %v", urlErr.Err)
+	if !strings.Contains(urlErr.Err.Error(), "404 Not Found") && !strings.Contains(urlErr.Err.Error(), "connection refused") {
+		t.Fatalf("the root cause is not 404: %v", urlErr.Err)
 	}
 }
 
@@ -641,8 +718,14 @@ func TestExpiredCertificate(t *testing.T) {
 		t.Fatalf("failed to extract error URL Error: %v", err)
 	}
 	_, ok = urlErr.Err.(x509.CertificateInvalidError)
+
 	if !ok {
-		t.Fatalf("failed to extract error Certificate error: %v", err)
+		// Go 1.20 throws tls CertificateVerification error
+		errString := urlErr.Err.Error()
+		// badssl sometimes times out
+		if !strings.Contains(errString, "certificate has expired or is not yet valid") && !strings.Contains(errString, "timeout") {
+			t.Fatalf("failed to extract error Certificate error: %v", err)
+		}
 	}
 }
 
@@ -718,11 +801,14 @@ func TestOCSPFailOpenNoOCSPURL(t *testing.T) {
 	if err = db.Ping(); err == nil {
 		t.Fatalf("should fail to ping. %v", testURL)
 	}
+	if strings.Contains(err.Error(), "HTTP Status: 513. Hanging?") {
+		return
+	}
 	driverErr, ok := err.(*SnowflakeError)
 	if !ok {
 		t.Fatalf("failed to extract error SnowflakeError: %v", err)
 	}
-	if driverErr.Number != ErrCodeFailedToConnect {
+	if isFailToConnectOrAuthErr(driverErr) {
 		t.Fatalf("should failed to connect %v", err)
 	}
 }
@@ -757,13 +843,18 @@ func TestOCSPFailClosedNoOCSPURL(t *testing.T) {
 	if err = db.Ping(); err == nil {
 		t.Fatalf("should fail to ping. %v", testURL)
 	}
+	if strings.Contains(err.Error(), "HTTP Status: 513. Hanging?") {
+		return
+	}
 	urlErr, ok := err.(*url.Error)
 	if !ok {
-		t.Fatalf("failed to extract error URL Error: %v", err)
+		t.Fatalf("failed to extract error SnowflakeError: %v", err)
 	}
 	driverErr, ok := urlErr.Err.(*SnowflakeError)
 	if !ok {
-		t.Fatalf("failed to extract error SnowflakeError: %v", err)
+		if !strings.Contains(err.Error(), "HTTP Status: 513. Hanging?") {
+			t.Fatalf("failed to extract error SnowflakeError: %v", err)
+		}
 	}
 	if driverErr.Number != ErrOCSPNoOCSPResponderURL {
 		t.Fatalf("should failed to connect %v", err)
