@@ -1,25 +1,35 @@
-// Copyright (c) 2017-2021 Snowflake Computing Inc. All right reserved.
+// Copyright (c) 2017-2022 Snowflake Computing Inc. All rights reserved.
 
 package gosnowflake
 
 import (
 	"crypto/rsa"
+	"crypto/x509"
 	"encoding/base64"
+	"encoding/pem"
+	"errors"
 	"fmt"
 	"net"
 	"net/http"
 	"net/url"
+	"os"
 	"strconv"
 	"strings"
 	"time"
 )
 
 const (
-	defaultClientTimeout  = 900 * time.Second // Timeout for network round trip + read out http response
-	defaultLoginTimeout   = 60 * time.Second  // Timeout for retry for login EXCLUDING clientTimeout
-	defaultRequestTimeout = 0 * time.Second   // Timeout for retry for request EXCLUDING clientTimeout
-	defaultJWTTimeout     = 60 * time.Second
-	defaultDomain         = ".snowflakecomputing.com"
+	defaultClientTimeout          = 900 * time.Second // Timeout for network round trip + read out http response
+	defaultJWTClientTimeout       = 10 * time.Second  // Timeout for network round trip + read out http response but used for JWT auth
+	defaultLoginTimeout           = 300 * time.Second // Timeout for retry for login EXCLUDING clientTimeout
+	defaultRequestTimeout         = 0 * time.Second   // Timeout for retry for request EXCLUDING clientTimeout
+	defaultJWTTimeout             = 60 * time.Second
+	defaultExternalBrowserTimeout = 120 * time.Second // Timeout for external browser login
+	defaultCloudStorageTimeout    = -1                // Timeout for calling cloud storage.
+	defaultMaxRetryCount          = 7                 // specifies maximum number of subsequent retries
+	defaultDomain                 = ".snowflakecomputing.com"
+	cnDomain                      = ".snowflakecomputing.cn"
+	topLevelDomainPrefix          = ".snowflakecomputing." // used to extract the domain from host
 )
 
 // ConfigBool is a type to represent true or false in the Config
@@ -62,12 +72,18 @@ type Config struct {
 
 	OktaURL *url.URL
 
-	LoginTimeout     time.Duration // Login retry timeout EXCLUDING network roundtrip and read out http response
-	RequestTimeout   time.Duration // request retry timeout EXCLUDING network roundtrip and read out http response
-	JWTExpireTimeout time.Duration // JWT expire after timeout
-	ClientTimeout    time.Duration // Timeout for network round trip + read out http response
+	LoginTimeout           time.Duration // Login retry timeout EXCLUDING network roundtrip and read out http response
+	RequestTimeout         time.Duration // request retry timeout EXCLUDING network roundtrip and read out http response
+	JWTExpireTimeout       time.Duration // JWT expire after timeout
+	ClientTimeout          time.Duration // Timeout for network round trip + read out http response
+	JWTClientTimeout       time.Duration // Timeout for network round trip + read out http response used when JWT token auth is taking place
+	ExternalBrowserTimeout time.Duration // Timeout for external browser login
+	CloudStorageTimeout    time.Duration // Timeout for a single call to a cloud storage provider
+	MaxRetryCount          int           // Specifies how many times non-periodic HTTP request can be retried
 
-	Application  string           // application name.
+	Application       string // application name.
+	DisableOCSPChecks bool   // driver doesn't check certificate revocation status
+	// Deprecated: InsecureMode use DisableOCSPChecks instead
 	InsecureMode bool             // driver doesn't check certificate revocation status
 	OCSPFailOpen OCSPFailOpenMode // OCSP Fail Open
 
@@ -80,11 +96,41 @@ type Config struct {
 	Transporter http.RoundTripper // RoundTripper to intercept HTTP requests and responses
 
 	DisableTelemetry bool // indicates whether to disable telemetry
+
+	Tracing string // sets logging level
+
+	TmpDirPath string // sets temporary directory used by a driver for operations like encrypting, compressing etc
+
+	MfaToken                       string     // Internally used to cache the MFA token
+	IDToken                        string     // Internally used to cache the Id Token for external browser
+	ClientRequestMfaToken          ConfigBool // When true the MFA token is cached in the credential manager. True by default in Windows/OSX. False for Linux.
+	ClientStoreTemporaryCredential ConfigBool // When true the ID token is cached in the credential manager. True by default in Windows/OSX. False for Linux.
+
+	DisableQueryContextCache bool // Should HTAP query context cache be disabled
+
+	IncludeRetryReason ConfigBool // Should retried request contain retry reason
+
+	ClientConfigFile string // File path to the client configuration json file
+
+	DisableConsoleLogin ConfigBool // Indicates whether console login should be disabled
+
+	DisableSamlURLCheck ConfigBool // Indicates whether the SAML URL check should be disabled
+}
+
+// Validate enables testing if config is correct.
+// A driver client may call it manually, but it is also called during opening first connection.
+func (c *Config) Validate() error {
+	if c.TmpDirPath != "" {
+		if _, err := os.Stat(c.TmpDirPath); err != nil {
+			return err
+		}
+	}
+	return nil
 }
 
 // ocspMode returns the OCSP mode in string INSECURE, FAIL_OPEN, FAIL_CLOSED
 func (c *Config) ocspMode() string {
-	if c.InsecureMode {
+	if c.DisableOCSPChecks || c.InsecureMode {
 		return ocspModeInsecure
 	} else if c.OCSPFailOpen == ocspFailOpenNotSet || c.OCSPFailOpen == OCSPFailOpenTrue {
 		// by default or set to true
@@ -95,26 +141,31 @@ func (c *Config) ocspMode() string {
 
 // DSN constructs a DSN for Snowflake db.
 func DSN(cfg *Config) (dsn string, err error) {
+	if strings.ToLower(cfg.Region) == "us-west-2" {
+		cfg.Region = ""
+	}
+	// in case account includes region
+	region, posDot := extractRegionFromAccount(cfg.Account)
+	if strings.ToLower(region) == "us-west-2" {
+		region = ""
+		cfg.Account = cfg.Account[:posDot]
+		logger.Info("Ignoring default region .us-west-2 in DSN from Account configuration.")
+	}
+	if region != "" {
+		if cfg.Region != "" {
+			return "", errRegionConflict()
+		}
+		cfg.Region = region
+		cfg.Account = cfg.Account[:posDot]
+	}
 	hasHost := true
 	if cfg.Host == "" {
 		hasHost = false
-		if cfg.Region == "us-west-2" {
-			cfg.Region = ""
-		}
 		if cfg.Region == "" {
 			cfg.Host = cfg.Account + defaultDomain
 		} else {
-			cfg.Host = cfg.Account + "." + cfg.Region + defaultDomain
+			cfg.Host = buildHostFromAccountAndRegion(cfg.Account, cfg.Region)
 		}
-	}
-	// in case account includes region
-	posDot := strings.Index(cfg.Account, ".")
-	if posDot > 0 {
-		if cfg.Region != "" {
-			return "", ErrInvalidRegion
-		}
-		cfg.Region = cfg.Account[posDot+1:]
-		cfg.Account = cfg.Account[:posDot]
 	}
 	err = fillMissingConfigParameters(cfg)
 	if err != nil {
@@ -153,6 +204,12 @@ func DSN(cfg *Config) (dsn string, err error) {
 	if cfg.PasscodeInPassword {
 		params.Add("passcodeInPassword", strconv.FormatBool(cfg.PasscodeInPassword))
 	}
+	if cfg.ClientTimeout != defaultClientTimeout {
+		params.Add("clientTimeout", strconv.FormatInt(int64(cfg.ClientTimeout/time.Second), 10))
+	}
+	if cfg.JWTClientTimeout != defaultJWTClientTimeout {
+		params.Add("jwtClientTimeout", strconv.FormatInt(int64(cfg.JWTClientTimeout/time.Second), 10))
+	}
 	if cfg.LoginTimeout != defaultLoginTimeout {
 		params.Add("loginTimeout", strconv.FormatInt(int64(cfg.LoginTimeout/time.Second), 10))
 	}
@@ -161,6 +218,15 @@ func DSN(cfg *Config) (dsn string, err error) {
 	}
 	if cfg.JWTExpireTimeout != defaultJWTTimeout {
 		params.Add("jwtTimeout", strconv.FormatInt(int64(cfg.JWTExpireTimeout/time.Second), 10))
+	}
+	if cfg.ExternalBrowserTimeout != defaultExternalBrowserTimeout {
+		params.Add("externalBrowserTimeout", strconv.FormatInt(int64(cfg.ExternalBrowserTimeout/time.Second), 10))
+	}
+	if cfg.CloudStorageTimeout != defaultCloudStorageTimeout {
+		params.Add("cloudStorageTimeout", strconv.FormatInt(int64(cfg.CloudStorageTimeout/time.Second), 10))
+	}
+	if cfg.MaxRetryCount != defaultMaxRetryCount {
+		params.Add("maxRetryCount", strconv.Itoa(cfg.MaxRetryCount))
 	}
 	if cfg.Application != clientType {
 		params.Add("application", cfg.Application)
@@ -187,10 +253,42 @@ func DSN(cfg *Config) (dsn string, err error) {
 	if cfg.InsecureMode {
 		params.Add("insecureMode", strconv.FormatBool(cfg.InsecureMode))
 	}
+	if cfg.DisableOCSPChecks {
+		params.Add("disableOCSPChecks", strconv.FormatBool(cfg.DisableOCSPChecks))
+	}
+	if cfg.Tracing != "" {
+		params.Add("tracing", cfg.Tracing)
+	}
+	if cfg.TmpDirPath != "" {
+		params.Add("tmpDirPath", cfg.TmpDirPath)
+	}
+	if cfg.DisableQueryContextCache {
+		params.Add("disableQueryContextCache", "true")
+	}
+	if cfg.IncludeRetryReason == ConfigBoolFalse {
+		params.Add("includeRetryReason", "false")
+	}
 
 	params.Add("ocspFailOpen", strconv.FormatBool(cfg.OCSPFailOpen != OCSPFailOpenFalse))
 
 	params.Add("validateDefaultParameters", strconv.FormatBool(cfg.ValidateDefaultParameters != ConfigBoolFalse))
+
+	if cfg.ClientRequestMfaToken != configBoolNotSet {
+		params.Add("clientRequestMfaToken", strconv.FormatBool(cfg.ClientRequestMfaToken != ConfigBoolFalse))
+	}
+
+	if cfg.ClientStoreTemporaryCredential != configBoolNotSet {
+		params.Add("clientStoreTemporaryCredential", strconv.FormatBool(cfg.ClientStoreTemporaryCredential != ConfigBoolFalse))
+	}
+	if cfg.ClientConfigFile != "" {
+		params.Add("clientConfigFile", cfg.ClientConfigFile)
+	}
+	if cfg.DisableConsoleLogin != configBoolNotSet {
+		params.Add("disableConsoleLogin", strconv.FormatBool(cfg.DisableConsoleLogin != ConfigBoolFalse))
+	}
+	if cfg.DisableSamlURLCheck != configBoolNotSet {
+		params.Add("disableSamlURLCheck", strconv.FormatBool(cfg.DisableSamlURLCheck != ConfigBoolFalse))
+	}
 
 	dsn = fmt.Sprintf("%v:%v@%v:%v", url.QueryEscape(cfg.User), url.QueryEscape(cfg.Password), cfg.Host, cfg.Port)
 	if params.Encode() != "" {
@@ -293,7 +391,7 @@ func ParseDSN(dsn string) (cfg *Config, err error) {
 			return
 		}
 	}
-	if cfg.Account == "" && strings.HasSuffix(cfg.Host, defaultDomain) {
+	if cfg.Account == "" && hostIncludesTopLevelDomain(cfg.Host) {
 		posDot := strings.Index(cfg.Host, ".")
 		if posDot > 0 {
 			cfg.Account = cfg.Host[:posDot]
@@ -347,26 +445,26 @@ func ParseDSN(dsn string) (cfg *Config, err error) {
 func fillMissingConfigParameters(cfg *Config) error {
 	posDash := strings.LastIndex(cfg.Account, "-")
 	if posDash > 0 {
-		if strings.Contains(cfg.Host, ".global.") {
+		if strings.Contains(strings.ToLower(cfg.Host), ".global.") {
 			cfg.Account = cfg.Account[:posDash]
 		}
 	}
 	if strings.Trim(cfg.Account, " ") == "" {
-		return ErrEmptyAccount
+		return errEmptyAccount()
 	}
 
-	if cfg.Authenticator != AuthTypeOAuth && strings.Trim(cfg.User, " ") == "" {
-		// oauth does not require a username
-		return ErrEmptyUsername
+	if authRequiresUser(cfg) && strings.TrimSpace(cfg.User) == "" {
+		return errEmptyUsername()
 	}
 
-	if cfg.Authenticator != AuthTypeExternalBrowser &&
-		cfg.Authenticator != AuthTypeOAuth &&
-		cfg.Authenticator != AuthTypeJwt &&
-		strings.Trim(cfg.Password, " ") == "" {
-		// no password parameter is required for EXTERNALBROWSER, OAUTH or JWT.
-		return ErrEmptyPassword
+	if authRequiresPassword(cfg) && strings.TrimSpace(cfg.Password) == "" {
+		return errEmptyPassword()
 	}
+
+	if authRequiresEitherPasswordOrToken(cfg) && strings.TrimSpace(cfg.Password) == "" && strings.TrimSpace(cfg.Token) == "" {
+		return errEmptyPasswordAndToken()
+	}
+
 	if strings.Trim(cfg.Protocol, " ") == "" {
 		cfg.Protocol = "https"
 	}
@@ -377,19 +475,24 @@ func fillMissingConfigParameters(cfg *Config) error {
 	cfg.Region = strings.Trim(cfg.Region, " ")
 	if cfg.Region != "" {
 		// region is specified but not included in Host
-		i := strings.Index(cfg.Host, defaultDomain)
+		domain, i := extractDomainFromHost(cfg.Host)
 		if i >= 1 {
 			hostPrefix := cfg.Host[0:i]
 			if !strings.HasSuffix(hostPrefix, cfg.Region) {
-				cfg.Host = hostPrefix + "." + cfg.Region + defaultDomain
+				cfg.Host = fmt.Sprintf("%v.%v%v", hostPrefix, cfg.Region, domain)
 			}
 		}
 	}
 	if cfg.Host == "" {
 		if cfg.Region != "" {
-			cfg.Host = cfg.Account + "." + cfg.Region + defaultDomain
+			cfg.Host = cfg.Account + "." + cfg.Region + getDomainBasedOnRegion(cfg.Region)
 		} else {
-			cfg.Host = cfg.Account + defaultDomain
+			region, _ := extractRegionFromAccount(cfg.Account)
+			if region != "" {
+				cfg.Host = cfg.Account + getDomainBasedOnRegion(region)
+			} else {
+				cfg.Host = cfg.Account + defaultDomain
+			}
 		}
 	}
 	if cfg.LoginTimeout == 0 {
@@ -404,6 +507,18 @@ func fillMissingConfigParameters(cfg *Config) error {
 	if cfg.ClientTimeout == 0 {
 		cfg.ClientTimeout = defaultClientTimeout
 	}
+	if cfg.JWTClientTimeout == 0 {
+		cfg.JWTClientTimeout = defaultJWTClientTimeout
+	}
+	if cfg.ExternalBrowserTimeout == 0 {
+		cfg.ExternalBrowserTimeout = defaultExternalBrowserTimeout
+	}
+	if cfg.CloudStorageTimeout == 0 {
+		cfg.CloudStorageTimeout = defaultCloudStorageTimeout
+	}
+	if cfg.MaxRetryCount == 0 {
+		cfg.MaxRetryCount = defaultMaxRetryCount
+	}
 	if strings.Trim(cfg.Application, " ") == "" {
 		cfg.Application = clientType
 	}
@@ -416,7 +531,12 @@ func fillMissingConfigParameters(cfg *Config) error {
 		cfg.ValidateDefaultParameters = ConfigBoolTrue
 	}
 
-	if strings.HasSuffix(cfg.Host, defaultDomain) && len(cfg.Host) == len(defaultDomain) {
+	if cfg.IncludeRetryReason == configBoolNotSet {
+		cfg.IncludeRetryReason = ConfigBoolTrue
+	}
+
+	domain, _ := extractDomainFromHost(cfg.Host)
+	if len(cfg.Host) == len(domain) {
 		return &SnowflakeError{
 			Number:      ErrCodeFailedToParseHost,
 			Message:     errMsgFailedToParseHost,
@@ -426,18 +546,76 @@ func fillMissingConfigParameters(cfg *Config) error {
 	return nil
 }
 
-// transformAccountToHost transforms host to account name
+func extractDomainFromHost(host string) (domain string, index int) {
+	i := strings.LastIndex(strings.ToLower(host), topLevelDomainPrefix)
+	if i >= 1 {
+		domain = host[i:]
+		return domain, i
+	}
+	return "", i
+}
+
+func getDomainBasedOnRegion(region string) string {
+	if strings.HasPrefix(strings.ToLower(region), "cn-") {
+		return cnDomain
+	}
+	return defaultDomain
+}
+
+func extractRegionFromAccount(account string) (region string, posDot int) {
+	posDot = strings.Index(strings.ToLower(account), ".")
+	if posDot > 0 {
+		return account[posDot+1:], posDot
+	}
+	return "", posDot
+}
+
+func hostIncludesTopLevelDomain(host string) bool {
+	return strings.Contains(strings.ToLower(host), topLevelDomainPrefix)
+}
+
+func buildHostFromAccountAndRegion(account, region string) string {
+	return account + "." + region + getDomainBasedOnRegion(region)
+}
+
+func authRequiresUser(cfg *Config) bool {
+	return cfg.Authenticator != AuthTypeOAuth &&
+		cfg.Authenticator != AuthTypeTokenAccessor &&
+		cfg.Authenticator != AuthTypeExternalBrowser &&
+		cfg.Authenticator != AuthTypePat
+}
+
+func authRequiresPassword(cfg *Config) bool {
+	return cfg.Authenticator != AuthTypeOAuth &&
+		cfg.Authenticator != AuthTypeTokenAccessor &&
+		cfg.Authenticator != AuthTypeExternalBrowser &&
+		cfg.Authenticator != AuthTypeJwt &&
+		cfg.Authenticator != AuthTypePat
+}
+
+func authRequiresEitherPasswordOrToken(cfg *Config) bool {
+	return cfg.Authenticator == AuthTypePat
+}
+
+// transformAccountToHost transforms account to host
 func transformAccountToHost(cfg *Config) (err error) {
-	if cfg.Port == 0 && !strings.HasSuffix(cfg.Host, defaultDomain) && cfg.Host != "" {
+	if cfg.Port == 0 && cfg.Host != "" && !hostIncludesTopLevelDomain(cfg.Host) {
 		// account name is specified instead of host:port
 		cfg.Account = cfg.Host
-		cfg.Host = cfg.Account + defaultDomain
-		cfg.Port = 443
-		posDot := strings.Index(cfg.Account, ".")
-		if posDot > 0 {
-			cfg.Region = cfg.Account[posDot+1:]
+		region, posDot := extractRegionFromAccount(cfg.Account)
+		if strings.ToLower(region) == "us-west-2" {
+			region = ""
 			cfg.Account = cfg.Account[:posDot]
+			logger.Info("Ignoring default region .us-west-2 from Account configuration.")
 		}
+		if region != "" {
+			cfg.Region = region
+			cfg.Account = cfg.Account[:posDot]
+			cfg.Host = buildHostFromAccountAndRegion(cfg.Account, cfg.Region)
+		} else {
+			cfg.Host = cfg.Account + defaultDomain
+		}
+		cfg.Port = 443
 	}
 	return nil
 }
@@ -493,7 +671,14 @@ func parseParams(cfg *Config, posQuestion int, dsn string) (err error) {
 // parseDSNParams parses the DSN "query string". Values must be url.QueryEscape'ed
 func parseDSNParams(cfg *Config, params string) (err error) {
 	logger.Infof("Query String: %v\n", params)
-	for _, v := range strings.Split(params, "&") {
+	paramsSlice := strings.Split(params, "&")
+	insecureModeIdx := findByPrefix(paramsSlice, "insecureMode")
+	disableOCSPChecksIdx := findByPrefix(paramsSlice, "disableOCSPChecks")
+	if insecureModeIdx > -1 && disableOCSPChecksIdx > -1 {
+		logger.Warn("duplicated insecureMode and disableOCSPChecks. disableOCSPChecks takes precedence")
+		paramsSlice = append(paramsSlice[:insecureModeIdx-1], paramsSlice[insecureModeIdx+1:]...)
+	}
+	for _, v := range paramsSlice {
 		param := strings.SplitN(v, "=", 2)
 		if len(param) != 2 {
 			continue
@@ -528,6 +713,16 @@ func parseDSNParams(cfg *Config, params string) (err error) {
 				return
 			}
 			cfg.PasscodeInPassword = vv
+		case "clientTimeout":
+			cfg.ClientTimeout, err = parseTimeout(value)
+			if err != nil {
+				return
+			}
+		case "jwtClientTimeout":
+			cfg.JWTClientTimeout, err = parseTimeout(value)
+			if err != nil {
+				return
+			}
 		case "loginTimeout":
 			cfg.LoginTimeout, err = parseTimeout(value)
 			if err != nil {
@@ -543,6 +738,21 @@ func parseDSNParams(cfg *Config, params string) (err error) {
 			if err != nil {
 				return err
 			}
+		case "externalBrowserTimeout":
+			cfg.ExternalBrowserTimeout, err = parseTimeout(value)
+			if err != nil {
+				return err
+			}
+		case "cloudStorageTimeout":
+			cfg.CloudStorageTimeout, err = parseTimeout(value)
+			if err != nil {
+				return err
+			}
+		case "maxRetryCount":
+			cfg.MaxRetryCount, err = strconv.Atoi(value)
+			if err != nil {
+				return err
+			}
 		case "application":
 			cfg.Application = value
 		case "authenticator":
@@ -551,12 +761,20 @@ func parseDSNParams(cfg *Config, params string) (err error) {
 				return err
 			}
 		case "insecureMode":
+			logInsecureModeDeprecationInfo()
 			var vv bool
 			vv, err = strconv.ParseBool(value)
 			if err != nil {
 				return
 			}
 			cfg.InsecureMode = vv
+		case "disableOCSPChecks":
+			var vv bool
+			vv, err = strconv.ParseBool(value)
+			if err != nil {
+				return
+			}
+			cfg.DisableOCSPChecks = vv
 		case "ocspFailOpen":
 			var vv bool
 			vv, err = strconv.ParseBool(value)
@@ -596,14 +814,87 @@ func parseDSNParams(cfg *Config, params string) (err error) {
 			} else {
 				cfg.ValidateDefaultParameters = ConfigBoolFalse
 			}
+		case "clientRequestMfaToken":
+			var vv bool
+			vv, err = strconv.ParseBool(value)
+			if err != nil {
+				return
+			}
+			if vv {
+				cfg.ClientRequestMfaToken = ConfigBoolTrue
+			} else {
+				cfg.ClientRequestMfaToken = ConfigBoolFalse
+			}
+		case "clientStoreTemporaryCredential":
+			var vv bool
+			vv, err = strconv.ParseBool(value)
+			if err != nil {
+				return
+			}
+			if vv {
+				cfg.ClientStoreTemporaryCredential = ConfigBoolTrue
+			} else {
+				cfg.ClientStoreTemporaryCredential = ConfigBoolFalse
+			}
+		case "tracing":
+			cfg.Tracing = value
+		case "tmpDirPath":
+			cfg.TmpDirPath = value
+		case "disableQueryContextCache":
+			var b bool
+			b, err = strconv.ParseBool(value)
+			if err != nil {
+				return
+			}
+			cfg.DisableQueryContextCache = b
+		case "includeRetryReason":
+			var vv bool
+			vv, err = strconv.ParseBool(value)
+			if err != nil {
+				return
+			}
+			if vv {
+				cfg.IncludeRetryReason = ConfigBoolTrue
+			} else {
+				cfg.IncludeRetryReason = ConfigBoolFalse
+			}
+		case "clientConfigFile":
+			cfg.ClientConfigFile = value
+		case "disableConsoleLogin":
+			var vv bool
+			vv, err = strconv.ParseBool(value)
+			if err != nil {
+				return
+			}
+			if vv {
+				cfg.DisableConsoleLogin = ConfigBoolTrue
+			} else {
+				cfg.DisableConsoleLogin = ConfigBoolFalse
+			}
+		case "disableSamlURLCheck":
+			var vv bool
+			vv, err = strconv.ParseBool(value)
+			if err != nil {
+				return
+			}
+			if vv {
+				cfg.DisableSamlURLCheck = ConfigBoolTrue
+			} else {
+				cfg.DisableSamlURLCheck = ConfigBoolFalse
+			}
 		default:
 			if cfg.Params == nil {
 				cfg.Params = make(map[string]*string)
 			}
-			cfg.Params[param[0]] = &value
+			// handle session variables $variable=value
+			cfg.Params[urlDecodeIfNeeded(param[0])] = &value
 		}
 	}
 	return
+}
+
+func logInsecureModeDeprecationInfo() {
+	logger.Warn("insecureMode is deprecated. Use disableOCSPChecks instead.")
 }
 
 func parseTimeout(value string) (time.Duration, error) {
@@ -614,4 +905,126 @@ func parseTimeout(value string) (time.Duration, error) {
 		return time.Duration(0), err
 	}
 	return time.Duration(vv * int64(time.Second)), nil
+}
+
+// ConfigParam is used to bind the name of the Config field with the environment variable and set the requirement for it
+type ConfigParam struct {
+	Name          string
+	EnvName       string
+	FailOnMissing bool
+}
+
+// GetConfigFromEnv is used to parse the environment variable values to specific fields of the Config
+func GetConfigFromEnv(properties []*ConfigParam) (*Config, error) {
+	var account, user, password, token, role, host, portStr, protocol, warehouse, database, schema, region, passcode, application string
+	var privateKey *rsa.PrivateKey
+	var err error
+	if len(properties) == 0 || properties == nil {
+		return nil, errors.New("missing configuration parameters for the connection")
+	}
+	for _, prop := range properties {
+		value, err := GetFromEnv(prop.EnvName, prop.FailOnMissing)
+		if err != nil {
+			return nil, err
+		}
+		switch prop.Name {
+		case "Account":
+			account = value
+		case "User":
+			user = value
+		case "Password":
+			password = value
+		case "Token":
+			token = value
+		case "Role":
+			role = value
+		case "Host":
+			host = value
+		case "Port":
+			portStr = value
+		case "Protocol":
+			protocol = value
+		case "Warehouse":
+			warehouse = value
+		case "Database":
+			database = value
+		case "Region":
+			region = value
+		case "Passcode":
+			passcode = value
+		case "Schema":
+			schema = value
+		case "Application":
+			application = value
+		case "PrivateKey":
+			privateKey, err = parsePrivateKeyFromFile(value)
+			if err != nil {
+				return nil, err
+			}
+		}
+	}
+
+	port := 443 // snowflake default port
+	if len(portStr) > 0 {
+		port, err = strconv.Atoi(portStr)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	cfg := &Config{
+		Account:     account,
+		User:        user,
+		Password:    password,
+		Token:       token,
+		Role:        role,
+		Host:        host,
+		Port:        port,
+		Protocol:    protocol,
+		Warehouse:   warehouse,
+		Database:    database,
+		Schema:      schema,
+		PrivateKey:  privateKey,
+		Region:      region,
+		Passcode:    passcode,
+		Application: application,
+		Params:      map[string]*string{},
+	}
+	return cfg, nil
+}
+
+func parsePrivateKeyFromFile(path string) (*rsa.PrivateKey, error) {
+	bytes, err := os.ReadFile(path)
+	if err != nil {
+		return nil, err
+	}
+	block, _ := pem.Decode(bytes)
+	if block == nil {
+		return nil, errors.New("failed to parse PEM block containing the private key")
+	}
+	privateKey, err := x509.ParsePKCS8PrivateKey(block.Bytes)
+	if err != nil {
+		return nil, err
+	}
+	pk, ok := privateKey.(*rsa.PrivateKey)
+	if !ok {
+		return nil, fmt.Errorf("interface convertion. expected type *rsa.PrivateKey, but got %T", privateKey)
+	}
+	return pk, nil
+}
+
+func extractAccountName(rawAccount string) string {
+	posDot := strings.Index(rawAccount, ".")
+	if posDot > 0 {
+		return strings.ToUpper(rawAccount[:posDot])
+	}
+	return strings.ToUpper(rawAccount)
+}
+
+func urlDecodeIfNeeded(param string) (decodedParam string) {
+	unescaped, err := url.QueryUnescape(param)
+	if err != nil {
+		return param
+	}
+	return unescaped
 }
