@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2021 Snowflake Computing Inc. All rights reserved.
+// Copyright (c) 2017-2022 Snowflake Computing Inc. All rights reserved.
 
 package gosnowflake
 
@@ -40,11 +40,12 @@ func (se *SnowflakeError) generateTelemetryExceptionData() *telemetryData {
 	data := &telemetryData{
 		Message: map[string]string{
 			typeKey:          sqlException,
+			sourceKey:        telemetrySource,
 			driverTypeKey:    "Go",
 			driverVersionKey: SnowflakeGoDriverVersion,
 			stacktraceKey:    maskSecrets(string(debug.Stack())),
 		},
-		Timestamp: time.Now().UnixNano(),
+		Timestamp: time.Now().UnixNano() / int64(time.Millisecond),
 	}
 	if se.QueryID != "" {
 		data.Message[queryIDKey] = se.QueryID
@@ -65,7 +66,7 @@ func (se *SnowflakeError) generateTelemetryExceptionData() *telemetryData {
 }
 
 func (se *SnowflakeError) sendExceptionTelemetry(sc *snowflakeConn, data *telemetryData) error {
-	if sc != nil {
+	if sc != nil && sc.telemetry != nil {
 		return sc.telemetry.addLog(data)
 	}
 	return nil // TODO oob telemetry
@@ -74,23 +75,41 @@ func (se *SnowflakeError) sendExceptionTelemetry(sc *snowflakeConn, data *teleme
 func (se *SnowflakeError) exceptionTelemetry(sc *snowflakeConn) *SnowflakeError {
 	data := se.generateTelemetryExceptionData()
 	if err := se.sendExceptionTelemetry(sc, data); err != nil {
-		logger.Debugf("failed to log to telemetry: %v", data)
+		logger.WithContext(sc.ctx).Debugf("failed to log to telemetry: %v", data)
 	}
 	return se
+}
+
+// return populated error fields replacing the default response
+func populateErrorFields(code int, data *execResponse) *SnowflakeError {
+	err := errUnknownError()
+	if code != -1 {
+		err.Number = code
+	}
+	if data.Data.SQLState != "" {
+		err.SQLState = data.Data.SQLState
+	}
+	if data.Message != "" {
+		err.Message = data.Message
+	}
+	if data.Data.QueryID != "" {
+		err.QueryID = data.Data.QueryID
+	}
+	return err
 }
 
 const (
 	/* connection */
 
-	// ErrCodeEmptyAccountCode is an error code for the case where a DNS doesn't include account parameter
+	// ErrCodeEmptyAccountCode is an error code for the case where a DSN doesn't include account parameter
 	ErrCodeEmptyAccountCode = 260000
-	// ErrCodeEmptyUsernameCode is an error code for the case where a DNS doesn't include user parameter
+	// ErrCodeEmptyUsernameCode is an error code for the case where a DSN doesn't include user parameter
 	ErrCodeEmptyUsernameCode = 260001
-	// ErrCodeEmptyPasswordCode is an error code for the case where a DNS doesn't include password parameter
+	// ErrCodeEmptyPasswordCode is an error code for the case where a DSN doesn't include password parameter
 	ErrCodeEmptyPasswordCode = 260002
-	// ErrCodeFailedToParseHost is an error code for the case where a DNS includes an invalid host name
+	// ErrCodeFailedToParseHost is an error code for the case where a DSN includes an invalid host name
 	ErrCodeFailedToParseHost = 260003
-	// ErrCodeFailedToParsePort is an error code for the case where a DNS includes an invalid port number
+	// ErrCodeFailedToParsePort is an error code for the case where a DSN includes an invalid port number
 	ErrCodeFailedToParsePort = 260004
 	// ErrCodeIdpConnectionError is an error code for the case where a IDP connection failed
 	ErrCodeIdpConnectionError = 260005
@@ -106,6 +125,16 @@ const (
 	ErrCodePrivateKeyParseError = 260010
 	// ErrCodeFailedToParseAuthenticator is an error code for the case where a DNS includes an invalid authenticator
 	ErrCodeFailedToParseAuthenticator = 260011
+	// ErrCodeClientConfigFailed is an error code for the case where clientConfigFile is invalid or applying client configuration fails
+	ErrCodeClientConfigFailed = 260012
+	// ErrCodeTomlFileParsingFailed is an error code for the case where parsing the toml file is failed because of invalid value.
+	ErrCodeTomlFileParsingFailed = 260013
+	// ErrCodeFailedToFindDSNInToml is an error code for the case where the DSN does not exist in the toml file.
+	ErrCodeFailedToFindDSNInToml = 260014
+	// ErrCodeInvalidFilePermission is an error code for the case where the user does not have 0600 permission to the toml file .
+	ErrCodeInvalidFilePermission = 260015
+	// ErrCodeEmptyPasswordAndToken is an error code for the case where a DSN do includes neither password nor token
+	ErrCodeEmptyPasswordAndToken = 260016
 
 	/* network */
 
@@ -136,6 +165,8 @@ const (
 
 	// ErrFailedToGetChunk is an error code for the case where it failed to get chunk of result set
 	ErrFailedToGetChunk = 262000
+	// ErrNonArrowResponseInArrowBatches is an error code for case where ArrowBatches mode is enabled, but response is not Arrow-based
+	ErrNonArrowResponseInArrowBatches = 262001
 
 	/* transaction*/
 
@@ -162,6 +193,14 @@ const (
 	ErrCompressionNotSupported = 264007
 	// ErrInternalNotMatchEncryptMaterial is an error code denoting the encryption material specified does not match
 	ErrInternalNotMatchEncryptMaterial = 264008
+	// ErrCommandNotRecognized is an error code denoting the PUT/GET command was not recognized
+	ErrCommandNotRecognized = 264009
+	// ErrFailedToConvertToS3Client is an error code denoting the failure of an interface to s3.Client conversion
+	ErrFailedToConvertToS3Client = 264010
+	// ErrNotImplemented is an error code denoting the file transfer feature is not implemented
+	ErrNotImplemented = 264011
+	// ErrInvalidPadding is an error code denoting the invalid padding of decryption key
+	ErrInvalidPadding = 264012
 
 	/* binding */
 
@@ -170,15 +209,31 @@ const (
 	// ErrBindUpload is an error code for the uploading process of bind elements to the stage
 	ErrBindUpload = 265002
 
+	/* async */
+
+	// ErrAsync is an error code for an unknown async error
+	ErrAsync = 266001
+
+	/* multi-statement */
+
+	// ErrNoResultIDs is an error code for empty result IDs for multi statement queries
+	ErrNoResultIDs = 267001
+
 	/* converter */
 
 	// ErrInvalidTimestampTz is an error code for the case where a returned TIMESTAMP_TZ internal value is invalid
 	ErrInvalidTimestampTz = 268000
-	// ErrInvalidOffsetStr is an error code for the case where a offset string is invalid. The input string must
+	// ErrInvalidOffsetStr is an error code for the case where an offset string is invalid. The input string must
 	// consist of sHHMI where one sign character '+'/'-' followed by zero filled hours and minutes
 	ErrInvalidOffsetStr = 268001
 	// ErrInvalidBinaryHexForm is an error code for the case where a binary data in hex form is invalid.
 	ErrInvalidBinaryHexForm = 268002
+	// ErrTooHighTimestampPrecision is an error code for the case where cannot convert Snowflake timestamp to arrow.Timestamp
+	ErrTooHighTimestampPrecision = 268003
+	// ErrNullValueInArray is an error code for the case where there are null values in an array without arrayValuesNullable set to true
+	ErrNullValueInArray = 268004
+	// ErrNullValueInMap is an error code for the case where there are null values in a map without mapValuesNullable set to true
+	ErrNullValueInMap = 268005
 
 	/* OCSP */
 
@@ -191,7 +246,7 @@ const (
 	// ErrOCSPNoOCSPResponderURL is an error code for the case where the OCSP responder URL is not attached.
 	ErrOCSPNoOCSPResponderURL = 269004
 
-	/* Query Status*/
+	/* query Status*/
 
 	// ErrQueryStatus when check the status of a query, receive error or no status
 	ErrQueryStatus = 279001
@@ -240,26 +295,101 @@ const (
 	errMsgOCSPInvalidValidity                = "invalid validity: producedAt: %v, thisUpdate: %v, nextUpdate: %v"
 	errMsgOCSPNoOCSPResponderURL             = "no OCSP server is attached to the certificate. %v"
 	errMsgBindColumnMismatch                 = "column %v has a different number of binds (%v) than column 1 (%v)"
+	errMsgNotImplemented                     = "not implemented"
+	errMsgFeatureNotSupported                = "feature is not supported: %v"
+	errMsgCommandNotRecognized               = "%v command not recognized"
+	errMsgLocalPathNotDirectory              = "the local path is not a directory: %v"
+	errMsgFileNotExists                      = "file does not exist: %v"
+	errMsgInvalidStageFs                     = "destination location type is not valid: %v"
+	errMsgInternalNotMatchEncryptMaterial    = "number of downloading files doesn't match the encryption materials. files=%v, encmat=%v"
+	errMsgFailedToConvertToS3Client          = "failed to convert interface to s3 client"
+	errMsgNoResultIDs                        = "no result IDs returned with the multi-statement query"
+	errMsgQueryStatus                        = "server ErrorCode=%s, ErrorMessage=%s"
+	errMsgInvalidPadding                     = "invalid padding on input"
+	errMsgClientConfigFailed                 = "client configuration failed: %v"
+	errMsgNullValueInArray                   = "for handling null values in arrays use WithArrayValuesNullable(ctx)"
+	errMsgNullValueInMap                     = "for handling null values in maps use WithMapValuesNullable(ctx)"
+	errMsgFailedToParseTomlFile              = "failed to parse toml file. the params %v occurred error with value %v"
+	errMsgFailedToFindDSNInTomlFile          = "failed to find DSN in toml file."
+	errMsgInvalidPermissionToTomlFile        = "file permissions different than read/write for user. Your Permission: %v"
+	errMsgNonArrowResponseInArrowBatches     = "arrow batches enabled, but the response is not Arrow based"
 )
 
-var (
-	// ErrEmptyAccount is returned if a DNS doesn't include account parameter.
-	ErrEmptyAccount = &SnowflakeError{
+// Returned if a DNS doesn't include account parameter.
+func errEmptyAccount() *SnowflakeError {
+	return &SnowflakeError{
 		Number:  ErrCodeEmptyAccountCode,
 		Message: "account is empty",
 	}
-	// ErrEmptyUsername is returned if a DNS doesn't include user parameter.
-	ErrEmptyUsername = &SnowflakeError{
+}
+
+// Returned if a DNS doesn't include user parameter.
+func errEmptyUsername() *SnowflakeError {
+	return &SnowflakeError{
 		Number:  ErrCodeEmptyUsernameCode,
 		Message: "user is empty",
 	}
-	// ErrEmptyPassword is returned if a DNS doesn't include password parameter.
-	ErrEmptyPassword = &SnowflakeError{
-		Number:  ErrCodeEmptyPasswordCode,
-		Message: "password is empty"}
+}
 
-	// ErrInvalidRegion is returned if a DSN's implicit region from account parameter and explicit region parameter conflict.
-	ErrInvalidRegion = &SnowflakeError{
+// Returned if a DNS doesn't include password parameter.
+func errEmptyPassword() *SnowflakeError {
+	return &SnowflakeError{
+		Number:  ErrCodeEmptyPasswordCode,
+		Message: "password is empty",
+	}
+}
+
+func errEmptyPasswordAndToken() *SnowflakeError {
+	return &SnowflakeError{
+		Number:  ErrCodeEmptyPasswordAndToken,
+		Message: "both password and token are empty",
+	}
+}
+
+// Returned if a DSN's implicit region from account parameter and explicit region parameter conflict.
+func errRegionConflict() *SnowflakeError {
+	return &SnowflakeError{
 		Number:  ErrCodeRegionOverlap,
-		Message: "two regions specified"}
-)
+		Message: "two regions specified",
+	}
+}
+
+// Returned if a DSN includes an invalid authenticator.
+func errFailedToParseAuthenticator() *SnowflakeError {
+	return &SnowflakeError{
+		Number:  ErrCodeFailedToParseAuthenticator,
+		Message: "failed to parse an authenticator",
+	}
+}
+
+// Returned if the server side returns an error without meaningful message.
+func errUnknownError() *SnowflakeError {
+	return &SnowflakeError{
+		Number:   -1,
+		SQLState: "-1",
+		Message:  "an unknown server side error occurred",
+		QueryID:  "-1",
+	}
+}
+
+func errNullValueInArray() *SnowflakeError {
+	return &SnowflakeError{
+		Number:  ErrNullValueInArray,
+		Message: errMsgNullValueInArray,
+	}
+}
+
+func errNullValueInMap() *SnowflakeError {
+	return &SnowflakeError{
+		Number:  ErrNullValueInMap,
+		Message: errMsgNullValueInMap,
+	}
+}
+
+func errNonArrowResponseForArrowBatches(queryID string) *SnowflakeError {
+	return &SnowflakeError{
+		QueryID: queryID,
+		Number:  ErrNonArrowResponseInArrowBatches,
+		Message: errMsgNonArrowResponseInArrowBatches,
+	}
+}
