@@ -1,10 +1,18 @@
-// Copyright (c) 2020-2021 Snowflake Computing Inc. All right reserved.
+// Copyright (c) 2020-2022 Snowflake Computing Inc. All rights reserved.
 
 package gosnowflake
 
 import (
 	"context"
+	"encoding/json"
+	"errors"
+	"io"
+	"net/http"
+	"net/url"
+	"os"
+	"reflect"
 	"testing"
+	"time"
 )
 
 func TestMultiStatementExecuteNoResultSet(t *testing.T) {
@@ -14,11 +22,8 @@ func TestMultiStatementExecuteNoResultSet(t *testing.T) {
 		"insert into test_multi_statement_txn values (1, 'a'), (2, 'b');\n" +
 		"commit;"
 
-	runTests(t, dsn, func(dbt *DBTest) {
-		dbt.mustExec("drop table if exists test_multi_statement_txn")
-		dbt.mustExec(`create or replace table test_multi_statement_txn(
-			c1 number, c2 string) as select 10, 'z'`)
-		defer dbt.mustExec("drop table if exists test_multi_statement_txn")
+	runDBTest(t, func(dbt *DBTest) {
+		dbt.mustExec(`create or replace table test_multi_statement_txn(c1 number, c2 string) as select 10, 'z'`)
 
 		res := dbt.mustExecContext(ctx, multiStmtQuery)
 		count, err := res.RowsAffected()
@@ -40,8 +45,10 @@ func TestMultiStatementQueryResultSet(t *testing.T) {
 
 	var v1, v2, v3 int64
 	var v4 string
-	runTests(t, dsn, func(dbt *DBTest) {
+
+	runDBTest(t, func(dbt *DBTest) {
 		rows := dbt.mustQueryContext(ctx, multiStmtQuery)
+		defer rows.Close()
 
 		// first statement
 		if rows.Next() {
@@ -111,7 +118,7 @@ func TestMultiStatementExecuteResultSet(t *testing.T) {
 		"select 2;\n" +
 		"rollback;"
 
-	runTests(t, dsn, func(dbt *DBTest) {
+	runDBTest(t, func(dbt *DBTest) {
 		dbt.mustExec("drop table if exists test_multi_statement_txn_rb")
 		dbt.mustExec(`create or replace table test_multi_statement_txn_rb(
 			c1 number, c2 string) as select 10, 'z'`)
@@ -135,13 +142,14 @@ func TestMultiStatementQueryNoResultSet(t *testing.T) {
 		"insert into test_multi_statement_txn values (1, 'a'), (2, 'b');\n" +
 		"commit;"
 
-	runTests(t, dsn, func(dbt *DBTest) {
+	runDBTest(t, func(dbt *DBTest) {
 		dbt.mustExec("drop table if exists test_multi_statement_txn")
 		dbt.mustExec(`create or replace table test_multi_statement_txn(
 			c1 number, c2 string) as select 10, 'z'`)
 		defer dbt.mustExec("drop table if exists tfmuest_multi_statement_txn")
 
-		dbt.mustQueryContext(ctx, multiStmtQuery)
+		rows := dbt.mustQueryContext(ctx, multiStmtQuery)
+		defer rows.Close()
 	})
 }
 
@@ -151,7 +159,7 @@ func TestMultiStatementExecuteMix(t *testing.T) {
 		"insert into test_multi values (1), (2);\n" +
 		"select cola from test_multi order by cola asc;"
 
-	runTests(t, dsn, func(dbt *DBTest) {
+	runDBTest(t, func(dbt *DBTest) {
 		dbt.mustExec("drop table if exists test_multi_statement_txn")
 		dbt.mustExec(`create or replace table test_multi_statement_txn(
 			c1 number, c2 string) as select 10, 'z'`)
@@ -175,7 +183,7 @@ func TestMultiStatementQueryMix(t *testing.T) {
 		"select cola from test_multi order by cola asc;"
 
 	var count, v int
-	runTests(t, dsn, func(dbt *DBTest) {
+	runDBTest(t, func(dbt *DBTest) {
 		dbt.mustExec("drop table if exists test_multi_statement_txn")
 		dbt.mustExec(`create or replace table test_multi_statement_txn(
 			c1 number, c2 string) as select 10, 'z'`)
@@ -222,7 +230,7 @@ func TestMultiStatementCountZero(t *testing.T) {
 	var v3 float64
 	var v4 bool
 
-	runTests(t, dsn, func(dbt *DBTest) {
+	runDBTest(t, func(dbt *DBTest) {
 		// first query
 		multiStmtQuery1 := "select 123;\n" +
 			"select '456';"
@@ -322,8 +330,8 @@ func TestMultiStatementCountZero(t *testing.T) {
 }
 
 func TestMultiStatementCountMismatch(t *testing.T) {
-	db := openDB(t)
-	defer db.Close()
+	conn := openConn(t)
+	defer conn.Close()
 
 	multiStmtQuery := "select 123;\n" +
 		"select 456;\n" +
@@ -331,7 +339,7 @@ func TestMultiStatementCountMismatch(t *testing.T) {
 		"select '000';"
 
 	ctx, _ := WithMultiStatement(context.Background(), 3)
-	if _, err := db.QueryContext(ctx, multiStmtQuery); err == nil {
+	if _, err := conn.QueryContext(ctx, multiStmtQuery); err == nil {
 		t.Fatal("should have failed to query multiple statements")
 	}
 }
@@ -342,7 +350,7 @@ func TestMultiStatementVaryingColumnCount(t *testing.T) {
 	ctx, _ := WithMultiStatement(context.Background(), 0)
 
 	var v1, v2 int
-	runTests(t, dsn, func(dbt *DBTest) {
+	runDBTest(t, func(dbt *DBTest) {
 		dbt.mustExec("create or replace table test_tbl(c1 int, c2 int)")
 		dbt.mustExec("insert into test_tbl values(1, 0)")
 		defer dbt.mustExec("drop table if exists test_tbl")
@@ -374,6 +382,194 @@ func TestMultiStatementVaryingColumnCount(t *testing.T) {
 			}
 		} else {
 			t.Error("failed to query")
+		}
+	})
+}
+
+// The total completion time should be similar to the duration of the query on Snowflake UI.
+func TestMultiStatementExecutePerformance(t *testing.T) {
+	ctx, _ := WithMultiStatement(context.Background(), 100)
+	runDBTest(t, func(dbt *DBTest) {
+		file, err := os.Open("test_data/multistatements.sql")
+		if err != nil {
+			t.Fatalf("failed opening file: %s", err)
+		}
+		defer file.Close()
+		statements, err := io.ReadAll(file)
+		if err != nil {
+			t.Fatalf("failed reading file: %s", err)
+		}
+
+		sql := string(statements)
+
+		start := time.Now()
+		res := dbt.mustExecContext(ctx, sql)
+		duration := time.Since(start)
+
+		count, err := res.RowsAffected()
+		if err != nil {
+			t.Fatalf("res.RowsAffected() returned error: %v", err)
+		}
+		if count != 0 {
+			t.Fatalf("expected 0 affected rows, got %d", count)
+		}
+		t.Logf("The total completion time was %v", duration)
+
+		file, err = os.Open("test_data/multistatements_drop.sql")
+		if err != nil {
+			t.Fatalf("failed opening file: %s", err)
+		}
+		defer file.Close()
+		statements, err = io.ReadAll(file)
+		if err != nil {
+			t.Fatalf("failed reading file: %s", err)
+		}
+		sql = string(statements)
+		dbt.mustExecContext(ctx, sql)
+	})
+}
+
+func TestUnitGetChildResults(t *testing.T) {
+	testcases := []struct {
+		ids   string
+		types string
+		out   []childResult
+	}{
+		{"", "", nil},
+		{"", "4096", nil},
+		{"01aa3265-0405-ab7c-0000-53b106343aba,02aa3265-0405-ab7c-0000-53b106343aba", "12544,12544", []childResult{
+			{"01aa3265-0405-ab7c-0000-53b106343aba", "12544"},
+			{"02aa3265-0405-ab7c-0000-53b106343aba", "12544"}}},
+		{"01aa3265-0405-ab7c-0000-53b106343aba,02aa3265-0405-ab7c-0000-53b106343aba,03aa3265-0405-ab7c-0000-53b106343aba", "25344,4096,12544", []childResult{
+			{"01aa3265-0405-ab7c-0000-53b106343aba", "25344"},
+			{"02aa3265-0405-ab7c-0000-53b106343aba", "4096"},
+			{"03aa3265-0405-ab7c-0000-53b106343aba", "12544"}}},
+	}
+	for _, test := range testcases {
+		t.Run(test.ids, func(t *testing.T) {
+			res := getChildResults(test.ids, test.types)
+			if !reflect.DeepEqual(res, test.out) {
+				t.Fatalf("Child result should be equal, expected %v, actual %v", res, test.out)
+			}
+		})
+	}
+}
+
+func funcGetQueryRespFail(_ context.Context, _ *snowflakeRestful, _ *url.URL, _ map[string]string, _ time.Duration) (*http.Response, error) {
+	return nil, errors.New("failed to get query response")
+}
+
+func funcGetQueryRespError(_ context.Context, _ *snowflakeRestful, _ *url.URL, _ map[string]string, _ time.Duration) (*http.Response, error) {
+	dd := &execResponseData{}
+	er := &execResponse{
+		Data:    *dd,
+		Message: "query failed",
+		Code:    "261000",
+		Success: false,
+	}
+	ba, err := json.Marshal(er)
+	if err != nil {
+		panic(err)
+	}
+
+	return &http.Response{
+		StatusCode: http.StatusOK,
+		Body:       &fakeResponseBody{body: ba},
+	}, nil
+}
+
+func TestUnitHandleMultiExec(t *testing.T) {
+	runSnowflakeConnTest(t, func(sct *SCTest) {
+		data := execResponseData{
+			ResultIDs:   "",
+			ResultTypes: "",
+		}
+		_, err := sct.sc.handleMultiExec(context.Background(), data)
+		if err == nil {
+			t.Fatalf("should have failed")
+		}
+		driverErr, ok := err.(*SnowflakeError)
+		if !ok {
+			t.Fatalf("should be snowflake error. err: %v", err)
+		}
+		if driverErr.Number != ErrNoResultIDs {
+			t.Fatalf("unexpected error code. expected: %v, got: %v", ErrNoResultIDs, driverErr.Number)
+		}
+
+		data = execResponseData{
+			ResultIDs:   "1eFhmhe23242kmfd540GgGre,1eFhmhe23242kmfd540GgGre",
+			ResultTypes: "12544,12544",
+		}
+		sct.sc.rest = &snowflakeRestful{
+			FuncGet:          funcGetQueryRespFail,
+			FuncCloseSession: closeSessionMock,
+			TokenAccessor:    getSimpleTokenAccessor(),
+		}
+		_, err = sct.sc.handleMultiExec(context.Background(), data)
+		if err == nil {
+			t.Fatalf("should have failed")
+		}
+
+		sct.sc.rest.FuncGet = funcGetQueryRespError
+		data.SQLState = "01112"
+		_, err = sct.sc.handleMultiExec(context.Background(), data)
+		if err == nil {
+			t.Fatalf("should have failed")
+		}
+		driverErr, ok = err.(*SnowflakeError)
+		if !ok {
+			t.Fatalf("should be snowflake error. err: %v", err)
+		}
+		if driverErr.Number != ErrFailedToPostQuery {
+			t.Fatalf("unexpected error code. expected: %v, got: %v", ErrFailedToPostQuery, driverErr.Number)
+		}
+	})
+}
+
+func TestUnitHandleMultiQuery(t *testing.T) {
+	runSnowflakeConnTest(t, func(sct *SCTest) {
+		data := execResponseData{
+			ResultIDs:   "",
+			ResultTypes: "",
+		}
+		rows := new(snowflakeRows)
+		err := sct.sc.handleMultiQuery(context.Background(), data, rows)
+		if err == nil {
+			t.Fatalf("should have failed")
+		}
+		driverErr, ok := err.(*SnowflakeError)
+		if !ok {
+			t.Fatalf("should be snowflake error. err: %v", err)
+		}
+		if driverErr.Number != ErrNoResultIDs {
+			t.Fatalf("unexpected error code. expected: %v, got: %v", ErrNoResultIDs, driverErr.Number)
+		}
+		data = execResponseData{
+			ResultIDs:   "1eFhmhe23242kmfd540GgGre,1eFhmhe23242kmfd540GgGre",
+			ResultTypes: "12544,12544",
+		}
+		sct.sc.rest = &snowflakeRestful{
+			FuncGet:          funcGetQueryRespFail,
+			FuncCloseSession: closeSessionMock,
+			TokenAccessor:    getSimpleTokenAccessor(),
+		}
+		err = sct.sc.handleMultiQuery(context.Background(), data, rows)
+		if err == nil {
+			t.Fatalf("should have failed")
+		}
+
+		sct.sc.rest.FuncGet = funcGetQueryRespError
+		data.SQLState = "01112"
+		err = sct.sc.handleMultiQuery(context.Background(), data, rows)
+		if err == nil {
+			t.Fatalf("should have failed")
+		}
+		driverErr, ok = err.(*SnowflakeError)
+		if !ok {
+			t.Fatalf("should be snowflake error. err: %v", err)
+		}
+		if driverErr.Number != ErrFailedToPostQuery {
+			t.Fatalf("unexpected error code. expected: %v, got: %v", ErrFailedToPostQuery, driverErr.Number)
 		}
 	})
 }
