@@ -1,4 +1,4 @@
-// Copyright (c) 2017-2021 Snowflake Computing Inc. All right reserved.
+// Copyright (c) 2017-2022 Snowflake Computing Inc. All rights reserved.
 
 package gosnowflake
 
@@ -12,7 +12,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"net"
 	"net/http"
 	"net/url"
@@ -31,11 +30,11 @@ func TestOCSP(t *testing.T) {
 	targetURL := []string{
 		"https://sfctest0.snowflakecomputing.com/",
 		"https://s3-us-west-2.amazonaws.com/sfc-snowsql-updates/?prefix=1.1/windows_x86_64",
-		"https://sfcdev1.blob.core.windows.net/",
+		"https://sfcdev2.blob.core.windows.net/",
 	}
 
 	transports := []*http.Transport{
-		snowflakeInsecureTransport,
+		snowflakeNoOcspTransport,
 		SnowflakeTransport,
 	}
 
@@ -43,26 +42,27 @@ func TestOCSP(t *testing.T) {
 		for _, tgt := range targetURL {
 			_ = os.Setenv(cacheServerEnabledEnv, enabled)
 			_ = os.Remove(cacheFileName) // clear cache file
-			ocspResponseCache = make(map[certIDKey][]interface{})
+			ocspResponseCache = make(map[certIDKey]*certCacheValue)
 			for _, tr := range transports {
-				c := &http.Client{
-					Transport: tr,
-					Timeout:   30 * time.Second,
-				}
-				req, err := http.NewRequest("GET", tgt, bytes.NewReader(nil))
-				if err != nil {
-					t.Fatalf("fail to create a request. err: %v", err)
-				}
-				res, err := c.Do(req)
-				if err != nil {
-					t.Fatalf("failed to GET contents. err: %v", err)
-				}
-				defer res.Body.Close()
-				_, err = ioutil.ReadAll(res.Body)
-				if err != nil {
-					t.Fatalf("failed to read content body for %v", tgt)
-				}
-
+				t.Run(fmt.Sprintf("%v_%v", tgt, enabled), func(t *testing.T) {
+					c := &http.Client{
+						Transport: tr,
+						Timeout:   30 * time.Second,
+					}
+					req, err := http.NewRequest("GET", tgt, bytes.NewReader(nil))
+					if err != nil {
+						t.Fatalf("fail to create a request. err: %v", err)
+					}
+					res, err := c.Do(req)
+					if err != nil {
+						t.Fatalf("failed to GET contents. err: %v", err)
+					}
+					defer res.Body.Close()
+					_, err = io.ReadAll(res.Body)
+					if err != nil {
+						t.Fatalf("failed to read content body for %v", tgt)
+					}
+				})
 			}
 		}
 	}
@@ -116,9 +116,11 @@ func TestUnitIsInValidityRange(t *testing.T) {
 		},
 	}
 	for _, tc := range testcases {
-		if tc.ret != isInValidityRange(currentTime, tc.thisTime, tc.nextTime) {
-			t.Fatalf("failed to check validity. should be: %v, currentTime: %v, thisTime: %v, nextTime: %v", tc.ret, currentTime, tc.thisTime, tc.nextTime)
-		}
+		t.Run(fmt.Sprintf("%v_%v", tc.thisTime, tc.nextTime), func(t *testing.T) {
+			if tc.ret != isInValidityRange(currentTime, tc.thisTime, tc.nextTime) {
+				t.Fatalf("failed to check validity. should be: %v, currentTime: %v, thisTime: %v, nextTime: %v", tc.ret, currentTime, tc.thisTime, tc.nextTime)
+			}
+		})
 	}
 }
 
@@ -126,46 +128,48 @@ func TestUnitEncodeCertIDGood(t *testing.T) {
 	targetURLs := []string{
 		"faketestaccount.snowflakecomputing.com:443",
 		"s3-us-west-2.amazonaws.com:443",
-		"sfcdev1.blob.core.windows.net:443",
+		"sfcdev2.blob.core.windows.net:443",
 	}
 	for _, tt := range targetURLs {
-		chainedCerts := getCert(tt)
-		for i := 0; i < len(chainedCerts)-1; i++ {
-			subject := chainedCerts[i]
-			issuer := chainedCerts[i+1]
-			ocspServers := subject.OCSPServer
-			if len(ocspServers) == 0 {
-				t.Fatalf("no OCSP server is found. cert: %v", subject.Subject)
+		t.Run(tt, func(t *testing.T) {
+			chainedCerts := getCert(tt)
+			for i := 0; i < len(chainedCerts)-1; i++ {
+				subject := chainedCerts[i]
+				issuer := chainedCerts[i+1]
+				ocspServers := subject.OCSPServer
+				if len(ocspServers) == 0 {
+					t.Fatalf("no OCSP server is found. cert: %v", subject.Subject)
+				}
+				ocspReq, err := ocsp.CreateRequest(subject, issuer, &ocsp.RequestOptions{})
+				if err != nil {
+					t.Fatalf("failed to create OCSP request. err: %v", err)
+				}
+				var ost *ocspStatus
+				_, ost = extractCertIDKeyFromRequest(ocspReq)
+				if ost.err != nil {
+					t.Fatalf("failed to extract cert ID from the OCSP request. err: %v", ost.err)
+				}
+				// better hash. Not sure if the actual OCSP server accepts this, though.
+				ocspReq, err = ocsp.CreateRequest(subject, issuer, &ocsp.RequestOptions{Hash: crypto.SHA512})
+				if err != nil {
+					t.Fatalf("failed to create OCSP request. err: %v", err)
+				}
+				_, ost = extractCertIDKeyFromRequest(ocspReq)
+				if ost.err != nil {
+					t.Fatalf("failed to extract cert ID from the OCSP request. err: %v", ost.err)
+				}
+				// tweaked request binary
+				ocspReq, err = ocsp.CreateRequest(subject, issuer, &ocsp.RequestOptions{Hash: crypto.SHA512})
+				if err != nil {
+					t.Fatalf("failed to create OCSP request. err: %v", err)
+				}
+				ocspReq[10] = 0 // random change
+				_, ost = extractCertIDKeyFromRequest(ocspReq)
+				if ost.err == nil {
+					t.Fatal("should have failed")
+				}
 			}
-			ocspReq, err := ocsp.CreateRequest(subject, issuer, &ocsp.RequestOptions{})
-			if err != nil {
-				t.Fatalf("failed to create OCSP request. err: %v", err)
-			}
-			var ost *ocspStatus
-			_, ost = extractCertIDKeyFromRequest(ocspReq)
-			if ost.err != nil {
-				t.Fatalf("failed to extract cert ID from the OCSP request. err: %v", ost.err)
-			}
-			// better hash. Not sure if the actual OCSP server accepts this, though.
-			ocspReq, err = ocsp.CreateRequest(subject, issuer, &ocsp.RequestOptions{Hash: crypto.SHA512})
-			if err != nil {
-				t.Fatalf("failed to create OCSP request. err: %v", err)
-			}
-			_, ost = extractCertIDKeyFromRequest(ocspReq)
-			if ost.err != nil {
-				t.Fatalf("failed to extract cert ID from the OCSP request. err: %v", ost.err)
-			}
-			// tweaked request binary
-			ocspReq, err = ocsp.CreateRequest(subject, issuer, &ocsp.RequestOptions{Hash: crypto.SHA512})
-			if err != nil {
-				t.Fatalf("failed to create OCSP request. err: %v", err)
-			}
-			ocspReq[10] = 0 // random change
-			_, ost = extractCertIDKeyFromRequest(ocspReq)
-			if ost.err == nil {
-				t.Fatal("should have failed")
-			}
-		}
+		})
 	}
 }
 
@@ -184,7 +188,7 @@ func TestUnitCheckOCSPResponseCache(t *testing.T) {
 	}
 	b64Key := base64.StdEncoding.EncodeToString([]byte("DUMMY_VALUE"))
 	currentTime := float64(time.Now().UTC().Unix())
-	ocspResponseCache[dummyKey0] = []interface{}{currentTime, b64Key}
+	ocspResponseCache[dummyKey0] = &certCacheValue{currentTime, b64Key}
 	subject := &x509.Certificate{}
 	issuer := &x509.Certificate{}
 	ost := checkOCSPResponseCache(&dummyKey, subject, issuer)
@@ -192,48 +196,36 @@ func TestUnitCheckOCSPResponseCache(t *testing.T) {
 		t.Fatalf("should have failed. expected: %v, got: %v", ocspMissedCache, ost.code)
 	}
 	// old timestamp
-	ocspResponseCache[dummyKey] = []interface{}{float64(1395054952), b64Key}
+	ocspResponseCache[dummyKey] = &certCacheValue{float64(1395054952), b64Key}
 	ost = checkOCSPResponseCache(&dummyKey, subject, issuer)
 	if ost.code != ocspCacheExpired {
 		t.Fatalf("should have failed. expected: %v, got: %v", ocspCacheExpired, ost.code)
 	}
 	// future timestamp
-	ocspResponseCache[dummyKey] = []interface{}{float64(1805054952), b64Key}
+	ocspResponseCache[dummyKey] = &certCacheValue{float64(1805054952), b64Key}
 	ost = checkOCSPResponseCache(&dummyKey, subject, issuer)
 	if ost.code != ocspFailedParseResponse {
 		t.Fatalf("should have failed. expected: %v, got: %v", ocspFailedDecodeResponse, ost.code)
 	}
 	// actual OCSP but it fails to parse, because an invalid issuer certificate is given.
-	actualOcspResponse := "MIIB0woBAKCCAcwwggHIBgkrBgEFBQcwAQEEggG5MIIBtTCBnqIWBBSxPsNpA/i/RwHUmCYaCALvY2QrwxgPMjAxNz" +
-		"A1MTYyMjAwMDBaMHMwcTBJMAkGBSsOAwIaBQAEFN+qEuMosQlBk+KfQoLOR0BClVijBBSxPsNpA/i/RwHUmCYaCALvY2QrwwIQBOHnp" +
-		"Nxc8vNtwCtCuF0Vn4AAGA8yMDE3MDUxNjIyMDAwMFqgERgPMjAxNzA1MjMyMjAwMDBaMA0GCSqGSIb3DQEBCwUAA4IBAQCuRGwqQsKy" +
-		"IAAGHgezTfG0PzMYgGD/XRDhU+2i08WTJ4Zs40Lu88cBeRXWF3iiJSpiX3/OLgfI7iXmHX9/sm2SmeNWc0Kb39bk5Lw1jwezf8hcI9+" +
-		"mZHt60vhUgtgZk21SsRlTZ+S4VXwtDqB1Nhv6cnSnfrL2A9qJDZS2ltPNOwebWJnznDAs2dg+KxmT2yBXpHM1kb0EOolWvNgORbgIgB" +
-		"koRzw/UU7zKsqiTB0ZN/rgJp+MocTdqQSGKvbZyR8d4u8eNQqi1x4Pk3yO/pftANFaJKGB+JPgKS3PQAqJaXcipNcEfqtl7y4PO6kqA" +
+	actualOcspResponse := "MIIB0woBAKCCAcwwggHIBgkrBgEFBQcwAQEEggG5MIIBtTCBnqIWBBSxPsNpA/i/RwHUmCYaCALvY2QrwxgPMjAxNz" + // pragma: allowlist secret
+		"A1MTYyMjAwMDBaMHMwcTBJMAkGBSsOAwIaBQAEFN+qEuMosQlBk+KfQoLOR0BClVijBBSxPsNpA/i/RwHUmCYaCALvY2QrwwIQBOHnp" + // pragma: allowlist secret
+		"Nxc8vNtwCtCuF0Vn4AAGA8yMDE3MDUxNjIyMDAwMFqgERgPMjAxNzA1MjMyMjAwMDBaMA0GCSqGSIb3DQEBCwUAA4IBAQCuRGwqQsKy" + // pragma: allowlist secret
+		"IAAGHgezTfG0PzMYgGD/XRDhU+2i08WTJ4Zs40Lu88cBeRXWF3iiJSpiX3/OLgfI7iXmHX9/sm2SmeNWc0Kb39bk5Lw1jwezf8hcI9+" + // pragma: allowlist secret
+		"mZHt60vhUgtgZk21SsRlTZ+S4VXwtDqB1Nhv6cnSnfrL2A9qJDZS2ltPNOwebWJnznDAs2dg+KxmT2yBXpHM1kb0EOolWvNgORbgIgB" + // pragma: allowlist secret
+		"koRzw/UU7zKsqiTB0ZN/rgJp+MocTdqQSGKvbZyR8d4u8eNQqi1x4Pk3yO/pftANFaJKGB+JPgKS3PQAqJaXcipNcEfqtl7y4PO6kqA" + // pragma: allowlist secret
 		"Jb4xI/OTXIrRA5TsT4cCioE"
 	// issuer is not a true issuer certificate
-	ocspResponseCache[dummyKey] = []interface{}{float64(currentTime - 1000), actualOcspResponse}
+	ocspResponseCache[dummyKey] = &certCacheValue{float64(currentTime - 1000), actualOcspResponse}
 	ost = checkOCSPResponseCache(&dummyKey, subject, issuer)
 	if ost.code != ocspFailedParseResponse {
 		t.Fatalf("should have failed. expected: %v, got: %v", ocspFailedParseResponse, ost.code)
 	}
 	// invalid validity
-	ocspResponseCache[dummyKey] = []interface{}{float64(currentTime - 1000), actualOcspResponse}
+	ocspResponseCache[dummyKey] = &certCacheValue{float64(currentTime - 1000), actualOcspResponse}
 	ost = checkOCSPResponseCache(&dummyKey, subject, nil)
 	if ost.code != ocspInvalidValidity {
 		t.Fatalf("should have failed. expected: %v, got: %v", ocspInvalidValidity, ost.code)
-	}
-	// wrong timestamp type
-	ocspResponseCache[dummyKey] = []interface{}{uint32(currentTime - 1000), 123456}
-	ost = checkOCSPResponseCache(&dummyKey, subject, issuer)
-	if ost.code != ocspFailedDecodeResponse {
-		t.Fatalf("should have failed. expected: %v, got: %v", ocspFailedDecodeResponse, ost.code)
-	}
-	// wrong value type
-	ocspResponseCache[dummyKey] = []interface{}{float64(currentTime - 1000), 123456}
-	ost = checkOCSPResponseCache(&dummyKey, subject, issuer)
-	if ost.code != ocspFailedDecodeResponse {
-		t.Fatalf("should have failed. expected: %v, got: %v", ocspFailedDecodeResponse, ost.code)
 	}
 }
 
@@ -314,8 +306,8 @@ func TestOCSPRetry(t *testing.T) {
 		body:    []byte{1, 2, 3},
 	}
 	res, b, st := retryOCSP(
-		context.TODO(),
-		client, fakeRequestFunc,
+		context.Background(),
+		client, emptyRequest,
 		dummyOCSPHost,
 		make(map[string]string), []byte{0}, certs[len(certs)-1], 10*time.Second)
 	if st.err == nil {
@@ -327,13 +319,51 @@ func TestOCSPRetry(t *testing.T) {
 		body:    []byte{1, 2, 3},
 	}
 	res, b, st = retryOCSP(
-		context.TODO(),
+		context.Background(),
 		client, fakeRequestFunc,
 		dummyOCSPHost,
 		make(map[string]string), []byte{0}, certs[len(certs)-1], 5*time.Second)
 	if st.err == nil {
 		fmt.Printf("should fail: %v, %v, %v\n", res, b, st)
 	}
+}
+
+func TestFullOCSPURL(t *testing.T) {
+	testcases := []tcFullOCSPURL{
+		{
+			url:               &url.URL{Host: "some-ocsp-url.com"},
+			expectedURLString: "some-ocsp-url.com",
+		},
+		{
+			url: &url.URL{
+				Host: "some-ocsp-url.com",
+				Path: "/some-path",
+			},
+			expectedURLString: "some-ocsp-url.com/some-path",
+		},
+		{
+			url: &url.URL{
+				Host: "some-ocsp-url.com",
+				Path: "some-path",
+			},
+			expectedURLString: "some-ocsp-url.com/some-path",
+		},
+	}
+
+	for _, testcase := range testcases {
+		t.Run("", func(t *testing.T) {
+			returnedStringURL := fullOCSPURL(testcase.url)
+			if returnedStringURL != testcase.expectedURLString {
+				t.Fatalf("failed to match returned OCSP url string; expected: %v, got: %v",
+					testcase.expectedURLString, returnedStringURL)
+			}
+		})
+	}
+}
+
+type tcFullOCSPURL struct {
+	url               *url.URL
+	expectedURLString string
 }
 
 func TestOCSPCacheServerRetry(t *testing.T) {
@@ -347,7 +377,7 @@ func TestOCSPCacheServerRetry(t *testing.T) {
 		body:    []byte{1, 2, 3},
 	}
 	res, st := checkOCSPCacheServer(
-		context.TODO(), client, fakeRequestFunc, dummyOCSPHost, 20*time.Second)
+		context.Background(), client, fakeRequestFunc, dummyOCSPHost, 20*time.Second)
 	if st.err == nil {
 		t.Errorf("should fail: %v", res)
 	}
@@ -357,7 +387,7 @@ func TestOCSPCacheServerRetry(t *testing.T) {
 		body:    []byte{1, 2, 3},
 	}
 	res, st = checkOCSPCacheServer(
-		context.TODO(), client, fakeRequestFunc, dummyOCSPHost, 10*time.Second)
+		context.Background(), client, fakeRequestFunc, dummyOCSPHost, 10*time.Second)
 	if st.err == nil {
 		t.Errorf("should fail: %v", res)
 	}
@@ -452,20 +482,22 @@ func TestCanEarlyExitForOCSP(t *testing.T) {
 	}
 
 	for idx, tt := range testcases {
-		ocspFailOpen = OCSPFailOpenTrue
-		expectedLen := len(tt.results)
-		if tt.resultLen > 0 {
-			expectedLen = tt.resultLen
-		}
-		r := canEarlyExitForOCSP(tt.results, expectedLen)
-		if !(tt.retFailOpen == nil && r == nil) && !(tt.retFailOpen != nil && r != nil && tt.retFailOpen.code == r.code) {
-			t.Fatalf("%d: failed to match return. expected: %v, got: %v", idx, tt.retFailOpen, r)
-		}
-		ocspFailOpen = OCSPFailOpenFalse
-		r = canEarlyExitForOCSP(tt.results, expectedLen)
-		if !(tt.retFailClosed == nil && r == nil) && !(tt.retFailClosed != nil && r != nil && tt.retFailClosed.code == r.code) {
-			t.Fatalf("%d: failed to match return. expected: %v, got: %v", idx, tt.retFailClosed, r)
-		}
+		t.Run("", func(t *testing.T) {
+			ocspFailOpen = OCSPFailOpenTrue
+			expectedLen := len(tt.results)
+			if tt.resultLen > 0 {
+				expectedLen = tt.resultLen
+			}
+			r := canEarlyExitForOCSP(tt.results, expectedLen)
+			if !(tt.retFailOpen == nil && r == nil) && !(tt.retFailOpen != nil && r != nil && tt.retFailOpen.code == r.code) {
+				t.Fatalf("%d: failed to match return. expected: %v, got: %v", idx, tt.retFailOpen, r)
+			}
+			ocspFailOpen = OCSPFailOpenFalse
+			r = canEarlyExitForOCSP(tt.results, expectedLen)
+			if !(tt.retFailClosed == nil && r == nil) && !(tt.retFailClosed != nil && r != nil && tt.retFailClosed.code == r.code) {
+				t.Fatalf("%d: failed to match return. expected: %v, got: %v", idx, tt.retFailClosed, r)
+			}
+		})
 	}
 }
 
@@ -516,7 +548,7 @@ func TestInitOCSPCacheFileCreation(t *testing.T) {
 	defer func() {
 		src, _ = os.Open(tmpFileName)
 		defer src.Close()
-		dst, _ = os.OpenFile(srcFileName, os.O_WRONLY, os.ModePerm)
+		dst, _ = os.OpenFile(srcFileName, os.O_WRONLY, readWriteFileMode)
 		defer dst.Close()
 		// copy temporary file contents back to original file
 		if _, err = io.Copy(dst, src); err != nil {
