@@ -1,4 +1,4 @@
-// Copyright (c) 2021 Snowflake Computing Inc. All right reserved.
+// Copyright (c) 2021-2022 Snowflake Computing Inc. All rights reserved.
 
 package gosnowflake
 
@@ -9,7 +9,8 @@ import (
 )
 
 const (
-	forceJSON = "ALTER SESSION SET GO_QUERY_RESULT_FORMAT = JSON"
+	forceARROW = "ALTER SESSION SET GO_QUERY_RESULT_FORMAT = ARROW"
+	forceJSON  = "ALTER SESSION SET GO_QUERY_RESULT_FORMAT = JSON"
 )
 
 func TestJSONInt(t *testing.T) {
@@ -25,7 +26,7 @@ func TestJSONFloat64(t *testing.T) {
 }
 
 func TestJSONVariousTypes(t *testing.T) {
-	runTests(t, dsn, func(dbt *DBTest) {
+	runDBTest(t, func(dbt *DBTest) {
 		dbt.mustExec(forceJSON)
 		rows := dbt.mustQuery(selectVariousTypes)
 		defer rows.Close()
@@ -178,7 +179,7 @@ func TestLargeSetResultWithCustomJSONDecoder(t *testing.T) {
 }
 
 func TestBindingJSONInterface(t *testing.T) {
-	runTests(t, dsn, func(dbt *DBTest) {
+	runDBTest(t, func(dbt *DBTest) {
 		dbt.mustExec(forceJSON)
 		rows := dbt.mustQuery(selectVariousTypes)
 		defer rows.Close()
