@@ -1,10 +1,9 @@
-// Copyright (c) 2017-2021 Snowflake Computing Inc. All right reserved.
+// Copyright (c) 2017-2022 Snowflake Computing Inc. All rights reserved.
 
 package gosnowflake
 
-//lint:file-ignore U1000 Ignore all unused code
-
 import (
+	"encoding/json"
 	"time"
 )
 
@@ -16,8 +15,10 @@ const (
 )
 
 type execBindParameter struct {
-	Type  string      `json:"type"`
-	Value interface{} `json:"value"`
+	Type   string         `json:"type"`
+	Value  interface{}    `json:"value"`
+	Format string         `json:"fmt,omitempty"`
+	Schema *bindingSchema `json:"schema,omitempty"`
 }
 
 type execRequest struct {
@@ -29,16 +30,55 @@ type execRequest struct {
 	Parameters   map[string]interface{}       `json:"parameters,omitempty"`
 	Bindings     map[string]execBindParameter `json:"bindings,omitempty"`
 	BindStage    string                       `json:"bindStage,omitempty"`
+	QueryContext requestQueryContext          `json:"queryContextDTO,omitempty"`
+}
+
+type requestQueryContext struct {
+	Entries []requestQueryContextEntry `json:"entries,omitempty"`
+}
+
+type requestQueryContextEntry struct {
+	Context   contextData `json:"context,omitempty"`
+	ID        int         `json:"id"`
+	Priority  int         `json:"priority"`
+	Timestamp int64       `json:"timestamp,omitempty"`
+}
+
+type contextData struct {
+	Base64Data string `json:"base64Data,omitempty"`
 }
 
 type execResponseRowType struct {
-	Name       string `json:"name"`
-	ByteLength int64  `json:"byteLength"`
-	Length     int64  `json:"length"`
-	Type       string `json:"type"`
-	Precision  int64  `json:"precision"`
-	Scale      int64  `json:"scale"`
-	Nullable   bool   `json:"nullable"`
+	Name       string          `json:"name"`
+	Fields     []fieldMetadata `json:"fields"`
+	ByteLength int64           `json:"byteLength"`
+	Length     int64           `json:"length"`
+	Type       string          `json:"type"`
+	Precision  int64           `json:"precision"`
+	Scale      int64           `json:"scale"`
+	Nullable   bool            `json:"nullable"`
+}
+
+func (ex *execResponseRowType) toFieldMetadata() fieldMetadata {
+	return fieldMetadata{
+		ex.Name,
+		ex.Type,
+		ex.Nullable,
+		int(ex.Length),
+		int(ex.Scale),
+		int(ex.Precision),
+		ex.Fields,
+	}
+}
+
+type fieldMetadata struct {
+	Name      string          `json:"name,omitempty"`
+	Type      string          `json:"type"`
+	Nullable  bool            `json:"nullable"`
+	Length    int             `json:"length"`
+	Scale     int             `json:"scale"`
+	Precision int             `json:"precision"`
+	Fields    []fieldMetadata `json:"fields,omitempty"`
 }
 
 type execResponseChunk struct {
@@ -68,6 +108,8 @@ type execResponseStageInfo struct {
 	Creds                 execResponseCredentials `json:"creds,omitempty"`
 	PresignedURL          string                  `json:"presignedUrl,omitempty"`
 	EndPoint              string                  `json:"endPoint,omitempty"`
+	UseS3RegionalURL      bool                    `json:"useS3RegionalUrl,omitempty"`
+	UseRegionalURL        bool                    `json:"useRegionalUrl,omitempty"`
 }
 
 // make all data field optional
@@ -112,6 +154,7 @@ type execResponseData struct {
 	Parallel                int64                 `json:"parallel,omitempty"`
 	Threshold               int64                 `json:"threshold,omitempty"`
 	AutoCompress            bool                  `json:"autoCompress,omitempty"`
+	Overwrite               bool                  `json:"overwrite,omitempty"`
 	SourceCompression       string                `json:"sourceCompression,omitempty"`
 	ShowEncryptionParameter bool                  `json:"clientShowEncryptionParameter,omitempty"`
 	EncryptionMaterial      encryptionWrapper     `json:"encryptionMaterial,omitempty"`
@@ -120,6 +163,9 @@ type execResponseData struct {
 	Command                 string                `json:"command,omitempty"`
 	Kind                    string                `json:"kind,omitempty"`
 	Operation               string                `json:"operation,omitempty"`
+
+	// HTAP
+	QueryContext json.RawMessage `json:"queryContext,omitempty"`
 }
 
 type execResponse struct {
