@@ -1,12 +1,14 @@
-// Copyright (c) 2017-2021 Snowflake Computing Inc. All right reserved.
+// Copyright (c) 2017-2022 Snowflake Computing Inc. All rights reserved.
 
 package gosnowflake
 
 import (
+	"context"
 	"database/sql/driver"
 	"io"
 	"reflect"
 	"strings"
+	"time"
 )
 
 const (
@@ -27,6 +29,13 @@ var (
 	maxChunkDownloaderErrorCounter = 5
 )
 
+// SnowflakeRows provides an API for methods exposed to the clients
+type SnowflakeRows interface {
+	GetQueryID() string
+	GetStatus() queryStatus
+	GetArrowBatches() ([]*ArrowBatch, error)
+}
+
 type snowflakeRows struct {
 	sc                  *snowflakeConn
 	ChunkDownloader     chunkDownloader
@@ -35,6 +44,16 @@ type snowflakeRows struct {
 	status              queryStatus
 	err                 error
 	errChannel          chan error
+	location            *time.Location
+	ctx                 context.Context
+	format              resultFormat
+}
+
+func (rows *snowflakeRows) getLocation() *time.Location {
+	if rows.location == nil && rows.sc != nil && rows.sc.cfg != nil {
+		rows.location = getCurrentLocation(rows.sc.cfg.Params)
+	}
+	return rows.location
 }
 
 type snowflakeValue interface{}
@@ -119,7 +138,7 @@ func (rows *snowflakeRows) Columns() []string {
 	if err := rows.waitForAsyncQueryStatus(); err != nil {
 		return make([]string, 0)
 	}
-	logger.Debug("Rows.Columns")
+	logger.WithContext(rows.ctx).Debug("Rows.Columns")
 	ret := make([]string, len(rows.ChunkDownloader.getRowType()))
 	for i, n := 0, len(rows.ChunkDownloader.getRowType()); i < n; i++ {
 		ret[i] = rows.ChunkDownloader.getRowType()[i].Name
@@ -131,9 +150,7 @@ func (rows *snowflakeRows) ColumnTypeScanType(index int) reflect.Type {
 	if err := rows.waitForAsyncQueryStatus(); err != nil {
 		return nil
 	}
-	return snowflakeTypeToGo(
-		getSnowflakeType(strings.ToUpper(rows.ChunkDownloader.getRowType()[index].Type)),
-		rows.ChunkDownloader.getRowType()[index].Scale)
+	return snowflakeTypeToGo(rows.ctx, getSnowflakeType(rows.ChunkDownloader.getRowType()[index].Type), rows.ChunkDownloader.getRowType()[index].Scale, rows.ChunkDownloader.getRowType()[index].Fields)
 }
 
 func (rows *snowflakeRows) GetQueryID() string {
@@ -144,8 +161,23 @@ func (rows *snowflakeRows) GetStatus() queryStatus {
 	return rows.status
 }
 
-func (rows *snowflakeRows) Next(dest []driver.Value) (err error) {
+// GetArrowBatches returns an array of ArrowBatch objects to retrieve data in arrow.Record format
+func (rows *snowflakeRows) GetArrowBatches() ([]*ArrowBatch, error) {
+	// Wait for all arrow batches before fetching.
+	// Otherwise, a panic error "invalid memory address or nil pointer dereference" will be thrown.
 	if err := rows.waitForAsyncQueryStatus(); err != nil {
+		return nil, err
+	}
+
+	if rows.format != arrowFormat {
+		return nil, errNonArrowResponseForArrowBatches(rows.queryID).exceptionTelemetry(rows.sc)
+	}
+
+	return rows.ChunkDownloader.getArrowBatches(), nil
+}
+
+func (rows *snowflakeRows) Next(dest []driver.Value) (err error) {
+	if err = rows.waitForAsyncQueryStatus(); err != nil {
 		return err
 	}
 	row, err := rows.ChunkDownloader.next()
@@ -165,7 +197,7 @@ func (rows *snowflakeRows) Next(dest []driver.Value) (err error) {
 		for i, n := 0, len(row.RowSet); i < n; i++ {
 			// could move to chunk downloader so that each go routine
 			// can convert data
-			err := stringToValue(&dest[i], rows.ChunkDownloader.getRowType()[i], row.RowSet[i])
+			err = stringToValue(rows.ctx, &dest[i], rows.ChunkDownloader.getRowType()[i], row.RowSet[i], rows.getLocation(), rows.sc.cfg.Params)
 			if err != nil {
 				return err
 			}
@@ -190,7 +222,9 @@ func (rows *snowflakeRows) NextResultSet() error {
 			return io.EOF
 		}
 		rows.ChunkDownloader = rows.ChunkDownloader.getNextChunkDownloader()
-		rows.ChunkDownloader.start()
+		if err := rows.ChunkDownloader.start(); err != nil {
+			return err
+		}
 	}
 	return rows.ChunkDownloader.nextResultSet()
 }
