@@ -1,4 +1,4 @@
-// Copyright (c) 2021 Snowflake Computing Inc. All right reserved.
+// Copyright (c) 2021-2022 Snowflake Computing Inc. All rights reserved.
 
 package gosnowflake
 
@@ -19,15 +19,15 @@ const (
 
 // implemented by localUtil and remoteStorageUtil
 type storageUtil interface {
-	createClient(*execResponseStageInfo, bool) cloudClient
+	createClient(*execResponseStageInfo, bool, *Config) (cloudClient, error)
 	uploadOneFileWithRetry(*fileMetadata) error
 	downloadOneFile(*fileMetadata) error
 }
 
 // implemented by snowflakeS3Util, snowflakeAzureUtil and snowflakeGcsUtil
 type cloudUtil interface {
-	createClient(*execResponseStageInfo, bool) cloudClient
-	getFileHeader(*fileMetadata, string) *fileHeader
+	createClient(*execResponseStageInfo, bool) (cloudClient, error)
+	getFileHeader(*fileMetadata, string) (*fileHeader, error)
 	uploadFile(string, *fileMetadata, *encryptMetadata, int, int64) error
 	nativeDownloadFile(*fileMetadata, string, int64) error
 }
@@ -35,22 +35,29 @@ type cloudUtil interface {
 type cloudClient interface{}
 
 type remoteStorageUtil struct {
+	cfg *Config
 }
 
-func (rsu *remoteStorageUtil) getNativeCloudType(cli string) cloudUtil {
+func (rsu *remoteStorageUtil) getNativeCloudType(cli string, cfg *Config) cloudUtil {
 	if cloudType(cli) == s3Client {
-		return &snowflakeS3Util{}
+		return &snowflakeS3Client{
+			cfg,
+		}
 	} else if cloudType(cli) == azureClient {
-		return &snowflakeAzureUtil{}
+		return &snowflakeAzureClient{
+			cfg,
+		}
 	} else if cloudType(cli) == gcsClient {
-		return &snowflakeGcsUtil{}
+		return &snowflakeGcsClient{
+			cfg,
+		}
 	}
 	return nil
 }
 
 // call cloud utils' native create client methods
-func (rsu *remoteStorageUtil) createClient(info *execResponseStageInfo, useAccelerateEndpoint bool) cloudClient {
-	utilClass := rsu.getNativeCloudType(info.LocationType)
+func (rsu *remoteStorageUtil) createClient(info *execResponseStageInfo, useAccelerateEndpoint bool, cfg *Config) (cloudClient, error) {
+	utilClass := rsu.getNativeCloudType(info.LocationType, cfg)
 	return utilClass.createClient(info, useAccelerateEndpoint)
 }
 
@@ -65,14 +72,14 @@ func (rsu *remoteStorageUtil) uploadOneFile(meta *fileMetadata) error {
 			if meta.realSrcStream != nil {
 				srcStream = meta.realSrcStream
 			}
-			encryptMeta, err = encryptStream(meta.encryptionMaterial, srcStream, &encryptedStream, 0)
+			encryptMeta, err = encryptStreamCBC(meta.encryptionMaterial, srcStream, &encryptedStream, 0)
 			if err != nil {
 				return err
 			}
 			meta.realSrcStream = &encryptedStream
 			dataFile = meta.realSrcFileName
 		} else {
-			encryptMeta, dataFile, err = encryptFile(meta.encryptionMaterial, meta.realSrcFileName, 0, meta.tmpDir)
+			encryptMeta, dataFile, err = encryptFileCBC(meta.encryptionMaterial, meta.realSrcFileName, 0, meta.tmpDir)
 			if err != nil {
 				return err
 			}
@@ -81,13 +88,21 @@ func (rsu *remoteStorageUtil) uploadOneFile(meta *fileMetadata) error {
 		dataFile = meta.realSrcFileName
 	}
 
-	utilClass := rsu.getNativeCloudType(meta.stageInfo.LocationType)
+	utilClass := rsu.getNativeCloudType(meta.stageInfo.LocationType, meta.sfa.sc.cfg)
 	maxConcurrency := int(meta.parallel)
 	var lastErr error
 	maxRetry := defaultMaxRetry
 	for retry := 0; retry < maxRetry; retry++ {
 		if !meta.overwrite {
-			header := utilClass.getFileHeader(meta, meta.dstFileName)
+			header, err := utilClass.getFileHeader(meta, meta.dstFileName)
+			if meta.resStatus == notFoundFile {
+				err := utilClass.uploadFile(dataFile, meta, encryptMeta, maxConcurrency, meta.options.MultiPartThreshold)
+				if err != nil {
+					logger.Warnf("Error uploading %v. err: %v", dataFile, err)
+				}
+			} else if err != nil {
+				return err
+			}
 			if header != nil && meta.resStatus == uploaded {
 				meta.dstFileSize = 0
 				meta.resStatus = skipped
@@ -95,7 +110,10 @@ func (rsu *remoteStorageUtil) uploadOneFile(meta *fileMetadata) error {
 			}
 		}
 		if meta.overwrite || meta.resStatus == notFoundFile {
-			utilClass.uploadFile(dataFile, meta, encryptMeta, maxConcurrency, meta.options.MultiPartThreshold)
+			err := utilClass.uploadFile(dataFile, meta, encryptMeta, maxConcurrency, meta.options.MultiPartThreshold)
+			if err != nil {
+				logger.Debugf("Error uploading %v. err: %v", dataFile, err)
+			}
 		}
 		if meta.resStatus == uploaded || meta.resStatus == renewToken || meta.resStatus == renewPresignedURL {
 			return nil
@@ -123,7 +141,7 @@ func (rsu *remoteStorageUtil) uploadOneFile(meta *fileMetadata) error {
 }
 
 func (rsu *remoteStorageUtil) uploadOneFileWithRetry(meta *fileMetadata) error {
-	utilClass := rsu.getNativeCloudType(meta.stageInfo.LocationType)
+	utilClass := rsu.getNativeCloudType(meta.stageInfo.LocationType, rsu.cfg)
 	retryOuter := true
 	for i := 0; i < 10; i++ {
 		// retry
@@ -134,7 +152,9 @@ func (rsu *remoteStorageUtil) uploadOneFileWithRetry(meta *fileMetadata) error {
 		if meta.resStatus == uploaded || meta.resStatus == skipped {
 			for j := 0; j < 10; j++ {
 				status := meta.resStatus
-				utilClass.getFileHeader(meta, meta.dstFileName)
+				if _, err := utilClass.getFileHeader(meta, meta.dstFileName); err != nil {
+					logger.Infof("error while getting file %v header. %v", meta.dstFileSize, err)
+				}
 				// check file header status and verify upload/skip
 				if meta.resStatus == notFoundFile {
 					time.Sleep(time.Second) // wait 1 second
@@ -162,9 +182,15 @@ func (rsu *remoteStorageUtil) uploadOneFileWithRetry(meta *fileMetadata) error {
 
 func (rsu *remoteStorageUtil) downloadOneFile(meta *fileMetadata) error {
 	fullDstFileName := path.Join(meta.localLocation, baseName(meta.dstFileName))
-	fullDstFileName = expandUser(fullDstFileName)
+	fullDstFileName, err := expandUser(fullDstFileName)
+	if err != nil {
+		return err
+	}
 	if !filepath.IsAbs(fullDstFileName) {
-		cwd, _ := os.Getwd()
+		cwd, err := os.Getwd()
+		if err != nil {
+			return err
+		}
 		fullDstFileName = filepath.Join(cwd, fullDstFileName)
 	}
 	baseDir, err := getDirectory()
@@ -177,8 +203,11 @@ func (rsu *remoteStorageUtil) downloadOneFile(meta *fileMetadata) error {
 		}
 	}
 
-	utilClass := rsu.getNativeCloudType(meta.stageInfo.LocationType)
-	header := utilClass.getFileHeader(meta, meta.srcFileName)
+	utilClass := rsu.getNativeCloudType(meta.stageInfo.LocationType, meta.sfa.sc.cfg)
+	header, err := utilClass.getFileHeader(meta, meta.srcFileName)
+	if err != nil {
+		return err
+	}
 	if header != nil {
 		meta.srcFileSize = header.contentLength
 	}
@@ -193,38 +222,37 @@ func (rsu *remoteStorageUtil) downloadOneFile(meta *fileMetadata) error {
 		if meta.resStatus == downloaded {
 			if meta.encryptionMaterial != nil {
 				if meta.presignedURL != nil {
-					header = utilClass.getFileHeader(meta, meta.srcFileName)
+					header, err = utilClass.getFileHeader(meta, meta.srcFileName)
+					if err != nil {
+						return err
+					}
 				}
-				tmpDstFileName, err := decryptFile(header.encryptionMetadata,
-					meta.encryptionMaterial, fullDstFileName, 0, meta.tmpDir)
-				if err != nil {
-					return err
+				if meta.options.GetFileToStream {
+					totalFileSize, err := decryptStreamCBC(header.encryptionMetadata,
+						meta.encryptionMaterial, 0, meta.dstStream, meta.sfa.streamBuffer)
+					if err != nil {
+						return err
+					}
+					meta.sfa.streamBuffer.Truncate(totalFileSize)
+					meta.dstFileSize = int64(totalFileSize)
+				} else {
+					tmpDstFileName, err := decryptFileCBC(header.encryptionMetadata,
+						meta.encryptionMaterial, fullDstFileName, 0, meta.tmpDir)
+					if err != nil {
+						return err
+					}
+					if err = os.Rename(tmpDstFileName, fullDstFileName); err != nil {
+						return err
+					}
 				}
-				if err = os.Rename(tmpDstFileName, fullDstFileName); err != nil {
-					return err
-				}
+
 			}
-			if fi, err := os.Stat(fullDstFileName); err == nil {
-				meta.dstFileSize = fi.Size()
+			if !meta.options.GetFileToStream {
+				if fi, err := os.Stat(fullDstFileName); err == nil {
+					meta.dstFileSize = fi.Size()
+				}
 			}
 			return nil
-		} else if meta.resStatus == renewPresignedURL {
-			return nil
-		} else if meta.resStatus == renewToken {
-			return nil
-		} else if meta.resStatus == needRetryWithLowerConcurrency {
-			maxConcurrency = meta.parallel - int64(retry)*meta.parallel/int64(maxRetry)
-			maxConcurrency = int64Max(defaultConcurrency, maxConcurrency)
-			meta.lastMaxConcurrency = int(maxConcurrency)
-			if !meta.noSleepingTime {
-				sleepingTime := intMin(int(math.Exp2(float64(retry))), 16)
-				time.Sleep(time.Duration(sleepingTime) * time.Second)
-			}
-		} else if meta.resStatus == needRetry {
-			if !meta.noSleepingTime {
-				sleepingTime := intMin(int(math.Exp2(float64(retry))), 16)
-				time.Sleep(time.Duration(sleepingTime) * time.Second)
-			}
 		}
 		lastErr = meta.lastError
 	}
