@@ -1,4 +1,4 @@
-// Copyright (c) 2021 Snowflake Computing Inc. All rights reserved.
+// Copyright (c) 2021-2022 Snowflake Computing Inc. All rights reserved.
 
 package gosnowflake
 
@@ -19,17 +19,19 @@ const (
 
 const (
 	typeKey          = "type"
+	sourceKey        = "source"
 	queryIDKey       = "QueryID"
 	driverTypeKey    = "DriverType"
 	driverVersionKey = "DriverVersion"
+	golangVersionKey = "GolangVersion"
 	sqlStateKey      = "SQLState"
 	reasonKey        = "reason"
 	errorNumberKey   = "ErrorNumber"
 	stacktraceKey    = "Stacktrace"
-	exceptionKey     = "Exception"
 )
 
 const (
+	telemetrySource      = "golang_driver"
 	sqlException         = "client_sql_exception"
 	connectionParameters = "client_connection_parameters"
 )
@@ -96,11 +98,12 @@ func (st *snowflakeTelemetry) sendBatch() error {
 	}
 	resp, err := st.sr.FuncPost(context.Background(), st.sr,
 		st.sr.getFullURL(telemetryPath, nil), headers, body,
-		defaultTelemetryTimeout, true)
+		defaultTelemetryTimeout, defaultTimeProvider, nil)
 	if err != nil {
 		logger.Info("failed to upload metrics to telemetry. err: %v", err)
 		return err
 	}
+	defer resp.Body.Close()
 	if resp.StatusCode != http.StatusOK {
 		err = fmt.Errorf("non-successful response from telemetry server: %v. "+
 			"disabling telemetry", resp.StatusCode)
