@@ -1,32 +1,53 @@
-// Copyright (c) 2017-2021 Snowflake Computing Inc. All right reserved.
+// Copyright (c) 2017-2022 Snowflake Computing Inc. All rights reserved.
 
 package gosnowflake
 
 import (
+	"context"
 	"database/sql/driver"
+	"errors"
 )
 
 type snowflakeTx struct {
-	sc *snowflakeConn
+	sc  *snowflakeConn
+	ctx context.Context
 }
 
-func (tx *snowflakeTx) Commit() (err error) {
-	if tx.sc == nil || tx.sc.rest == nil {
-		return driver.ErrBadConn
+type txCommand int
+
+const (
+	commit txCommand = iota
+	rollback
+)
+
+func (cmd txCommand) string() (string, error) {
+	switch cmd {
+	case commit:
+		return "COMMIT", nil
+	case rollback:
+		return "ROLLBACK", nil
 	}
-	_, err = tx.sc.exec(tx.sc.ctx, "COMMIT", false /* noResult */, false /* isInternal */, false /* describeOnly */, nil)
+	return "", errors.New("unsupported transaction command")
+}
+
+func (tx *snowflakeTx) Commit() error {
+	return tx.execTxCommand(commit)
+}
+
+func (tx *snowflakeTx) Rollback() error {
+	return tx.execTxCommand(rollback)
+}
+
+func (tx *snowflakeTx) execTxCommand(command txCommand) (err error) {
+	txStr, err := command.string()
 	if err != nil {
 		return
 	}
-	tx.sc = nil
-	return
-}
-
-func (tx *snowflakeTx) Rollback() (err error) {
 	if tx.sc == nil || tx.sc.rest == nil {
 		return driver.ErrBadConn
 	}
-	_, err = tx.sc.exec(tx.sc.ctx, "ROLLBACK", false /* noResult */, false /* isInternal */, false /* describeOnly */, nil)
+	isInternal := isInternal(tx.ctx)
+	_, err = tx.sc.exec(tx.ctx, txStr, false /* noResult */, isInternal, false /* describeOnly */, nil)
 	if err != nil {
 		return
 	}
