@@ -1,44 +1,52 @@
-// Copyright (c) 2017-2021 Snowflake Computing Inc. All right reserved.
+// Copyright (c) 2017-2022 Snowflake Computing Inc. All rights reserved.
 
 package gosnowflake
 
 import (
 	"context"
 	"database/sql/driver"
+	"fmt"
 	"io"
 	"math/rand"
+	"os"
 	"strings"
 	"sync"
 	"time"
 
-	"github.com/google/uuid"
+	"github.com/apache/arrow-go/v18/arrow/memory"
 )
 
 type contextKey string
 
 const (
-	// multiStatementCount controls the number of queries to execute in a single API call
-	multiStatementCount contextKey = "MULTI_STATEMENT_COUNT"
-	// asyncMode tells the server to not block the request on executing the entire query
-	asyncMode contextKey = "ASYNC_MODE_QUERY"
-	// queryIDChannel is the channel to receive the query ID from
-	queryIDChannel contextKey = "QUERY_ID_CHANNEL"
-	// snowflakeRequestIDKey is optional context key to specify request id
-	snowflakeRequestIDKey contextKey = "SNOWFLAKE_REQUEST_ID"
-	// fetchResultByID the queryID of query result to fetch
-	fetchResultByID contextKey = "SF_FETCH_RESULT_BY_ID"
-	// fileStreamFile is the address of the file to be uploaded via PUT
-	fileStreamFile contextKey = "STREAMING_PUT_FILE"
-	// fileTransferOptions allows the user to pass in custom
-	fileTransferOptions contextKey = "FILE_TRANSFER_OPTIONS"
-	// enableHigherPrecision returns numbers with higher precision in a *big format
-	enableHigherPrecision contextKey = "ENABLE_HIGHER_PRECISION"
+	multiStatementCount              contextKey = "MULTI_STATEMENT_COUNT"
+	asyncMode                        contextKey = "ASYNC_MODE_QUERY"
+	queryIDChannel                   contextKey = "QUERY_ID_CHANNEL"
+	snowflakeRequestIDKey            contextKey = "SNOWFLAKE_REQUEST_ID"
+	fetchResultByID                  contextKey = "SF_FETCH_RESULT_BY_ID"
+	fileStreamFile                   contextKey = "STREAMING_PUT_FILE"
+	fileGetStream                    contextKey = "STREAMING_GET_FILE"
+	fileTransferOptions              contextKey = "FILE_TRANSFER_OPTIONS"
+	enableHigherPrecision            contextKey = "ENABLE_HIGHER_PRECISION"
+	enableArrowBatchesUtf8Validation contextKey = "ENABLE_ARROW_BATCHES_UTF8_VALIDATION"
+	arrowBatches                     contextKey = "ARROW_BATCHES"
+	arrowAlloc                       contextKey = "ARROW_ALLOC"
+	arrowBatchesTimestampOption      contextKey = "ARROW_BATCHES_TIMESTAMP_OPTION"
+	queryTag                         contextKey = "QUERY_TAG"
+	enableStructuredTypes            contextKey = "ENABLE_STRUCTURED_TYPES"
+	mapValuesNullable                contextKey = "MAP_VALUES_NULLABLE"
+	arrayValuesNullable              contextKey = "ARRAY_VALUES_NULLABLE"
 )
 
 const (
 	describeOnly        contextKey = "DESCRIBE_ONLY"
+	internalQuery       contextKey = "INTERNAL_QUERY"
 	cancelRetry         contextKey = "CANCEL_RETRY"
 	streamChunkDownload contextKey = "STREAM_CHUNK_DOWNLOAD"
+)
+
+var (
+	defaultTimeProvider = &unixTimeProvider{}
 )
 
 // WithMultiStatement returns a context that allows the user to execute the desired number of sql queries in one query
@@ -57,7 +65,7 @@ func WithQueryIDChan(ctx context.Context, c chan<- string) context.Context {
 }
 
 // WithRequestID returns a new context with the specified snowflake request id
-func WithRequestID(ctx context.Context, requestID uuid.UUID) context.Context {
+func WithRequestID(ctx context.Context, requestID UUID) context.Context {
 	return context.WithValue(ctx, snowflakeRequestIDKey, requestID)
 }
 
@@ -76,6 +84,11 @@ func WithFileStream(ctx context.Context, reader io.Reader) context.Context {
 	return context.WithValue(ctx, fileStreamFile, reader)
 }
 
+// WithFileGetStream returns a context that contains the address of the file stream to be GET
+func WithFileGetStream(ctx context.Context, writer io.Writer) context.Context {
+	return context.WithValue(ctx, fileGetStream, writer)
+}
+
 // WithFileTransferOptions returns a context that contains the address of file transfer options
 func WithFileTransferOptions(ctx context.Context, options *SnowflakeFileTransferOptions) context.Context {
 	return context.WithValue(ctx, fileTransferOptions, options)
@@ -89,17 +102,90 @@ func WithDescribeOnly(ctx context.Context) context.Context {
 // WithHigherPrecision returns a context that enables higher precision by
 // returning a *big.Int or *big.Float variable when querying rows for column
 // types with numbers that don't fit into its native Golang counterpart
+// When used in combination with WithArrowBatches, original BigDecimal in arrow batches will be preserved.
 func WithHigherPrecision(ctx context.Context) context.Context {
 	return context.WithValue(ctx, enableHigherPrecision, true)
 }
 
+// WithArrowBatches returns a context that allows users to retrieve
+// arrow.Record download workers upon querying
+func WithArrowBatches(ctx context.Context) context.Context {
+	return context.WithValue(ctx, arrowBatches, true)
+}
+
+// WithArrowAllocator returns a context embedding the provided allocator
+// which will be utilized by chunk downloaders when constructing Arrow
+// objects.
+func WithArrowAllocator(ctx context.Context, pool memory.Allocator) context.Context {
+	return context.WithValue(ctx, arrowAlloc, pool)
+}
+
+// WithOriginalTimestamp in combination with WithArrowBatches returns a context
+// that allows users to retrieve arrow.Record with original timestamp struct returned by Snowflake.
+// It can be used in case arrow.Timestamp cannot fit original timestamp values.
+//
+// Deprecated: please use WithArrowBatchesTimestampOption instead.
+func WithOriginalTimestamp(ctx context.Context) context.Context {
+	return context.WithValue(ctx, arrowBatchesTimestampOption, UseOriginalTimestamp)
+}
+
+// WithArrowBatchesTimestampOption in combination with WithArrowBatches returns a context
+// that allows users to retrieve arrow.Record with different timestamp options.
+// UseNanosecondTimestamp: arrow.Timestamp in nanosecond precision, could cause ErrTooHighTimestampPrecision if arrow.Timestamp cannot fit original timestamp values.
+// UseMicrosecondTimestamp: arrow.Timestamp in microsecond precision
+// UseMillisecondTimestamp: arrow.Timestamp in millisecond precision
+// UseSecondTimestamp: arrow.Timestamp in second precision
+// UseOriginalTimestamp: original timestamp struct returned by Snowflake. It can be used in case arrow.Timestamp cannot fit original timestamp values.
+func WithArrowBatchesTimestampOption(ctx context.Context, option snowflakeArrowBatchesTimestampOption) context.Context {
+	return context.WithValue(ctx, arrowBatchesTimestampOption, option)
+}
+
+// WithArrowBatchesUtf8Validation in combination with WithArrowBatches returns a context that
+// will validate and replace invalid UTF-8 characters in string columns with the replacement character
+// Theoretically, this should not be necessary, because arrow string column is only intended to contain valid UTF-8 characters.
+// However, in practice, it is possible that the data in the string column is not valid UTF-8.
+func WithArrowBatchesUtf8Validation(ctx context.Context) context.Context {
+	return context.WithValue(ctx, enableArrowBatchesUtf8Validation, true)
+
+}
+
+// WithQueryTag returns a context that will set the given tag as the QUERY_TAG
+// parameter on any queries that are run
+func WithQueryTag(ctx context.Context, tag string) context.Context {
+	return context.WithValue(ctx, queryTag, tag)
+}
+
+// WithStructuredTypesEnabled changes how structured types are returned.
+// Without this context structured types are returned as strings.
+// With this context enabled, structured types are returned as native Go types.
+func WithStructuredTypesEnabled(ctx context.Context) context.Context {
+	return context.WithValue(ctx, enableStructuredTypes, true)
+}
+
+// WithMapValuesNullable changes how map values are returned.
+// Instead of simple values (like string) sql.NullXXX wrappers (like sql.NullString) are used.
+func WithMapValuesNullable(ctx context.Context) context.Context {
+	return context.WithValue(ctx, mapValuesNullable, true)
+}
+
+// WithArrayValuesNullable changes how array values are returned.
+// Instead of simple values (like string) sql.NullXXX wrappers (like sql.NullString) are used.
+func WithArrayValuesNullable(ctx context.Context) context.Context {
+	return context.WithValue(ctx, arrayValuesNullable, true)
+}
+
+// WithInternal sets the internal query flag.
+func WithInternal(ctx context.Context) context.Context {
+	return context.WithValue(ctx, internalQuery, true)
+}
+
 // Get the request ID from the context if specified, otherwise generate one
-func getOrGenerateRequestIDFromContext(ctx context.Context) uuid.UUID {
-	requestID, ok := ctx.Value(snowflakeRequestIDKey).(uuid.UUID)
-	if ok && requestID != uuid.Nil {
+func getOrGenerateRequestIDFromContext(ctx context.Context) UUID {
+	requestID, ok := ctx.Value(snowflakeRequestIDKey).(UUID)
+	if ok && requestID != nilUUID {
 		return requestID
 	}
-	return uuid.New()
+	return NewUUID()
 }
 
 // integer min
@@ -217,12 +303,54 @@ func escapeForCSV(value string) string {
 	return value
 }
 
-func randomString(n int) string {
-	rand.Seed(time.Now().UnixNano())
-	alpha := []rune("abcdefghijklmnopqrstuvwxyz")
-	b := make([]rune, n)
-	for i := range b {
-		b[i] = alpha[rand.Intn(len(alpha))]
+// GetFromEnv is used to get the value of an environment variable from the system
+func GetFromEnv(name string, failOnMissing bool) (string, error) {
+	if value := os.Getenv(name); value != "" {
+		return value, nil
 	}
-	return string(b)
+	if failOnMissing {
+		return "", fmt.Errorf("%v environment variable is not set", name)
+	}
+	return "", nil
+}
+
+type currentTimeProvider interface {
+	currentTime() int64
+}
+
+type unixTimeProvider struct {
+}
+
+func (utp *unixTimeProvider) currentTime() int64 {
+	return time.Now().UnixMilli()
+}
+
+func contains[T comparable](s []T, e T) bool {
+	for _, v := range s {
+		if v == e {
+			return true
+		}
+	}
+	return false
+}
+
+func chooseRandomFromRange(min float64, max float64) float64 {
+	return rand.Float64()*(max-min) + min
+}
+
+func withLowerKeys[T any](in map[string]T) map[string]T {
+	out := make(map[string]T)
+	for k, v := range in {
+		out[strings.ToLower(k)] = v
+	}
+	return out
+}
+
+func findByPrefix(in []string, prefix string) int {
+	for i, v := range in {
+		if strings.HasPrefix(v, prefix) {
+			return i
+		}
+	}
+	return -1
 }
